### 2018-11-23 #266 @Robinlovelace
stats19
The goal of stats19 is to make it easy to work with road crash data. Specifically it enables access to and processing of the UK’s official road traffic casualty database, which is called STATS19.

```
Package: stats19
Title: Work with open road traffic casualty data from Great Britain
Version: 0.1.0.9000
    person("Robin", "Lovelace", email = "rob00x@gmail.com", role = c("aut", "cre"),
    person("Malcolm", "Morgan", email = "m.morgan1@leeds.ac.uk", role = c("aut")),
    person("Layik", "Hama", email = "layik.hama@gmail.com", role = c("aut"))
  )
Description: Tools to help process and analysie the UK road collision data also known as Stats19.
    sf,
    foreign,
    readxl
```

-   URL for the package (the development repository, not a stylized html page): https://github.com/ITSLeeds/stats19

- **data retrieval**: downloads and formats publicly available road safety data that is difficult (impossible for most people) to use effectively without

Academic, industry and public sector researchers investigating road safety, specifically people wanting to perform natural experimental studies to find out how best to make the transport system safer. It will also be of interest for people interested in large point-on-network data for methodological applications.

Previous packages/code-bases:

- code in the [bikeR](https://github.com/Robinlovelace/bikeR) repo underlying an academic paper on collisions involving cyclists
- functions in [**stplanr**](https://github.com/ropensci/stplanr/blob/master/R/load-stats19.R) for downloading Stats19 data
- updated functions related to the [CyIPT](https://github.com/cyipt/stats19) project

 but this package takes the best of the preceding packages and adds new features including:

- automated and reproducible data formatting code based on DfT's official [guidelines](http://data.dft.gov.uk/road-accidents-safety-data/Road-Accident-Safety-Data-Guide.xls) (before labels were hard-coded)
- continuous integration
- integration with **sf** (the **stplanr** implementation used **sp**)
- inclusion of much more data (back to 1974 and until 2017, previous work only covered 1985:2015)

- Currently the style uses `=` assignment, which differs from the recommended [style](http://r-pkgs.had.co.nz/r.html)
