### 2018-11-23 ITSLeeds/stats19#11 @Robinlovelace
get_stats19()
See #3
