### 2018-11-23 a63cef8
Add proper check for var matching
