+++ README.Rmd
@@ -28,7 +28,7 @@ The package builds on previous work including:
 
 - code in the [bikeR](https://github.com/Robinlovelace/bikeR) repo underlying an academic paper on collisions involving cyclists [@lovelace_who_2016]
 - functions in [**stplanr**](https://github.com/ropensci/stplanr/blob/master/R/load-stats19.R) for downloading Stats19 data
-- updated functions related to the [CyIPT](https://github.com/cyipt/stats19) project.
+- updated functions related to the [CyIPT](https://github.com/cyipt/stats19) project
 
 ## Installation
 
@@ -75,7 +75,7 @@ crashes_2017 = format_accidents(crashes_2017_raw)
 
 What just happened?
 We read-in the 'raw' Stats19 data without cleaning messy column names or re-categorising the outputs.
-`format_accidents()` does this hard work, automating the process of matching column names with variable names and labels in a [`.xls` file](data.dft.gov.uk/road-accidents-safety-data/Road-Accident-Safety-Data-Guide.xls) provided by the DfT.
+`format_accidents()` does this hard work, automating the process of matching column names with variable names and labels in a [`.xls` file](http://data.dft.gov.uk/road-accidents-safety-data/Road-Accident-Safety-Data-Guide.xls) provided by the DfT.
 This means `crashes_2017` is much more usable than `crashes_2017_raw`, as shown below, which selects three records and four variables for messy and clean datasets:
 
 ```{r}
@@ -86,9 +86,10 @@ crashes_2017_raw[random_n, key_vars]
 crashes_2017[random_n, key_vars]
 ```
 
-More data can be read-in as follows:
+<!-- More data can be read-in as follows: -->
 
-```{r, eval=FALSE}
+```{r, eval=FALSE, echo=FALSE}
+# old code to be up-dated
 d14 = "Stats19_Data_2005-2014"
 crashes_2005_2014 = read_accidents(data_dir = d14)
 crashes_2005_2014_f = format_stats19_2005_2014_ac(crashes_2005_2014)
