+++ README.md
@@ -27,7 +27,7 @@ The package builds on previous work including:
     [**stplanr**](https://github.com/ropensci/stplanr/blob/master/R/load-stats19.R)
     for downloading Stats19 data
   - updated functions related to the
-    [CyIPT](https://github.com/cyipt/stats19) project.
+    [CyIPT](https://github.com/cyipt/stats19) project
 
 ## Installation
 
@@ -35,7 +35,7 @@ Install and attach the latest version with:
 
 ``` r
 devtools::install_github("ITSLeeds/stats19")
-#> Skipping install of 'stats19' from a github remote, the SHA1 (2b9dec1e) has not changed since last install.
+#> Skipping install of 'stats19' from a github remote, the SHA1 (a63cef80) has not changed since last install.
 #>   Use `force = TRUE` to force installation
 library(stats19)
 ```
@@ -92,7 +92,7 @@ What just happened? We read-in the ‘raw’ Stats19 data without cleaning
 messy column names or re-categorising the outputs. `format_accidents()`
 does this hard work, automating the process of matching column names
 with variable names and labels in a [`.xls`
-file](data.dft.gov.uk/road-accidents-safety-data/Road-Accident-Safety-Data-Guide.xls)
+file](http://data.dft.gov.uk/road-accidents-safety-data/Road-Accident-Safety-Data-Guide.xls)
 provided by the DfT. This means `crashes_2017` is much more usable than
 `crashes_2017_raw`, as shown below, which selects three records and four
 variables for messy and clean datasets:
@@ -105,33 +105,19 @@ crashes_2017_raw[random_n, key_vars]
 #> # A tibble: 3 x 4
 #>   Police_Force Accident_Severity Speed_limit Light_Conditions
 #>          <int>             <int>       <int>            <int>
-#> 1            6                 3          30                4
-#> 2           45                 3          30                1
-#> 3           10                 2          30                5
+#> 1           45                 2          50                1
+#> 2            4                 3          30                1
+#> 3           44                 3          30                1
 crashes_2017[random_n, key_vars]
 #> # A tibble: 3 x 4
-#>   police_force       accident_severity speed_limit light_conditions       
-#>   <chr>              <chr>                   <int> <chr>                  
-#> 1 Greater Manchester Slight                     30 Darkness - lights lit  
-#> 2 Surrey             Slight                     30 Daylight               
-#> 3 Northumbria        Serious                    30 Darkness - lights unlit
+#>   police_force accident_severity speed_limit light_conditions
+#>   <chr>        <chr>                   <int> <chr>           
+#> 1 Surrey       Serious                    50 Daylight        
+#> 2 Lancashire   Slight                     30 Daylight        
+#> 3 Hampshire    Slight                     30 Daylight
 ```
 
-More data can be read-in as follows:
-
-``` r
-d14 = "Stats19_Data_2005-2014"
-crashes_2005_2014 = read_accidents(data_dir = d14)
-crashes_2005_2014_f = format_stats19_2005_2014_ac(crashes_2005_2014)
-d15 = "RoadSafetyData_2015"
-crashes_2015 = read_accidents(data_dir = d15, filename = "Accidents_2015.csv")
-crashes_2015_f = format_stats19_2015_ac(crashes_2015)
-d16 = "dftRoadSafety_Accidents_2016"
-crashes_2016 = read_accidents(data_dir = d16, filename = "dftRoadSafety_Accidents_2016.csv")
-crashes_2016_f = format_stats19_2016_ac(crashes_2016)
-all_crashes = rbind(crashes_2015_f, crashes_2016_f, crashes_2017_f)
-table(ac$Accident_Severity)
-```
+<!-- More data can be read-in as follows: -->
 
 ## References
 
