+++ NAMESPACE
@@ -5,6 +5,7 @@ export(dl_stats19_2005_2014)
 export(dl_stats19_2015)
 export(dl_stats19_2016_ac)
 export(dl_stats19_2017_ac)
+export(find_file_name)
 export(format_accidents)
 export(format_column_names)
 export(format_stats19_2005_2014_ac)
