### 2018-11-23 552817f
Update readme - better vars + description
