+++ README.Rmd
@@ -74,12 +74,13 @@ crashes_2017 = format_accidents(crashes_2017_raw)
 ```
 
 What just happened?
-We read-in the 'raw' Stats19 data without cleaning messy column names or re-categorising the outputs.
+We read-in data on all road crashes recorded by the police in 2017 across Great Britain.
+`read_accidents()` imports the 'raw' Stats19 data without cleaning messy column names or re-categorising the outputs.
 `format_accidents()` does this work, automating the process of matching column names with variable names and labels in a [`.xls` file](http://data.dft.gov.uk/road-accidents-safety-data/Road-Accident-Safety-Data-Guide.xls) provided by the DfT.
-This means `crashes_2017` is much more usable than `crashes_2017_raw`, as shown below, which selects three records and four variables for messy and clean datasets:
+This means `crashes_2017` is much more usable than `crashes_2017_raw`, as shown below, which shows three records and some key variables in the messy and clean datasets:
 
 ```{r}
-key_patt = "severity|speed|force|light|junction"
+key_patt = "severity|speed|light|junction_detail"
 key_vars = grep(key_patt, x = names(crashes_2017_raw), ignore.case = TRUE)
 random_n = sample(x = nrow(crashes_2017_raw), size = 3)
 crashes_2017_raw[random_n, key_vars]
