### 2018-11-23 36e42ce
Refactor: dl_stats19, test coverage and cleanup, closes #4, see #12
