+++ tests/testthat/test-dl_stats19.R
@@ -1,0 +1,5 @@
+context("test-dl_stats19")
+
+test_that("dl_stats19_2017_ac works", {
+  expect_output(dl_stats19_2017_ac())
+})
