### 2018-11-23 8e3a4b7
Show crossing var in readme
