+++ man/dl_stats19.Rd
@@ -4,9 +4,11 @@
 \alias{dl_stats19}
 \title{Download Stats19 data for a year or range of two years.}
 \usage{
-dl_stats19(years = "", type = "Accidents")
+dl_stats19(file_name = NULL, years = "", type = "Accidents")
 }
 \arguments{
+\item{file_name}{The file name to download, above two will be ignore.}
+
 \item{years}{Either a single year or a two year range, defaults to 2 years ago}
 
 \item{type}{One of 'Accidents', 'Casualties', 'Vehicles'; defaults to 'Accidents'#' @export}
