### 2018-11-23 65222dd
Near complete stats19 interctive download selection
