### 2018-11-23 a05d476
Near complete stats19 interctive download selection
