+++ .travis.yml
@@ -3,7 +3,13 @@ cache: packages
 sudo: required
 dist: trusty
 
+r_github_packages:
+    - jimhester/covr
+
 before_install:
   - sudo add-apt-repository ppa:ubuntugis/ppa --yes
   - sudo apt-get --yes --force-yes update -qq
   - sudo apt-get install --yes libudunits2-dev libproj-dev libgeos-dev libgdal-dev
+
+after_success:
+  - Rscript -e 'covr::codecov()'
