### 2018-11-23 76c3282
Coverage for format.R
