### 2018-11-23 17e39da
Refactor dl_stats19 methods in #12, and fix .zip in main function
