### 2018-11-23 7c84412
Refactor dl_stats19 methods in #12, and fix .zip in main function
