+++ man/dl_2005_2014.Rd
@@ -1,14 +1,14 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/dl_stats19.R
-\name{dl_stats19_2005_2014}
-\alias{dl_stats19_2005_2014}
-\alias{dl_stats19_2015}
-\alias{dl_stats19_2016_ac}
-\alias{dl_stats19_2017_ac}
+\name{dl_2005_2014}
+\alias{dl_2005_2014}
+\alias{dl_2015}
+\alias{dl_2016_accidents}
+\alias{dl_2017_accidents}
 \title{Download Stats19 data}
 \usage{
 
-  dl_stats19_2005_2014(zip_url = paste0("http://data.dft.gov.uk.s3.amazonaws.com/",
+  dl_2005_2014(zip_url = paste0("http://data.dft.gov.uk.s3.amazonaws.com/",
   "road-accidents-safety-data/Stats19_Data_2005-2014.zip"),
   exdir = "Stats19_Data_2005-2014")
 }
@@ -30,11 +30,11 @@ Ensure you have a fast internet connection and at least 100 Mb space
 
 \examples{
 \dontrun{
-dl_stats19_2005_2014()
+dl_2005_2014()
 
 # Load all stats19 datasets
 ac = read_accidents()
-ca = read_stats19_2005_2014_ca()
+ca = read_2005_2014_ca()
 ve = read_stats19_2005_2014_ve()
 # now you can analyse the UK's stats19 data in a single table
 }
