+++ NAMESPACE
@@ -1,10 +1,10 @@
 # Generated by roxygen2: do not edit by hand
 
+export(dl_2005_2014)
+export(dl_2015)
+export(dl_2016_accidents)
+export(dl_2017_accidents)
 export(dl_schema)
-export(dl_stats19_2005_2014)
-export(dl_stats19_2015)
-export(dl_stats19_2016_ac)
-export(dl_stats19_2017_ac)
 export(find_file_name)
 export(format_accidents)
 export(format_column_names)
