### 2018-11-27 a7d8007
Cleanup read_stats19, as per #13
