+++ R/read_stats19.R
@@ -38,306 +38,3 @@ read_accidents = function(data_dir = "dftRoadSafetyData_Accidents_2017", filenam
   ac
 
 }
-#' Format UK 'Stats19' road traffic casualty data
-#'
-#' @section Details:
-#' This is a helper function to format raw stats19 data
-#'
-#' @param ac Data frame representing the raw Stats19 data read-in with `read_csv()`
-#' @param factorize Should some results be returned as factors? `FALSE` by default
-#' @aliases format_stats19_2015_ac format_stats19_2016_ac
-#' @export
-#' @examples
-#' \dontrun{
-#' ac = read_accidents()
-#' sapply(ac, class)
-#' ac_formatted = format_stats19_2005_2014_ac(ac)
-#' sapply(ac_formatted, class)
-#' }
-format_stats19_2005_2014_ac = function(ac, factorize = FALSE) {
-  ac$Accident_Severity =
-    factor(ac$Accident_Severity, labels = c("Fatal", "Serious", "Slight"))
-  ac$Police_Force =
-    factor(ac$Police_Force,
-           labels =
-             c(
-               "Metropolitan Police", "Cumbria", "Lancashire", "Merseyside",
-               "Greater Manchester", "Cheshire", "Northumbria", "Durham", "North Yorkshire",
-               "West Yorkshire", "South Yorkshire", "Humberside", "Cleveland",
-               "West Midlands", "Staffordshire", "West Mercia", "Warwickshire",
-               "Derbyshire", "Nottinghamshire", "Lincolnshire", "Leicestershire",
-               "Northamptonshire", "Cambridgeshire", "Norfolk", "Suffolk", "Bedfordshire",
-               "Hertfordshire", "Essex", "Thames Valley", "Hampshire", "Surrey",
-               "Kent", "Sussex", "City of London", "Devon and Cornwall", "Avon and Somerset",
-               "Gloucestershire", "Wiltshire", "Dorset", "North Wales", "Gwent",
-               "South Wales", "Dyfed-Powys", "Northern", "Grampian", "Tayside",
-               "Fife", "Lothian and Borders", "Central", "Strathclyde", "Dumfries and Galloway"
-             )
-    )
-  ac$`1st_Road_Class` =
-    factor(ac$`1st_Road_Class`,
-           labels = c("Motorway", "A(M)", "A", "B", "C", "Unclassified")
-    )
-  ac$Road_Type =
-    factor(ac$Road_Type,
-           labels = c(
-             "Roundabout", "One way street", "Dual carriageway", "Single carriageway",
-             "Slip road", "Unknown"
-           )
-    )
-  ac$Junction_Detail =
-    factor(ac$Junction_Detail,
-           labels =
-             c(
-               "Data missing or out of range", "Not at junction or within 20 metres",
-               "Roundabout", "Mini-roundabout", "T or staggered junction", "Slip road",
-               "Crossroads", "More than 4 arms (not roundabout)", "Private drive or entrance",
-               "Other junction"
-             )
-    )
-  ac$Light_Conditions =
-    factor(ac$Light_Conditions,
-           labels = c(
-             "Daylight", "Darkness - lights lit", "Darkness - lights unlit",
-             "Darkness - no lighting", "Darkness - lighting unknown"
-           )
-    )
-  ac$Weather_Conditions =
-    factor(ac$Weather_Conditions,
-           labels = c(
-             "Data missing or out of range", "Fine no high winds", "Raining no high winds",
-             "Snowing no high winds", "Fine + high winds", "Raining + high winds",
[…]
