### 2018-11-27 73db753
Cleanup read_stats19, as per #13
