+++ man/format_stats19_2005_2014_ca.Rd
@@ -1,24 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/read_stats19.R
-\name{format_stats19_2005_2014_ca}
-\alias{format_stats19_2005_2014_ca}
-\title{Format UK 'Stats19' road traffic casualty data}
-\usage{
-format_stats19_2005_2014_ca(ca)
-}
-\arguments{
-\item{ca}{Dataframe representing the raw Stats19 data read-in with `read_csv()`.}
-}
-\description{
-Format UK 'Stats19' road traffic casualty data
-}
-\section{Details}{
-
-This is a helper function to format raw stats19 data
-}
-
-\examples{
-\dontrun{
-ca = format_stats19_2005_2014_ca(ca)
-}
-}
