+++ NAMESPACE
@@ -8,12 +8,5 @@ export(dl_schema)
 export(find_file_name)
 export(format_accidents)
 export(format_column_names)
-export(format_stats19_2005_2014_ac)
-export(format_stats19_2005_2014_ca)
-export(format_stats19_2005_2014_ve)
-export(format_stats19_2015_ac)
-export(format_stats19_2016_ac)
 export(read_accidents)
 export(read_schema)
-export(read_stats19_ca)
-export(read_stats19_ve)
