### 2018-11-27 6a7bea5
Cover read_accidents,see #13
