### 2018-11-27 ITSLeeds/stats19#11 @layik
So this function would:
1. call `dl_stats19` with say accidents then >
2. call `read_accidents`
3. call `format_accidents`

Returns a clean `df` ready to be punched.
