### 2018-11-27 c0e3026
Skip download tests on no internet connection using curl package
