+++ tests/testthat/test-read_stats19.R
@@ -1,6 +1,14 @@
 context("test-read_stats19")
+# To run download functions you need an internet connection.
+# pref a fast one
+skip_download = function() {
+  connected = !is.null(curl::nslookup("r-project.org", error = FALSE))
+  if(!connected)
+    skip("No connection to run download function.")
+}
 
 test_that("read_stats19 works", {
+  skip_download()
   # download real data
   acc_2016 = file_names$dftRoadSafety_Accidents_2016.zip
   dl_stats19(file_name = acc_2016)
