### 2018-11-27 a239968
Skip download tests on no internet connection using curl package
