### 2018-11-27 7157693
Update with new dl_ funs in readme
Heads-up @layik - they look great
