### 2018-11-27 4927f22
Update documentation for dl_stats19
