+++ man/dl_stats19.Rd
@@ -27,7 +27,8 @@ have a sufficient internet access and hard disk space.
 
 \examples{
 \dontrun{
-dl_stats19(years = 2017)
-# now you can analyse the UK's stats19 data in a single table
+dl_stats19(years = 2017) # interactively select files...
+# now you can read-in the data
+dl_stats19(years = 2004)
 }
 }
