### 2018-11-27 06daf59
Refactor dl_stats19 code
