### 2018-11-27 93ca0c7
Add basic read_vehicles, may need more column types + tests
