+++ tests/testthat/test-read_stats19.R
@@ -1,13 +1,6 @@
 context("test-read_stats19")
-# To run download functions you need an internet connection.
-# pref a fast one
-skip_download = function() {
-  connected = !is.null(curl::nslookup("r-project.org", error = FALSE))
-  if(!connected)
-    skip("No connection to run download function.")
-}
 
-test_that("read_stats19 works", {
+test_that("read_accidents works", {
   skip_download()
   # download real data
   acc_2016 = file_names$dftRoadSafety_Accidents_2016.zip
@@ -22,4 +15,24 @@ test_that("read_stats19 works", {
   raw_read = read.csv(file.path(path, sub(".zip", ".csv", acc_2016)))
   expect_equal(nrow(read), nrow(raw_read))
   expect_error(read_accidents("junk"))
+})
+
+test_that("read_vehicles works", {
+  skip_download()
+  # download real data
+  veh_2016 = file_names$dftRoadSafetyData_Vehicles_2016.zip
+  dl_stats19(file_name = veh_2016)
+  path = file.path(
+    tempdir(), sub(".zip", "", veh_2016))
+  # read it
+  read = read_vehicles(
+    data_dir = path,
+    filename = "Veh.csv"
+  )
+  raw_read = read.csv(file.path(path, "Veh.csv"))
+  expect_false(identical(
+    class(read$Accident_Index),
+    class(raw_read$Accident_Index)
+    ))
+  expect_error(read_vehicles("junk"))
 })
