### 2018-11-27 446f186
Add basic read_vehicles, may need more column types + tests
