+++ NAMESPACE
@@ -10,3 +10,4 @@ export(format_accidents)
 export(format_column_names)
 export(read_accidents)
 export(read_schema)
+export(read_vehicles)
