### 2018-11-27 016f6f5
Update find_file_name() docs, tests
