### 2018-11-27 ITSLeeds/stats19#11 @Robinlovelace
Yes and maybe provide some additional things, e.g. allow the user to specify if they want the output to be in an `sf` object.
