### 2018-11-27 8f418aa
Add skip to global vars for tests
