### 2018-11-27 fecc647
Fix issues with docs
