+++ man/find_file_name.Rd
@@ -14,3 +14,9 @@ find_file_name(years = "2016", type = "")
 \description{
 Currently, there are 52 file names to download data from.
 }
+\examples{
+find_file_name(2016)
+find_file_name(2016, type = "accident")
+find_file_name(1979)
+find_file_name(2016:2017)
+}
