### 2018-11-27 e41b060
Fix issues with docs
