### 2018-11-28 0712bc4
format_sf - close #9
