+++ man/format_sf.Rd
@@ -1,0 +1,25 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/format.R
+\name{format_sf}
+\alias{format_sf}
+\title{Format convert stats19 data into spatial (sf) object}
+\usage{
+format_sf(x, lonlat = FALSE)
+}
+\arguments{
+\item{x}{Data frame created with `read_accidents()`}
+
+\item{lonlat}{Should the results be returned in longitude/latitude?
+By default `FALSE`, meaning the British National Grid (EPSG code: 27700)
+is used.}
+}
+\description{
+Format convert stats19 data into spatial (sf) object
+}
+\examples{
+\dontrun{
+x = read_accidents()
+x_sf = format_sf(x)
+sf:::plot.sf(x)
+}
+}
