+++ NAMESPACE
@@ -9,6 +9,7 @@ export(dl_stats19)
 export(find_file_name)
 export(format_accidents)
 export(format_column_names)
+export(format_sf)
 export(read_accidents)
 export(read_schema)
 export(read_vehicles)
