### 2018-11-28 43ab452
Clean up handy dl_ functions, see #12
