+++ tests/testthat/test-dl_stats19.R
@@ -1,10 +1,5 @@
 context("test-dl_stats19")
 
-test_that("dl_2017_accidents works", {
-  skip_download()
-  expect_output(dl_2017_accidents())
-})
-
 test_that("dl_stats19 works for junk", {
   expect_output(dl_stats19(type = "junk"))
   expect_output(dl_stats19(years = "2999", type = "junk"))
