### 2018-11-28 4e08e95
Clean up handy dl_ functions, see #12
