+++ NAMESPACE
@@ -1,9 +1,5 @@
 # Generated by roxygen2: do not edit by hand
 
-export(dl_2005_2014)
-export(dl_2015)
-export(dl_2016_accidents)
-export(dl_2017_accidents)
 export(dl_schema)
 export(dl_stats19)
 export(find_file_name)
