### 2018-11-28 7303e1f
Cover format_sf using sample data
