### 2018-11-28 673b5a9
More covr for format_sf
