### 2018-11-28 bc5dc5e
Add read_casualties
