+++ R/read_stats19.R
@@ -40,7 +40,6 @@ read_accidents = function(data_dir = file.path(tempdir(), "dftRoadSafetyData_Acc
   ac
 
 }
-
 #' Import and Stats19 data on vehicles
 #'
 #' @section Details:
@@ -71,4 +70,35 @@ read_vehicles = function(data_dir = file.path(tempdir(), "dftRoadSafetyData_Vehi
     Accident_Index = readr::col_character()
   ))
   ve
+}
+#' Import and Stats19 data on casualties
+#'
+#' @section Details:
+#' The function returns a data frame, in which each record is a reported casualty
+#' in the stats19 dataset.
+#'
+#'
+#' @param data_dir Character string representing where the data is stored.
+#' If empty, R will attempt to download and unzip the data for you.
+#' @param filename Character string of the filename of the .csv to read in - default value
+#' is `Veh.csv` from default data_dir
+#'
+#' @export
+#' @examples
+#' \dontrun{
+#' casualties = read_casualties()
+#' }
+read_casualties = function(data_dir = file.path(tempdir(), "dftRoadSafetyData_Casualties_2017"),
+                         filename = "Veh.csv") {
+
+  if (!filename %in% list.files(data_dir)) {
+    stop("No data found. Change data_dir or Run dl_stats19*() functions first.")
+  }
+
+  # read the data in
+  ca = readr::read_csv(file.path(data_dir, filename), col_types = readr::cols(
+    .default = readr::col_integer(),
+    Accident_Index = readr::col_character()
+  ))
+  ca
 }
