+++ man/read_casualties.Rd
@@ -1,0 +1,30 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/read_stats19.R
+\name{read_casualties}
+\alias{read_casualties}
+\title{Import and Stats19 data on casualties}
+\usage{
+read_casualties(data_dir = file.path(tempdir(),
+  "dftRoadSafetyData_Casualties_2017"), filename = "Veh.csv")
+}
+\arguments{
+\item{data_dir}{Character string representing where the data is stored.
+If empty, R will attempt to download and unzip the data for you.}
+
+\item{filename}{Character string of the filename of the .csv to read in - default value
+is `Veh.csv` from default data_dir}
+}
+\description{
+Import and Stats19 data on casualties
+}
+\section{Details}{
+
+The function returns a data frame, in which each record is a reported casualty
+in the stats19 dataset.
+}
+
+\examples{
+\dontrun{
+casualties = read_casualties()
+}
+}
