+++ NAMESPACE
@@ -7,5 +7,6 @@ export(format_accidents)
 export(format_column_names)
 export(format_sf)
 export(read_accidents)
+export(read_casualties)
 export(read_schema)
 export(read_vehicles)
