### 2018-11-28 3c66dbc
Add read_casualties
