### 2018-11-28 71cf328
Add working read_casualties
