### 2018-11-28 cc5a99f
Add first pass of format_casualties
