+++ man/format_casualties.Rd
@@ -1,30 +1,30 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/format.R
-\name{format_accidents}
-\alias{format_accidents}
-\title{Format stats19 'accidents' data}
+\name{format_casualties}
+\alias{format_casualties}
+\title{Format stats19 casualties}
 \usage{
-format_accidents(x, factorize = FALSE)
+format_casualties(x, factorize = FALSE)
 }
 \arguments{
-\item{x}{Data frame created with `read_accidents()`}
+\item{x}{Data frame created with `read_casualties()`}
 
 \item{factorize}{Should some results be returned as factors? `FALSE` by default}
 }
 \description{
-Format stats19 'accidents' data
+Format stats19 casualties
 }
 \section{Details}{
 
-This is a helper function to format raw stats19 data
+This function formats raw stats19 data
 }
 
 \examples{
 \dontrun{
-x = read_accidents()
+x = read_casualties()
 sapply(x, class)
 table(x$Accident_Severity)
-crashes = format_accidents(x)
+casualties = format_casualties(x)
 sapply(crashes, class)
 table(crashes$accident_severity)
 }
