+++ NAMESPACE
@@ -4,6 +4,7 @@ export(dl_schema)
 export(dl_stats19)
 export(find_file_name)
 export(format_accidents)
+export(format_casualties)
 export(format_column_names)
 export(format_sf)
 export(read_accidents)
