### 2018-11-28 28502c8
Add first pass of format_casualties
