### 2018-11-28 dda8a48
Update format_casualties
