### 2018-11-28 70736c5
Update format_casualties
