### 2018-11-28 1d1e304
Tidy-up format_casualties code
