### 2018-11-28 76c77de
Add format_vehicles
