+++ NAMESPACE
@@ -7,6 +7,7 @@ export(format_accidents)
 export(format_casualties)
 export(format_column_names)
 export(format_sf)
+export(format_vehicles)
 export(read_accidents)
 export(read_casualties)
 export(read_schema)
