### 2018-11-28 90cd38e
Update docs format_vehicles (travis)
