### 2018-11-28 582aeea
README using smart dl and read functions, needs push to update self again
