+++ README.md
@@ -1,103 +1,97 @@
 
-[![Travis build
-status](https://travis-ci.org/ITSLeeds/stats19.svg?branch=master)](https://travis-ci.org/ITSLeeds/stats19)
-[![codecov](https://codecov.io/gh/ITSLeeds/stats19/branch/master/graph/badge.svg)](https://codecov.io/gh/ITSLeeds/stats19)
+[![Travis build status](https://travis-ci.org/ITSLeeds/stats19.svg?branch=master)](https://travis-ci.org/ITSLeeds/stats19) [![codecov](https://codecov.io/gh/ITSLeeds/stats19/branch/master/graph/badge.svg)](https://codecov.io/gh/ITSLeeds/stats19)
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
+stats19
+=======
 
-# stats19
+The goal of **stats19** is to make it easy to work with road crash data. Specifically it enables access to and processing of the UK's official road traffic casualty database, which is called [STATS19](https://data.gov.uk/dataset/cb7ae6f0-4be6-4935-9277-47e5ce24a11f/road-safety-data). The name comes from the form used by the police to record car crashes and other incidents resulting in casualties on the roads.
 
-The goal of **stats19** is to make it easy to work with road crash data.
-Specifically it enables access to and processing of the UK’s official
-road traffic casualty database, which is called
-[STATS19](https://data.gov.uk/dataset/cb7ae6f0-4be6-4935-9277-47e5ce24a11f/road-safety-data).
-The name comes from the form used by the police to record car crashes
-and other incidents resulting in casualties on the roads.
-
-A description of the stats19 data and variables they contain can be
-found in a
-[document](http://data.dft.gov.uk/road-accidents-safety-data/Brief-guide-to%20road-accidents-and-safety-data.doc)
-hosted by the UK’s Department for Transport (DfT).
+A description of the stats19 data and variables they contain can be found in a [document](http://data.dft.gov.uk/road-accidents-safety-data/Brief-guide-to%20road-accidents-and-safety-data.doc) hosted by the UK's Department for Transport (DfT).
 
 The package builds on previous work including:
 
-  - code in the [bikeR](https://github.com/Robinlovelace/bikeR) repo
-    underlying an academic paper on collisions involving cyclists
-    (Lovelace, Roberts, and Kellar 2016)
-  - functions in
-    [**stplanr**](https://github.com/ropensci/stplanr/blob/master/R/load-stats19.R)
-    for downloading Stats19 data
-  - updated functions related to the
-    [CyIPT](https://github.com/cyipt/stats19) project
+-   code in the [bikeR](https://github.com/Robinlovelace/bikeR) repo underlying an academic paper on collisions involving cyclists (Lovelace, Roberts, and Kellar 2016)
+-   functions in [**stplanr**](https://github.com/ropensci/stplanr/blob/master/R/load-stats19.R) for downloading Stats19 data
+-   updated functions related to the [CyIPT](https://github.com/cyipt/stats19) project
 
-## Installation
+Installation
+------------
 
 Install and attach the latest version with:
 
 ``` r
 devtools::install_github("ITSLeeds/stats19")
-#> Skipping install of 'stats19' from a github remote, the SHA1 (c0e30261) has not changed since last install.
-#>   Use `force = TRUE` to force installation
+#> Using GitHub PAT from envvar GITHUB_PAT
+#> Downloading GitHub repo ITSLeeds/stats19@master
+#> from URL https://api.github.com/repos/ITSLeeds/stats19/zipball/master
+#> Installation failed: Bad credentials (401)
 library(stats19)
 ```
 
[…]
