### 2018-11-28 b6fa820
Followup README udpate
