### 2018-11-28 b4ab1d9
Tests for all formats in #10
