### 2018-11-28 561364c
Test Travis, seems to also fail on install_github
