### 2018-11-28 aa7c416
Helper function to scan disk location in prep for #16
