+++ man/locate_files.Rd
@@ -1,0 +1,22 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/utils.R
+\name{locate_files}
+\alias{locate_files}
+\title{Locate a file on disk}
+\usage{
+locate_files(data_dir = tempdir(), type = "", years = "")
+}
+\arguments{
+\item{data_dir}{super directory where dataset(s) were first downloaded to.}
+
+\item{type}{One of 'Accidents', 'Casualties', 'Vehicles'; defaults to 'Accidents', ignores case.}
+
+\item{years}{Provide a year, part of it or maximum range of two years.}
+}
+\description{
+Helper function to locate files. Given below params, the function
+returns 0 or more files found at location/names given.
+}
+\examples{
+locate_files(years = 2016)
+}
