+++ man/find_file_name.Rd
@@ -4,7 +4,8 @@
 \alias{find_file_name}
 \title{Find a file name within stats19::file_names.}
 \usage{
-find_file_name(years = "2016", type = "")
+find_file_name(years = as.integer(format(Sys.Date(), "\%Y")) - 2,
+  type = "")
 }
 \arguments{
 \item{years}{Provide a year, part of it or maximum range of two years. Defaults to two years ago.}
