+++ NAMESPACE
@@ -8,6 +8,7 @@ export(format_casualties)
 export(format_column_names)
 export(format_sf)
 export(format_vehicles)
+export(locate_files)
 export(read_accidents)
 export(read_casualties)
 export(read_schema)
