### 2018-11-28 58e1f24
Helper function to scan disk location in prep for #16
