### 2018-11-28 1c1ba60
Minor, better test case needed
