### 2018-11-29 10a9141
Large: almost done with #16, refactor locate_files
