+++ tests/testthat/test-read_stats19.R
@@ -5,14 +5,16 @@ test_that("read_accidents works", {
   # download real data
   acc_2016 = file_names$dftRoadSafety_Accidents_2016.zip
   dl_stats19(file_name = acc_2016)
-  path = file.path(
-    tempdir(), sub(".zip", "", acc_2016))
+  # make sure we have a csv file to read
+  path = locate_one_file(
+    tempdir(), years = 2016, filename = sub(".zip", ".csv", acc_2016))
   # read it
   read = read_accidents(
-    data_dir = path,
+    years = 2016, # make it unique
+    data_dir = tempdir(),
     filename = sub(".zip", ".csv", acc_2016)
   )
-  raw_read = read.csv(file.path(path, sub(".zip", ".csv", acc_2016)))
+  raw_read = read.csv(path)
   expect_equal(nrow(read), nrow(raw_read))
   expect_error(read_accidents("junk"))
 })
