+++ man/read_schema.Rd
@@ -8,13 +8,10 @@ read_schema(data_dir = tempdir(),
   filename = "Road-Accident-Safety-Data-Guide.xls", sheet = NULL)
 }
 \arguments{
-\item{data_dir}{Character string representing where the data is stored.
-Default value assumes in recent session `tempdir` there is
-a `dftRoadSafetyData_Accidents_2017` directory.}
+\item{data_dir}{Where sets of downloaded data would be found.}
 
-\item{filename}{Character string of the filename of the .csv to read in - default values
-are those downloaded from the UK Department for Transport (DfT).
-Default value is `Acc.csv` under above default `data_dir`.}
+\item{filename}{Character string of the filename of the .csv to read, if this is given, type and
+years determin whether there is a target to read, otherwise disk scan would be needed.}
 
 \item{sheet}{integer to be added if you want to download a single sheet}
 }
