+++ NAMESPACE
@@ -9,6 +9,7 @@ export(format_column_names)
 export(format_sf)
 export(format_vehicles)
 export(locate_files)
+export(locate_one_file)
 export(read_accidents)
 export(read_casualties)
 export(read_schema)
