### 2018-11-29 ae32efe
Large: almost done with #16, refactor locate_files
