### 2018-11-29 035aaeb
Add data_dir to dl_stats19, closes #15
