### 2018-11-29 23afde9
Add data_dir to dl_stats19, closes #15
