+++ man/download_and_unzip.Rd
@@ -4,12 +4,14 @@
 \alias{download_and_unzip}
 \title{Download and unzip given appropriate params}
 \usage{
-download_and_unzip(exdir, zip_url)
+download_and_unzip(exdir, zip_url, data_dir = tempdir())
 }
 \arguments{
 \item{exdir}{Required zip name also used as destination of csv folder}
 
 \item{zip_url}{Required full path of file to download}
+
+\item{data_dir}{Parent directory of exdir}
 }
 \description{
 Downloads dftRoadSafety_2016_Accidents.zip to /tmp/blahDIR
