### 2018-11-29 3452104
Refactor read_* functions, closes #16
