### 2018-11-29 dd81e88
Refactor read_* functions, closes #16
