+++ man/read_accidents.Rd
@@ -8,7 +8,7 @@ read_accidents(filename = "", data_dir = tempdir(), years = "")
 }
 \arguments{
 \item{filename}{Character string of the filename of the .csv to read, if this is given, type and
-years determin whether there is a target to read, otherwise disk scan would be needed.}
+years determine whether there is a target to read, otherwise disk scan would be needed.}
 
 \item{data_dir}{Where sets of downloaded data would be found.}
 
