+++ man/check_input_file.Rd
@@ -1,31 +1,22 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/read_stats19.R
-\name{read_casualties}
-\alias{read_casualties}
-\title{Import and Stats19 data on casualties}
+\name{check_input_file}
+\alias{check_input_file}
+\title{Local helper to be reused.}
 \usage{
-read_casualties(filename = "", data_dir = tempdir(), years = "")
+check_input_file(filename = NULL, type = NULL, data_dir = NULL,
+  years = NULL)
 }
 \arguments{
 \item{filename}{Character string of the filename of the .csv to read, if this is given, type and
 years determine whether there is a target to read, otherwise disk scan would be needed.}
+
+\item{type}{One of 'Accidents', 'Casualties', 'Vehicles'; defaults to 'Accidents'#'}
 
 \item{data_dir}{Where sets of downloaded data would be found.}
 
 \item{years}{Either a single year or a two year range, defaults to 2 years ago}
 }
 \description{
-Import and Stats19 data on casualties
-}
-\section{Details}{
-
-The function returns a data frame, in which each record is a reported casualty
-in the stats19 dataset.
-}
-
-\examples{
-\dontrun{
-dl_stats19(years = 2017, type = "casualties")
-casualties = read_casualties()
-}
+Local helper to be reused.
 }
