+++ README.md
@@ -22,10 +22,8 @@ Install and attach the latest version with:
 
 ``` r
 devtools::install_github("ITSLeeds/stats19")
-#> Downloading GitHub repo ITSLeeds/stats19@master
-#> from URL https://api.github.com/repos/ITSLeeds/stats19/zipball/master
-#> Installing stats19
-#> ...
+#> Skipping install of 'stats19' from a github remote, the SHA1 (3452104a) has not changed since last install.
+#>   Use `force = TRUE` to force installation
 library(stats19)
 ```
 
@@ -45,7 +43,7 @@ dl_stats19(years = 2017, type = "Accidents")
 #> Attempt downloading from:
 #> http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/dftRoadSafetyData_Accidents_2017.zip
 #> happy to go (Y = enter, N = esc)?
-#> [1] "Data saved at: /tmp/RtmpC654bb/dftRoadSafetyData_Accidents_2017/Acc.csv"
+#> [1] "Data saved at: /var/folders/z7/l4z5fwqs2ksfv22ghh2n9smh0000gp/T//RtmpbgjuO5/dftRoadSafetyData_Accidents_2017/Acc.csv"
 ```
 
 Currently, these files are downloaded to a default location of "tempdir" which is a platform independent "safe" location to download the data in. Once downloaded, they are unzipped under original DfT file names. The function prints out the location and final file name(s) of unzipped files(s) as shown above.
@@ -77,16 +75,16 @@ dl_stats19(file_name = paste0(d17, ".zip"))
 #> http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/dftRoadSafetyData_Accidents_2017.zip
 #> happy to go (Y = enter, N = esc)?
 #> Data already exists in data_dir, not downloading
-#> [1] "Data saved at: /tmp/RtmpC654bb/dftRoadSafetyData_Accidents_2017/Acc.csv"
-crashes_2017_raw = read_accidents(data_dir = file.path(tempdir(), d17), filename = "Acc.csv")
+#> [1] "Data saved at: /var/folders/z7/l4z5fwqs2ksfv22ghh2n9smh0000gp/T//RtmpbgjuO5/dftRoadSafetyData_Accidents_2017/Acc.csv"
+crashes_2017_raw = read_accidents(years = 2017, filename = "Acc.csv")
 #> Warning: 20 parsing failures.
-#>   row       col expected actual                                                       file
-#> 45334 Longitude a double   NULL '/tmp/RtmpC654bb/dftRoadSafetyData_Accidents_2017/Acc.csv'
-#> 45334 Latitude  a double   NULL '/tmp/RtmpC654bb/dftRoadSafetyData_Accidents_2017/Acc.csv'
-#> 45721 Longitude a double   NULL '/tmp/RtmpC654bb/dftRoadSafetyData_Accidents_2017/Acc.csv'
-#> 45721 Latitude  a double   NULL '/tmp/RtmpC654bb/dftRoadSafetyData_Accidents_2017/Acc.csv'
-#> 46052 Longitude a double   NULL '/tmp/RtmpC654bb/dftRoadSafetyData_Accidents_2017/Acc.csv'
-#> ..... ......... ........ ...... ..........................................................
+#>   row       col expected actual                                                                                                    file
[…]
