### 2018-11-30 aa21ede
DESC file update, #17
