+++ DESCRIPTION
@@ -1,13 +1,14 @@
 Package: stats19
 Title: Work with open road traffic casualty data from Great Britain
-Version: 0.1.0.9000
+Version: 0.1
 Authors@R: c(
     person("Robin", "Lovelace", email = "rob00x@gmail.com", role = c("aut", "cre"),
     comment = c(ORCID = "0000-0001-5679-6536")),
     person("Malcolm", "Morgan", email = "m.morgan1@leeds.ac.uk", role = c("aut")),
-    person("Layik", "Hama", email = "layik.hama@gmail.com", role = c("aut"))
+    person("Layik", "Hama", email = "layik.hama@gmail.com", role = c("aut"),
+    comment = c(ORCID = "0000-0003-1912-4890"))
   )
-Description: Tools to help process and analysie the UK road collision data also known as Stats19.
+Description: Tools to help process and analysie the UK road collision data also known as STATS19.
 Depends: R (>= 3.5.0)
 License: GPL-3
 Encoding: UTF-8
