### 2018-11-30 32a61ab
Redundant generate_file_name, preps for #17
