+++ tests/testthat/test-utils.R
@@ -4,18 +4,6 @@ test_that("geturl works", {
   expect_equal(get_url(), "http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/")
 })
 
-test_that("generate_file_name works", {
-  expect_equal(generate_file_name(), "dftRoadSafety_Accidents_2016")
-  expect_equal(generate_file_name(zip = TRUE), "dftRoadSafety_Accidents_2016.zip")
-  expect_equal(generate_file_name(years = c("2005", "2009")),
-               "dftRoadSafety_Accidents_2005_2009")
-  expect_equal(generate_file_name(years = c("2009", "2005")),
-               "dftRoadSafety_Accidents_2005_2009")
-  expect_equal(generate_file_name(type = ""), "dftRoadSafety_2016")
-  expect_error(generate_file_name(years = c("2005", "2009", "2010")))
-
-})
-
 test_that("find_file_name works", {
   expect_equal(find_file_name(type = "accid"),
                "dftRoadSafety_Accidents_2016.zip")
