### 2018-11-30 c3e49ef
Redundant generate_file_name, preps for #17
