+++ man/generate_file_name.Rd
@@ -1,29 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/utils.R
-\name{generate_file_name}
-\alias{generate_file_name}
-\title{Zip file builder
-default "dftRoadSafety_Accidents_2016"}
-\usage{
-generate_file_name(years = as.integer(format(Sys.Date(), "\%Y")) - 2,
-  type = "Accidents", dft = "dft", zip = FALSE)
-}
-\arguments{
-\item{years}{Either a single year or a two year range, defaults to 2 years ago}
-
-\item{type}{One of 'Accidents', 'Casualties', 'Vehicles'; defaults to 'Accidents'}
-
-\item{dft}{The prefix with the file name 'dft'}
-
-\item{zip}{Add the '.zip' to the file name? Defaults to `FALSE`}
-}
-\description{
-Zip file builder
-default "dftRoadSafety_Accidents_2016"
-}
-\examples{
-\dontrun{
-generate_file_name()
-generate_file_name(zip = TRUE)
-}
-}
