### 2018-11-30 014d013
Check --as-cran, #17
