### 2018-11-30 dc5ca8e
Minor, rhub check 1, #17
