+++ tests/testthat/test-utils.R
@@ -1,7 +1,10 @@
 context("test-utils")
 
 test_that("geturl works", {
-  expect_equal(get_url(), "http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/")
+  expect_equal(get_url(),
+               file.path("http://data.dft.gov.uk.s3.amazonaws.com",
+                         "road-accidents-safety-data/")
+               )
 })
 
 test_that("find_file_name works", {
