### 2018-12-02 7d2bcb3
Shorten vignette
