### 2018-12-02 a2e1c66
Update package vignette
