+++ vignettes/stats19.Rmd
@@ -18,42 +18,53 @@ knitr::opts_chunk$set(
 )
 ```
 
-## 1. Introduction
+## Introduction
 
-**stats19** enables access to and processing of the UK's official road traffic casualty database, [STATS19](https://data.gov.uk/dataset/cb7ae6f0-4be6-4935-9277-47e5ce24a11f/road-safety-data).
+**stats19** enables access to and processing of the UK's official road traffic casualty database, [STATS19](https://data.gov.uk/dataset/cb7ae6f0-4be6-4935-9277-47e5ce24a11f/road-safety-data). 
+A description of variables in the database can be found in a [document](http://data.dft.gov.uk/road-accidents-safety-data/Brief-guide-to%20road-accidents-and-safety-data.doc) hosted by the UK's Department for Transport (DfT).
+The datasets are collectively called STATS19 after the form used to report them, which can be found [here](http://docs.adrn.ac.uk/888043/mrdoc/pdf/888043_stats19-road-accident-injury-statistics-report-form.pdf). 
+This vignette focusses on how to use the **stats19** package to work with STATS19 data.
 
-A description of the stats19 data and variables they contain can be found in a [document](http://data.dft.gov.uk/road-accidents-safety-data/Brief-guide-to%20road-accidents-and-safety-data.doc) hosted by the UK's Department for Transport (DfT).
-This vignette focusses on how to use this package to work with STATS19 data.
-
-The development version is hosted on [GitHub](https://github.com/ITSLeeds/stats19).
-It can be installed as follows:
+The development version is hosted on [GitHub](https://github.com/ITSLeeds/stats19) and can be installed and attached as follows:
 
 ```{r, eval=FALSE}
 devtools::install_github("ITSLeeds/stats19")
 ```
 
-### 1.1 package layout
-
-The work in this package has been divided into four main sections which will be discussed in detail in section (2) of this manuscript. Currently basics of `dl_stats19` which deals with downloading datasets, `read_*` functions and `format_` functions have been written. The aim is to also dedicated a section to analyzing the datasets. We also plan to combine the datasets with other datasets from [CDRC](www.cdrc.ac.uk) and other sources, but that may not be part of this package per se.
-
-### 1.2 outputs
-
-Robin is best placed for this part.
-
-## 2. Get stats19
-
-Install and attach the latest development version with:
-
-```{r install}
-# devtools::install_github("ITSLeeds/stats19")
+```{r}
 library(stats19)
 ```
 
-As stated in the introduction the package enables you to download road safety datasets published by the Department of Transport (DfT) of United Kingdom officially called [STATS19](https://data.gov.uk/dataset/cb7ae6f0-4be6-4935-9277-47e5ce24a11f/road-safety-data).
+The package has 3 main types of functions, corresponding with the main tasks undertaken by the package:
 
-### 2.1 Download files
+- Download functions beginning with `dl_`
+- Read functions beginning with `read_`
+- Format functions beginning with `format_`
 
-The main function to download individual datasets published by the DfT, is to use `dl_stats19`. The function takes three parameters. If `file_name` has been provided, this means the user is aware of what to download and the other two parameters will be ignored. Otherwise, you can use `years` and `type` to "search" through the file names which are also included in the package for convenience, called `stats19::file_names` as a named list.
+Multiple functions are needed for each step because of the structure of STATS19 data, which are divided into 3 tables, as described in an article that analyses 2005-2015 STATS19 data with reproducible R code by Christoph Freier ([2018](https://kawameicha.gitlab.io/inCodeWeTrust/2018/01/10-years-and-3.3-million-accidented-vehicles-in-the-uk/)):
+
+1. "accident circumstances, with details about location, severity, weather, etc; 
[…]
