### 2018-12-03 018ffdd
Finish vignette, close #8
