+++ vignettes/references.bib
@@ -10,3 +10,37 @@
   keywords = {Risk,geographical factors,corresponding author,risk,s institution,first author,university of sheffield,Cycling,safety,msc,robin lovelace,mr,exposure,Safety,Exposure,Geographical factors},
   file = {/home/robin/Zotero/storage/GDJH6AUM/S136984781500039X.html;/home/robin/Zotero/storage/XC7A9MA2/S136984781500039X.html},
 }
+
+@Article{edwards_relationship_1998,
+  title = {The {{Relationship Between Road Accident Severity}} and {{Recorded Weather}}},
+  volume = {29},
+  issn = {0022-4375},
+  doi = {10.1016/S0022-4375(98)00051-6},
+  abstract = {Road accident severity may be influenced by a number of factors. This article investigates the relationship between weather and road accidents in England and Wales. The weather information recorded on Police Accident Report Forms was taken as the prevailing weather at the time of the accident. At the local authority level, accident severity for the various adverse weather categories of rain, fog, and high winds is compared with the nonhazardous condition of fine weather. Severity ratios are then calculated. Findings establish that accident severity decreases significantly in rain compared with fine weather, while severity in fog shows geographical variation. Evidence for accident severity in high winds remains inconclusive.},
+  number = {4},
+  journal = {Journal of Safety Research},
+  author = {Julia B Edwards},
+  month = {dec},
+  year = {1998},
+  keywords = {Accident severity,fog,high winds,rain,weather},
+  pages = {249-262},
+  file = {/home/robin/Zotero/storage/R6ICNAUZ/S0022437598000516.html},
+}
+
+@Article{grundy_effect_2009,
+  title = {Effect of 20 Mph Traffic Speed Zones on Road Injuries in {{London}}, 1986-2006: Controlled Interrupted Time Series Analysis},
+  volume = {339},
+  copyright = {\textcopyright{} Grundy et al 2009. This is an open-access article distributed under the terms of the Creative Commons Attribution Non-commercial License, which permits use, distribution, and reproduction in any medium, provided the original work is properly cited, the use is non commercial and is otherwise in compliance with the license. See: http://creativecommons.org/licenses/by-nc/2.0/  and  http://creativecommons.org/licenses/by-nc/2.0/legalcode.},
+  issn = {0959-8138, 1468-5833},
+  shorttitle = {Effect of 20 Mph Traffic Speed Zones on Road Injuries in {{London}}, 1986-2006},
+  doi = {10.1136/bmj.b4469},
[…]
