### 2018-12-03 13786aa
Finish vignette, close #8
