### 2018-12-03 966fa42
Fix #19
