+++ README.md
@@ -31,40 +31,14 @@ The package builds on previous work including:
 
 ## Installation
 
-Install and attach the latest version with:
+Install and attach the latest version
+with:
 
 ``` r
 devtools::install_github("ITSLeeds/stats19")
-#> Downloading GitHub repo ITSLeeds/stats19@master
-#> Skipping 1 packages ahead of CRAN: sf
-#>   
-   checking for file ‘/tmp/Rtmp5TFmDP/remotes755e6b0c2ae9/ITSLeeds-stats19-753e0e6/DESCRIPTION’ ...
-  
-✔  checking for file ‘/tmp/Rtmp5TFmDP/remotes755e6b0c2ae9/ITSLeeds-stats19-753e0e6/DESCRIPTION’
-#> 
-  
-─  preparing ‘stats19’:
-#> 
-  
-   checking DESCRIPTION meta-information ...
-  
-✔  checking DESCRIPTION meta-information
-#> 
-  
-─  checking for LF line-endings in source and make files and shell scripts
-#> 
-  
-─  checking for empty or unneeded directories
-#> ─  looking to see if a ‘data/datalist’ file should be added
-#> 
-  
-─  building ‘stats19_0.1.tar.gz’
-#> 
-  
-   
-#> 
-#> Installing package into '/home/robin/R/x86_64-pc-linux-gnu-library/3.5'
-#> (as 'lib' is unspecified)
+```
+
+``` r
 library(stats19)
 ```
 
@@ -89,7 +63,7 @@ dl_stats19(years = 2017, type = "Accidents")
 #> Attempt downloading from:
 #> http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/dftRoadSafetyData_Accidents_2017.zip
 #> happy to go (Y = enter, N = esc)?
-#> [1] "Data saved at: /tmp/Rtmp5TFmDP/dftRoadSafetyData_Accidents_2017/Acc.csv"
+#> [1] "Data saved at: /tmp/Rtmp3zMtBR/dftRoadSafetyData_Accidents_2017/Acc.csv"
 ```
 
 Currently, these files are downloaded to a default location of “tempdir”
@@ -127,19 +101,10 @@ dl_stats19(file_name = paste0(d17, ".zip"))
 #> http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/dftRoadSafetyData_Accidents_2017.zip
 #> happy to go (Y = enter, N = esc)?
 #> Data already exists in data_dir, not downloading
-#> [1] "Data saved at: /tmp/Rtmp5TFmDP/dftRoadSafetyData_Accidents_2017/Acc.csv"
+#> [1] "Data saved at: /tmp/Rtmp3zMtBR/dftRoadSafetyData_Accidents_2017/Acc.csv"
 crashes_2017_raw = read_accidents(years = 2017, filename = "Acc.csv")
 #> Reading in:
-#> /tmp/Rtmp5TFmDP/dftRoadSafetyData_Accidents_2017/Acc.csv
-#> Warning: 20 parsing failures.
-#>   row       col expected actual                                                       file
-#> 45334 Longitude a double   NULL '/tmp/Rtmp5TFmDP/dftRoadSafetyData_Accidents_2017/Acc.csv'
-#> 45334 Latitude  a double   NULL '/tmp/Rtmp5TFmDP/dftRoadSafetyData_Accidents_2017/Acc.csv'
-#> 45721 Longitude a double   NULL '/tmp/Rtmp5TFmDP/dftRoadSafetyData_Accidents_2017/Acc.csv'
[…]
