+++ R/read_stats19.R
@@ -13,7 +13,8 @@
 #' @export
 #' @examples
 #' \dontrun{
-#' ac = read_accidents()
+#' dl_stats19(years = 2017, type = "Accidents")
+#' ac = read_accidents(years = "2017")
 #' }
 read_accidents = function(filename = "",
                           data_dir = tempdir(),
@@ -28,16 +29,21 @@ read_accidents = function(filename = "",
   message("Reading in: ")
   message(path)
   # read the data in
-  ac = readr::read_csv(path, col_types = readr::cols(
-    .default = readr::col_integer(),
-    Accident_Index = readr::col_character(),
-    Longitude = readr::col_double(),
-    Latitude = readr::col_double(),
-    Date = readr::col_character(),
-    Time = readr::col_character(),
-    `Local_Authority_(Highway)` = readr::col_character(),
-    LSOA_of_Accident_Location = readr::col_character()
-  ))
+  suppressWarnings({
+    ac = readr::read_csv(
+      path,
+      col_types = readr::cols(
+        .default = readr::col_integer(),
+        Accident_Index = readr::col_character(),
+        Longitude = readr::col_double(),
+        Latitude = readr::col_double(),
+        Date = readr::col_character(),
+        Time = readr::col_character(),
+        `Local_Authority_(Highway)` = readr::col_character(),
+        LSOA_of_Accident_Location = readr::col_character()
+      )
+    )
+  })
 
   ac
 
