+++ vignettes/references.bib
@@ -44,3 +44,20 @@
   file = {/home/robin/Zotero/storage/MFRGUFTA/Grundy et al. - 2009 - Effect of 20 mph traffic speed zones on road injur.pdf;/home/robin/Zotero/storage/UJV6CYNA/bmj.b4469.full.html},
   pmid = {20007666},
 }
+
+@Article{sarkar_street_2018,
+  title = {Street Morphology and Severity of Road Casualties: {{A}} 5-Year Study of {{Greater London}}},
+  volume = {12},
+  issn = {1556-8318},
+  shorttitle = {Street Morphology and Severity of Road Casualties},
+  doi = {10.1080/15568318.2017.1402972},
+  abstract = {Built environment factors, especially street-scale design and traffic casualties, are intrinsically interlinked. Starting from Alker Tripp's seminal ideas about city design, street morphology, and accident risk, this article summarizes results from an increasingly sophisticated line of enquiry at the boundaries between transport geography, network modeling, urban geography, and planning. It goes on to specify what we believe to be the most comprehensive study yet, based on five years' worth of road casualty data from London; GIS data on street morphology and physical features captured at a street-link unit of analysis; socio-economics and other determinants of accidents; and individual data about casualty victims. We test hypotheses about links between urban morphology and casualty severity using multi-level models with individual victim attributes at level-one, street-link morphology attributes (various measure of network connectivity) at level-two, and neighborhood descriptors at level-three.Results indicate that street-level morphology and design (expressed in terms of betweenness, divergence ratio, and hull radius), together with traffic volume and physical features of streets are all significantly associated with odds of ``Killed and Seriously Injured'' (KSI) causality incidents. We find the strongest evidence yet recorded that London's 20-MPH speed-restricted residential zones reduce the incidence of KSI; while neighborhood-level factors such as population density, deprivation of living environment, and access to services are also significant predictors of KSI indicating that selective urban territorial enclosure can save lives.},
+  number = {7},
+  journal = {International Journal of Sustainable Transportation},
+  author = {Chinmoy Sarkar and Chris Webster and Sarika Kumari},
+  month = {aug},
+  year = {2018},
+  keywords = {KSI,morphometrics,multilevel,STATS19,Traffic casualty severity,urban design},
+  pages = {510-525},
+  file = {/home/robin/Zotero/storage/CKF3JGXX/Sarkar et al. - 2018 - Street morphology and severity of road casualties.pdf;/home/robin/Zotero/storage/FNTF9I3H/15568318.2017.html},
+}
