### 2018-12-03 #266 @Robinlovelace
Heads-up @sckott `format_sf()` is now in there and the package now has a proper vignette. Are there any other things we should add at this stage for peer review? Thanks.
