### 2018-12-03 d7dfbd8
find_file_name: 1974 to 2004 semantic search addressing #21
