### 2018-12-03 8fff826
find_file_name: 1974 to 2004 semantic search addressing #21
