+++ R/utils.R
@@ -45,6 +45,18 @@ get_directory = function() {
 #' @export
 find_file_name = function(years =  as.integer(format(Sys.Date(), "%Y")) - 2,
                           type = "") {
+  # semantically accept years between 1974 and 2003 inclusive
+  # see https://github.com/ITSLeeds/stats19/issues/21
+  # TODO: not pure function, changing parameter value
+  # first OR last value are between 1974 and 2004
+  if((!is.na(as.integer(years[1])) &&
+      1974 <= as.integer(years[1]) && as.integer(years[1]) <= 2004) |
+     (!is.na(as.integer(years[length(years)])) &&
+      1974 <= as.integer(years[length(years)]) && as.integer(years[length(years)]) <= 2004)) {
+    if(as.integer(years[1]) != 2004)
+      # ignore subsequent years and reassignment is fine if is.element(2004, years)
+      years = 2004
+  }
   stopifnot(length(years) <= 2)
   file_names_vec = unlist(stats19::file_names, use.names = FALSE)
   result = file_names_vec[grep(years[1], file_names_vec, ignore.case = TRUE)]
