### 2018-12-03 a497390
Add warning for Stats19-Data1979-2004.zip
