+++ DESCRIPTION
@@ -11,6 +11,8 @@ Authors@R: c(
 Description: Tools to help process and analysie the UK road collision data also known as STATS19.
 Depends: R (>= 3.5.0)
 License: GPL-3
+URL: https://github.com/ITSLeeds/stats19
+BugReports: https://github.com/ITSLeeds/stats19/issues
 Encoding: UTF-8
 LazyData: true
 Imports: 
