### 2018-12-03 e8d3408
Improve style thanks to goodpractice
