### 2018-12-03 6a5327a
Improve style thanks to goodpractice
