+++ R/format.R
@@ -8,12 +8,9 @@
 #' @export
 #' @examples
 #' \dontrun{
+#' dl_stats19(years = 2017, type = "accident")
 #' x = read_accidents()
-#' sapply(x, class)
-#' table(x$Accident_Severity)
 #' crashes = format_accidents(x)
-#' sapply(crashes, class)
-#' table(crashes$accident_severity)
 #' }
 #' @export
 format_accidents = function(x, factorize = FALSE) {
@@ -37,16 +34,7 @@ format_accidents = function(x, factorize = FALSE) {
   # lkp$new_name[perfect_matches]
   # lkp$new_name[!perfect_matches]
 
-  # format 1 column for testing
-  # col_name_tmp = "accident_severity"
-  # sel_col = agrep(pattern = col_name_tmp, x = stats19_schema$variable, max.distance = 3)
-  # lookup_col_name = stats19_schema$variable[sel_col]
-  # lookup = stats19_schema[stats19_schema$variable == lookup_col_name, 1:2]
-  # x$accident_severity = lookup$label[match(x$accident_severity, lookup$code)]
-
   # doing it as a for loop for now as easier to debug - could convert to lapply
-  i = 1 # for testing
-  i = 6 # for testing
   for(i in vars_to_change) {
     # format 1 column for testing
     lookup_col_name = lkp$schema_variable[lkp$new_name == old_names[i]]
@@ -68,12 +56,9 @@ format_accidents = function(x, factorize = FALSE) {
 #' @export
 #' @examples
 #' \dontrun{
+#' dl_stats19(years = 2017, type = "casualties")
 #' x = read_casualties()
-#' sapply(x, class)
-#' table(x$Casualty_Severity)
 #' casualties = format_casualties(x)
-#' sapply(casualties, class)
-#' table(casualties$casualty_severity)
 #' }
 #' @export
 format_casualties = function(x, factorize = FALSE) {
@@ -95,9 +80,6 @@ format_casualties = function(x, factorize = FALSE) {
   message("Changing these variables: ", old_names[vkeep])
   message("Not changing these variables: ", old_names[!vkeep])
 
-  # # doing it as a for loop for now as easier to debug - could convert to lapply
-  # i = 1 # for testing
-  # i = 6 # for testing
   for(i in vars_to_change) {
     # format 1 column for testing
     lookup_col_name = lkp$schema_variable[lkp$new_name == old_names[i]]
@@ -119,12 +101,9 @@ format_casualties = function(x, factorize = FALSE) {
 #' @export
 #' @examples
 #' \dontrun{
-#' x = read_vehicles()
-#' sapply(x, class)
-#' table(x$Vehicle_Type)
+#' dl_stats19(years = 2017, type = "vehicles")
+#' vehicles_raw = read_vehicles()
 #' vehicles = format_vehicles(x)
-#' sapply(vehicles, class)
-#' table(vehicles$vehicle_type)
 #' }
 #' @export
 format_vehicles = function(x, factorize = FALSE) {
@@ -146,9 +125,6 @@ format_vehicles = function(x, factorize = FALSE) {
   message("Changing these variables: ", old_names[vkeep])
   message("Not changing these variables: ", old_names[!vkeep])
 
-  # # doing it as a for loop for now as easier to debug - could convert to lapply
-  # i = 1 # for testing
-  # i = 6 # for testing
   for(i in vars_to_change) {
     # format 1 column for testing
     lookup_col_name = lkp$schema_variable[lkp$new_name == old_names[i]]
@@ -200,18 +176,21 @@ read_schema = function(
   data_dir = tempdir(),
   filename = "Road-Accident-Safety-Data-Guide.xls",
   sheet = NULL
-  ) {
+) {
+
[…]
