### 2018-12-03 bbc096e
Remove 3 dependencies
