### 2018-12-05 427657f
Update style of @mpadge code
