+++ man/dl_stats19.Rd
@@ -4,8 +4,8 @@
 \alias{dl_stats19}
 \title{Download Stats19 data for a year or range of two years.}
 \usage{
-dl_stats19(file_name = NULL, years = "", type = "",
-  data_dir = tempdir())
+dl_stats19(file_name = NULL, years = as.integer(format(Sys.Date(),
+  "\%Y")) - 2, type = "Accidents", data_dir = tempdir())
 }
 \arguments{
 \item{file_name}{The file name to download, above two will be ignore.}
