+++ tests/testthat/test-dl_stats19.R
@@ -1,8 +1,8 @@
 context("test-dl_stats19")
 
 test_that("dl_stats19 works for junk", {
-  expect_output(dl_stats19(type = "junk"))
-  expect_output(dl_stats19(years = "2999", type = "junk"))
+  expect_error(dl_stats19(type = "junk"))
+  expect_error(dl_stats19(years = "2999", type = "junk"))
 })
 
 test_that("dl_stats19 works for default", {
