+++ man/download_and_unzip.Rd
@@ -13,6 +13,9 @@ download_and_unzip(exdir, zip_url, data_dir = tempdir())
 
 \item{data_dir}{Parent directory of exdir}
 }
+\value{
+Names of file added to `data_dir`
+}
 \description{
 Downloads dftRoadSafety_2016_Accidents.zip to /tmp/blahDIR
 Unzips dftRoadSafety_2016_Accidents.zip to
