+++ R/utils.R
@@ -30,11 +30,24 @@ get_directory = function() {
   directory
 }
 
-#' Find a file name within stats19::file_names.
+#' check and convert year argument
+check_year <- function (year) {
+  if (is.null (year) | length (year) > 1)
+    stop ("Please specify a single year between 1979 and ",
+          format (format(Sys.Date(), "%Y")))
+  year <- as.numeric (year)
+  current_year <- as.numeric (format(Sys.Date(), "%Y"))
+  if (!year %in% 1979:current_year)
+    stop ("Please provide a year between 1979 and ", current_year)
+
+  return (year)
+}
+
+#' Find file names within stats19::file_names.
 #'
 #' Currently, there are 52 file names to download data from.
 #'
-#' @param years Provide a year, part of it or maximum range of two years. Defaults to two years ago.
+#' @param years Years for which data are to be found
 #' @param type One of 'Accidents', 'Casualties', 'Vehicles'; defaults to 'Accidents', ignores case.
 #'
 #' @examples
@@ -43,20 +56,20 @@ get_directory = function() {
 #' find_file_name(1979)
 #' find_file_name(2016:2017)
 #' @export
-find_file_name = function(years =  as.integer(format(Sys.Date(), "%Y")) - 2,
-                          type = NULL) {
+find_file_name = function(years = NULL, type = NULL) {
 
-  current_year <- format(Sys.Date(), "%Y")
-  if (!any (years %in% 1979:current_year))
-    stop("stats19 data are only available from 1979 onwards")
+  years <- vapply (years, check_year, numeric (1))
   file_names_vec = unlist(stats19::file_names, use.names = FALSE)
   result <- NULL
   # see https://github.com/ITSLeeds/stats19/issues/21
   if (any (years %in% 1979:2004)) {
       result <- c (result, file_names_vec [grep ("1979", file_names_vec)])
   }
-  index = unlist(lapply(years, function(i) grep(i, file_names_vec,
-                                                ignore.case = TRUE)))
+  if (is.null (years))
+      index <- seq (file_names_vec)
+  else
+      index = unlist(lapply(years, function(i) grep(i, file_names_vec,
+                                                    ignore.case = TRUE)))
   result = c (result, file_names_vec[index])
 
   if (!is.null (type))
@@ -73,7 +86,7 @@ find_file_name = function(years =  as.integer(format(Sys.Date(), "%Y")) - 2,
 #' Helper function to locate files. Given below params, the function
 #' returns 0 or more files found at location/names given.
 #'
-#' @param years Provide a year, part of it or maximum range of two years.
+#' @param year Years for which data are to be found
 #' @param type One of 'Accidents', 'Casualties', 'Vehicles'; defaults to 'Accidents', ignores case.
 #' @param data_dir Super directory where dataset(s) were first downloaded to.
 #' @param quiet Print out messages (files found)
@@ -86,25 +99,33 @@ find_file_name = function(years =  as.integer(format(Sys.Date(), "%Y")) - 2,
 #' @export
 locate_files = function(data_dir = tempdir(),
                         type = "Accidents",
-                        years = as.integer(format(Sys.Date(), "%Y")) - 2,
+                        years = NULL,
                         quiet = FALSE) {
   stopifnot(dir.exists(data_dir))
   file_names = find_file_name(years = years, type = type)
   file_names <- tools::file_path_sans_ext (file_names)
   dir_files <- list.dirs (data_dir)
   files_on_disk <- NULL
-  if (any (grepl (file_names, dir_files)))
-    files_on_disk <- dir_files [grep (file_names, dir_files)]
+  # check is any file names match those on disk
+  gr <- vapply (file_names, function (i) any (grepl (i, dir_files)),
+                logical (1))
[…]
