+++ man/locate_files.Rd
@@ -4,17 +4,17 @@
 \alias{locate_files}
 \title{Locate a file on disk}
 \usage{
-locate_files(data_dir = tempdir(), type = "Accidents",
-  years = as.integer(format(Sys.Date(), "\%Y")) - 2, quiet = FALSE)
+locate_files(data_dir = tempdir(), type = "Accidents", years = NULL,
+  quiet = FALSE)
 }
 \arguments{
 \item{data_dir}{Super directory where dataset(s) were first downloaded to.}
 
 \item{type}{One of 'Accidents', 'Casualties', 'Vehicles'; defaults to 'Accidents', ignores case.}
 
-\item{years}{Provide a year, part of it or maximum range of two years.}
-
 \item{quiet}{Print out messages (files found)}
+
+\item{year}{Years for which data are to be found}
 }
 \value{
 Depending on @param return: full path of a single file found, list of directories
