+++ R/read_stats19.R
@@ -8,23 +8,24 @@
 #' @param filename Character string of the filename of the .csv to read, if this is given, type and
 #' years determine whether there is a target to read, otherwise disk scan would be needed.
 #' @param data_dir Where sets of downloaded data would be found.
-#' @param years Either a single year or a two year range, defaults to 2 years ago
+#' @param year Single year for which data are to be read
 #'
 #' @export
 #' @examples
 #' \dontrun{
-#' dl_stats19(years = 2017, type = "Accidents")
-#' ac = read_accidents(years = "2017")
+#' dl_stats19(year = 2017, type = "Accidents")
+#' ac = read_accidents(year = 2017)
 #' }
 read_accidents = function(filename = "",
                           data_dir = tempdir(),
-                          years = "") {
+                          year = NULL) {
   # check inputs
+  year <- check_year (year)
   path = check_input_file(
     filename = filename,
     type = "accidents",
     data_dir = data_dir,
-    years = years
+    year = year
   )
   message("Reading in: ")
   message(path)
@@ -48,32 +49,30 @@ read_accidents = function(filename = "",
   ac
 
 }
+
 #' Import and Stats19 data on vehicles
 #'
 #' @section Details:
 #' The function returns a data frame, in which each record is a reported vehicle in the
 #' stats19 dataset for the data_dir and filename provided.
 #'
-#'
-#' @param filename Character string of the filename of the .csv to read, if this is given, type and
-#' years determine whether there is a target to read, otherwise disk scan would be needed.
-#' @param data_dir Where sets of downloaded data would be found.
-#' @param years Either a single year or a two year range, defaults to 2 years ago
+#' @inheritParams read_accidents
 #'
 #' @export
 #' @examples
 #' \dontrun{
 #' ve = read_vehicles()
 #' }
-read_vehicles = function(filename = "",
+read_vehicles = function(filename = NULL,
                          data_dir = tempdir(),
-                         years = "") {
+                         year = NULL) {
   # check inputs
+  year <- check_year (year)
   path = check_input_file(
     filename = filename,
     type = "vehicles",
     data_dir = data_dir,
-    years = years
+    year = year
   )
   # read the data in
   ve = readr::read_csv(path, col_types = readr::cols(
@@ -82,16 +81,14 @@ read_vehicles = function(filename = "",
   ))
   ve
 }
+
 #' Import and Stats19 data on casualties
 #'
 #' @section Details:
 #' The function returns a data frame, in which each record is a reported casualty
 #' in the stats19 dataset.
 #'
-#' @param filename Character string of the filename of the .csv to read, if this is given, type and
-#' years determine whether there is a target to read, otherwise disk scan would be needed.
-#' @param data_dir Where sets of downloaded data would be found.
-#' @param years Either a single year or a two year range, defaults to 2 years ago
+#' @inheritParams read_accidents
 #'
 #' @export
 #' @examples
@@ -99,16 +96,17 @@ read_vehicles = function(filename = "",
 #' dl_stats19(years = 2017, type = "casualties")
 #' casualties = read_casualties()
 #' }
-read_casualties = function(filename = "",
+read_casualties = function(filename = NULL,
                            data_dir = tempdir(),
-                           years = "") {
+                           year = NULL) {
 
   # check inputs
+  year <- check_year (year)
   path = check_input_file(
     filename = filename,
     type = "casualties",
     data_dir = data_dir,
-    years = years
+    year = year
   )
   # read the data in
   ca = readr::read_csv(path, col_types = readr::cols(
@@ -123,25 +121,25 @@ read_casualties = function(filename = "",
 #' @param filename Character string of the filename of the .csv to read, if this is given, type and
[…]
