### 2018-12-05 859d9be
Polishing-up changes from @mpadge :shower:
