### 2018-12-05 2f1ed44
Polishing-up changes from @mpadge :shower:
