+++ man/dl_stats19.Rd
@@ -38,6 +38,6 @@ have a sufficient internet access and hard disk space.
 dl_stats19(year = 2017) # interactively select files...
 
 # now you can read-in the data
-dl_stats19(year = 2004)
+dl_stats19(year = 2009)
 }
 }
