+++ man/check_year.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/utils.R
+\name{check_year}
+\alias{check_year}
+\title{check and convert year argument}
+\usage{
+check_year(year)
+}
+\description{
+check and convert year argument
+}
