### 2018-12-05 8346f68
Minor fixes for travis
