+++ DESCRIPTION
@@ -6,7 +6,8 @@ Authors@R: c(
     comment = c(ORCID = "0000-0001-5679-6536")),
     person("Malcolm", "Morgan", email = "m.morgan1@leeds.ac.uk", role = c("aut")),
     person("Layik", "Hama", email = "layik.hama@gmail.com", role = c("aut"),
-    comment = c(ORCID = "0000-0003-1912-4890"))
+    comment = c(ORCID = "0000-0003-1912-4890")),
+    person("Mark", "Padgham", role=c("aut"))
   )
 Description: Tools to help process and analysie the UK road collision data also known as STATS19.
 Depends: R (>= 3.5.0)
