### 2018-12-05 327e436
style pkg with styler::equals_style
