### 2018-12-05 f0c2e82
style pkg with styler::equals_style
