### 2018-12-06 7781441
Update readme: fix code, add badges!
