+++ README.md
@@ -1,9 +1,15 @@
 
+[![The API of a maturing package has been roughed out, but finer details
+likely to
+change.](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)
 [![Travis build
 status](https://travis-ci.org/ITSLeeds/stats19.svg?branch=master)](https://travis-ci.org/ITSLeeds/stats19)
 [![codecov](https://codecov.io/gh/ITSLeeds/stats19/branch/master/graph/badge.svg)](https://codecov.io/gh/ITSLeeds/stats19)
 [![Gitter
 chat](https://badges.gitter.im/ITSLeeds/stats19.png)](https://gitter.im/stats19/Lobby?source=orgpage)
+[![](http://www.r-pkg.org/badges/version/stats19)](http://www.r-pkg.org/pkg/stats19)
+[![CRAN RStudio mirror
+downloads](http://cranlogs.r-pkg.org/badges/stats19)](http://www.r-pkg.org/pkg/stats19)
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
@@ -42,7 +48,8 @@ devtools::install_github("ITSLeeds/stats19")
 ``` r
 library(stats19)
 #> Data provided under the conditions of the Open Government License.
-#> If you use data accessed from this package, please acknowledge the Department for Transport and link to:
+#> If you use data from this package, mention the source
+#> (Department for Transport), cite the package and link to:
 #> www.nationalarchives.gov.uk/doc/open-government-licence/version/3/.
 ```
 
@@ -61,13 +68,11 @@ The following code chunk, for example, downloads and unzips a .zip file
 containing Stats19 data from 2017:
 
 ``` r
-dl_stats19(years = 2017, type = "Accidents")
-#> File to download:
-#> dftRoadSafetyData_Accidents_2017.zip
+dl_stats19(year = 2017, type = "Accidents")
+#> File to download: dftRoadSafetyData_Accidents_2017.zip
 #> Attempt downloading from:
-#> http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/dftRoadSafetyData_Accidents_2017.zip
-#> happy to go (Y = enter, N = esc)?
-#> [1] "Data saved at: /tmp/RtmpSz9JUm/dftRoadSafetyData_Accidents_2017/Acc.csv"
+#>    http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/dftRoadSafetyData_Accidents_2017.zip
+#> Data saved at /tmp/Rtmpkpois6/dftRoadSafetyData_Accidents_2017.zip/Acc.csv
 ```
 
 Currently, these files are downloaded to a default location of “tempdir”
@@ -81,14 +86,12 @@ providing just the year for example, would result in options presented
 to you:
 
 ``` r
-dl_stats19(years = 2017)
-#> More than one file found:
-#> Please type corresponding file number:
-#>     [1] dftRoadSafetyData_Vehicles_2017.zip
-#>     [2] dftRoadSafetyData_Casualties_2017.zip
-#>     [3] dftRoadSafetyData_Accidents_2017.zip
-#> 1 - 3:
-#> You made an invalid choice
+dl_stats19(year = 2017)
+#> File to download: dftRoadSafetyData_Accidents_2017.zip
+#> Attempt downloading from:
+#>    http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/dftRoadSafetyData_Accidents_2017.zip
+#> Data already exists in data_dir, not downloading
[…]
