### 2018-12-06 fc226cc
Update readme: fix code, add badges!
