### 2018-12-07 3673a57
Refactor check_year - cc @mpadge and @layik
