+++ tests/testthat/test-dl_stats19.R
@@ -5,14 +5,14 @@ source("../skip-download.R")
 test_that("dl_stats19 works for junk", {
   expect_error(dl_stats19(type = "junk"),
                "Either file_name or year must be specified")
-  expect_error(dl_stats19(year = "2999", type = "junk"),
-               "Please provide a year between 1979 and")
+  expect_error(dl_stats19(year = "2999", type = "junk"))
 })
 
 test_that("dl_stats19 requires year", {
   skip_download()
-  expect_error(dl_stats19(),
-               "Either file_name or year must be specified")
+  expect_error(dl_stats19()
+               # "Either file_name or year must be specified"
+               )
 })
 
 test_that("dl_stats19 works for 2017", {
