+++ man/check_year.Rd
@@ -12,3 +12,7 @@ check_year(year)
 \description{
 check and convert year argument
 }
+\examples{
+# check_year("2018") # fails
+# check_year(2017)
+}
