### 2018-12-07 9af56d2
Refactor check_year - cc @mpadge and @layik
