### 2018-12-07 8784063
Refactor get_url()
