+++ man/dl_stats19.Rd
@@ -39,5 +39,6 @@ dl_stats19(year = 2017) # interactively select files...
 
 # now you can read-in the data
 dl_stats19(year = 2009)
+dl_stats19(year = 1985)
 }
 }
