### 2018-12-07 bc1ed16
Refactor get_url()
