+++ man/get_domain.Rd
@@ -1,11 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/utils.R
-\name{get_domain}
-\alias{get_domain}
-\title{Return only the domain}
-\usage{
-get_domain()
-}
-\description{
-Return only the domain
-}
