### 2018-12-09 #266 @daranzolin
### Prolegomena

Thanks to rOpenSci for the invitation to review, and special thanks to Robin, Malcolm, Layik, and Mark for this important package. I'm already a huge fan.

***

> I would tease a bit more of what's in these data sets. I wasn't entirely sure until I downloaded and opened the supporting documentation. If I were searching for this kind of data, and I didn't know what STATS19 was, I'd like to know I'm in the right place after scanning the README. Maybe a map?

> I couldn't load the vignette from the console:

```
vignette(package = "stats19")
#> no vignettes found
```
> Admittedly, I also couldn't load the vignette for my own rOpenSci package, so I'm not sure what's missing.

> Several of the examples failed:

* `read_casualties` -- argument for `dl_stats19` should be `year` instead of `years`
* `read_vehicles` -- I think year must be specified and presupposes an existing download of Vehicles data, e.g. `ve = read_vehicles(year = 2017)`

```
stats19::read_vehicles()
#> Error in if (year %in% 1980:2003) {: argument is of length zero
```
* `check_input_file` -- no example included
* `format_accidents` -- same as above, argument for `dl_stats19` should be `year` instead of `years`
* `format_casualties` -- same as above, argument for `dl_stats19` should be `year` instead of `years`
* `format_vehicles` -- same as above, argument for `dl_stats19` should be `year` instead of `years`
* `format_sf` -- `read_accidents` requires a year argument

> I couldn't find any explicit contributing guidelines in the README, and there is no CONTRIBUTING document.

##### For packages co-submitting to JOSS

- [X] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)

> There is no paper.md.

> Test results:
```
══ Results ═════════════════════════════════════════════════════════════════════════════════════════════════

OK:       37
```

---

> A superb and essential package--we need this data and we need it in these formats. The download-format-read-explore workflow is intuitive and relatively frictionless. I have only some brief comments:

* I wonder you could possibly merge the formatting and reading step with a `raw = TRUE` or `format = TRUE` argument in the `read_*` functions. But perhaps that's my tendency towards abstraction. Something like `ac = read_accidents(year = 2017, format = TRUE)`

* My personal preference would be to have the schema from `dl_schema` lazily loaded with the package.

* [According to the vignette,](https://github.com/ITSLeeds/stats19/blob/master/vignettes/stats19.Rmd) the `dl_*` functions are interactive, although the interactivity is commented out in the code. Will the interactivity be returning? Or does the vignette need to be updated?

* Out of curiosity, what's happening with https://github.com/cyipt/stats19? It was updated recently.

* I confess I wish the package name was more expressive--stats19 sounds like an introductory statistics class.

* This data will be used to make many maps. I personally would love a nudge in that direction in either the README or the vignette.
