### 2018-12-11 #266 @mpadge
Yeah, thanks @daranzolin for an impressively prompt review. I'll leave the response proper to @layik and @Robinlovelace, but will confess guilty here to having commented out the interactivity in the `dl_*` functions. That's all my doing and I should already have reinstated it. Apologies to all, and yes, it will be returning.
