### 2018-12-11 #266 @Robinlovelace
Many thanks for the quick yet detailed review @daranzolin. I'm working on it this week and will get back to you on each of those constructive comments asap. More soon...
