### 2018-12-11 257c9b3
Refactor dl utils
prepare for interactivity (again...)
