### 2018-12-11 b671633
Refactor dl utils
prepare for interactivity (again...)
