+++ man/phrase.Rd
@@ -1,0 +1,14 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/dl_stats19.R
+\name{phrase}
+\alias{phrase}
+\title{Generate a phrase for data download purposes}
+\usage{
+phrase(data_dir)
+}
+\description{
+Generate a phrase for data download purposes
+}
+\examples{
+stats19:::phrase(tempdir())
+}
