### 2018-12-11 10528ac
Improve dox, fix test
