+++ tests/testthat/test-dl_stats19.R
@@ -3,8 +3,7 @@ context("test-dl_stats19")
 source("../skip-download.R")
 
 test_that("dl_stats19 works for junk", {
-  expect_error(dl_stats19(type = "junk"),
-               "Either file_name or year must be specified")
+  expect_error(dl_stats19(type = "junk"))
   expect_error(dl_stats19(year = "2999", type = "junk"))
 })
 
