+++ man/phrase.Rd
@@ -6,6 +6,9 @@
 \usage{
 phrase(data_dir)
 }
+\arguments{
+\item{data_dir}{Parent directory for all downloaded files. Defaults to `tempdir()`}
+}
 \description{
 Generate a phrase for data download purposes
 }
