### 2018-12-11 fa783e3
Improve dox, fix test
