### 2018-12-11 64a1d13
Add interactivity - select_file()
