+++ R/dl_stats19.R
@@ -41,8 +41,12 @@ dl_stats19 = function(year = NULL,
     fnames = find_file_name(years = year, type = type)
     # todo: add menu here...
     if(length(fnames) > 1) {
-      message("Heads-up: more than one file found, select one:")
-      fnames = fnames[1]
+      if(interactive()) {
+        fnames = select_file(fnames)
+      } else {
+        message("More than one file found, selecting the first.")
+        fnames = fnames[1]
+      }
     }
     zip_url = get_url(fnames) # no need for the .zip here
   } else {
@@ -146,4 +150,15 @@ dl_schema = function(data_dir = tempdir()) {
   destfile = file.path(data_dir, "Road-Accident-Safety-Data-Guide.xls")
   utils::download.file(u, destfile = destfile)
   # download and unzip the data if it's not present
+}
+
+#' Interactively select from options
+#' @param fnames File names to select from
+#' @examples
+#' # fnames = c("f1", "f2")
+#' # stats19:::select_file(fnames)
+select_file = function(fnames) {
+  message("Multiple matches. Which do you want to download?")
+  selection = utils::menu(choices = fnames)
+  fnames[selection]
 }
