### 2018-12-11 #266 @Robinlovelace
Heads-up all (especially @daranzolin, @mpadge and @layik - the originator of interactive downloads in **stats19**), I've re-added interactivity, this time using `menu()` which is best practice for this kind of thing and fewer lines of code.

Please check out the video to see how it works here:

[![asciicast](https://asciinema.org/a/4yhLMS8B256KypvEqpKPl29Fs.svg)](https://asciinema.org/a/4yhLMS8B256KypvEqpKPl29Fs)

And if you're interested in the code used to do this, see here: https://github.com/ITSLeeds/stats19/pull/36

Feedback on behaviour or code welcome before I merge this. Many thanks @layik for implementing this 1st time around.
