+++ man/match_type.Rd
@@ -1,18 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/dl_stats19.R
-\name{match_type}
-\alias{match_type}
-\title{Match type to types provided by stats19}
-\usage{
-match_type(type)
-}
-\arguments{
-\item{type}{One of 'Accidents', 'Casualties', 'Vehicles'; defaults to 'Accidents'#'}
-}
-\description{
-Match type to types provided by stats19
-}
-\examples{
-stats19:::match_type(type = "accidents")
-stats19:::match_type(type = "nomatch")
-}
