+++ man/dl_stats19.Rd
@@ -36,10 +36,10 @@ have a sufficient internet access and hard disk space.
 \examples{
 \dontrun{
 dl_stats19(year = 2017) # interactively select files...
-
 # now you can read-in the data
 dl_stats19(year = 2009)
 dl_stats19(year = 2009, type = "casualties")
+dl_stats19(type = "casualties")
 dl_stats19(year = 1985)
 }
 }
