+++ man/find_file_name.Rd
@@ -18,6 +18,8 @@ Currently, there are 52 file names to download data from.
 find_file_name(2016)
 find_file_name(2016, type = "Accidents")
 find_file_name(1985, type = "Accidents")
+find_file_name(type = "cas")
+find_file_name(type = "accid")
 find_file_name(1979)
 find_file_name(2016:2017)
 }
