### 2018-12-11 edbe39e
Fix #38
