+++ tests/testthat/test-utils.R
@@ -44,11 +44,10 @@ test_that("locate_files & locate_one_file works", {
   x2 = locate_one_file(filename = "Cas.csv", year = 2017, type = "cas")
   expect_true(length(x1) == 1)
   # more tests on locate_files
-  expect_error(locate_files(), "No files of that type exist")
+  expect_error(locate_files())
   # from clean start
   unlink(tempdir(), recursive = TRUE)
   dir.create(tempdir())
-  expect_error(locate_files(), "No files of that type exist")
   expect_error(locate_files(data_dir = "/junking"),
                "dir.exists\\(data_dir\\) is not TRUE")
 })
