### 2018-12-11 3f1b3fd
Fix typo - return 2005:2014 data for 2006
