### 2018-12-11 a6c606c
Fix for tests
