+++ man/check_year.Rd
@@ -15,5 +15,6 @@ check and convert year argument
 \examples{
 # check_year("2018") # fails
 # check_year(2017)
+# check_year(2006)
 # check_year(1985)
 }
