### 2018-12-11 b376a4b
Fix for tests
