+++ tests/testthat/test-utils.R
@@ -17,9 +17,10 @@ test_that("find_file_name works", {
                "RoadSafetyData_Accidents_2015.zip")
   expect_equal(find_file_name(years = c("2005", "2006")),
                "Stats19_Data_2005-2014.zip")
-  expect_silent(find_file_name(years = c("2009", "2014", "2015")))
-  expect_equal(length (find_file_name(years = c("2009", "2014", "2015"),
-                                      type = "Accident")), 4)
+  expect_silent(find_file_name(years = c("2014", "2015")))
+  expect_message(find_file_name(years = 2009))
+  expect_equal(length(find_file_name(years = c("2014", "2015"),
+                                      type = "Accident")), 2)
   # cover https://github.com/ITSLeeds/stats19/issues/21
   expect_equal (length (find_file_name(years = 1980:2001)), 1)
   # start OR end year is between 74 and 04
