+++ man/find_file_name.Rd
@@ -20,6 +20,6 @@ find_file_name(2016, type = "Accidents")
 find_file_name(1985, type = "Accidents")
 find_file_name(type = "cas")
 find_file_name(type = "accid")
-find_file_name(1979)
+find_file_name(2006)
 find_file_name(2016:2017)
 }
