+++ README.md
@@ -69,10 +69,10 @@ containing Stats19 data from 2017:
 
 ``` r
 dl_stats19(year = 2017, type = "Accidents")
-#> File to download: dftRoadSafetyData_Accidents_2017.zip
+#> Files identified: dftRoadSafetyData_Accidents_2017.zip
 #> Attempt downloading from:
 #>    http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/dftRoadSafetyData_Accidents_2017.zip
-#> Data saved at /tmp/Rtmpkpois6/dftRoadSafetyData_Accidents_2017.zip/Acc.csv
+#> Data saved at /tmp/RtmpXdUCj8/dftRoadSafetyData_Accidents_2017.zip/Acc.csv
 ```
 
 Currently, these files are downloaded to a default location of “tempdir”
@@ -87,12 +87,16 @@ to you:
 
 ``` r
 dl_stats19(year = 2017)
-#> File to download: dftRoadSafetyData_Accidents_2017.zip
-#> Attempt downloading from:
-#>    http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/dftRoadSafetyData_Accidents_2017.zip
-#> Data already exists in data_dir, not downloading
-#> Data saved at /tmp/Rtmpkpois6/dftRoadSafetyData_Accidents_2017.zip/Acc.csv
 ```
+
+    Multiple matches. Which do you want to download?
+    
+    1: dftRoadSafetyData_Vehicles_2017.zip
+    2: dftRoadSafetyData_Casualties_2017.zip
+    3: dftRoadSafetyData_Accidents_2017.zip
+    
+    Selection: 
+    Enter an item from the menu, or 0 to exit
 
 ## Reading-in data
 
@@ -102,14 +106,14 @@ download went OK):
 ``` r
 d17 = "dftRoadSafetyData_Accidents_2017"
 dl_stats19(file_name = paste0(d17, ".zip"))
-#> File to download: dftRoadSafetyData_Accidents_2017.zip
+#> Files identified: dftRoadSafetyData_Accidents_2017.zip
 #> Attempt downloading from:
 #>    http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/dftRoadSafetyData_Accidents_2017.zip
 #> Data already exists in data_dir, not downloading
-#> Data saved at /tmp/Rtmpkpois6/dftRoadSafetyData_Accidents_2017.zip/Acc.csv
+#> Data saved at /tmp/RtmpXdUCj8/dftRoadSafetyData_Accidents_2017.zip/Acc.csv
 crashes_2017_raw = read_accidents(year = 2017, filename = "Acc.csv")
 #> Reading in:
-#> /tmp/Rtmpkpois6/dftRoadSafetyData_Accidents_2017/Acc.csv
+#> /tmp/RtmpXdUCj8/dftRoadSafetyData_Accidents_2017/Acc.csv
 crashes_2017 = format_accidents(crashes_2017_raw)
 ```
 
@@ -132,16 +136,16 @@ crashes_2017_raw[random_n, key_vars]
 #> # A tibble: 3 x 4
 #>   Accident_Severity Speed_limit `Pedestrian_Crossing-Hum… Light_Conditions
 #>               <int>       <int>                     <int>            <int>
-#> 1                 3          40                         0                4
-#> 2                 3          30                         0                4
-#> 3                 2          60                         0                1
+#> 1                 2          30                         0                4
+#> 2                 3          30                         0                1
+#> 3                 3          20                         0                4
 crashes_2017[random_n, key_vars]
 #> # A tibble: 3 x 4
 #>   accident_severity speed_limit pedestrian_crossing_hu… light_conditions  
[…]
