### 2018-12-12 c566682
More test coverage per #30
