### 2018-12-12 ITSLeeds/stats19#11 @layik
@Robinlovelace can we shelve this please? There will be lots of ways people would want to "get" stats19. I suggest we let users re-open/reference this issue if they need our help.
