### 2018-12-12 ITSLeeds/stats19#11 @Robinlovelace
Sure - not a priority. Pending feedback from  https://github.com/ropensci/onboarding/issues/266 I'm happy for it to be shelved.
