### 2018-12-12 c8df4e7
Move utils funs into utils.R
