+++ man/check_input_file.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/read_stats19.R
+% Please edit documentation in R/read.R
 \name{check_input_file}
 \alias{check_input_file}
 \title{Local helper to be reused.}
