### 2018-12-12 1c4840d
Generalise tests - #41
