### 2018-12-12 3c652a9
Code cleanup pair programming with Robin.
