### 2018-12-12 c14f458
Code cleanup pair programming with Robin.
