+++ man/vehicles_sample.Rd
@@ -1,0 +1,23 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{vehicles_sample}
+\alias{vehicles_sample}
+\alias{vehicles_sample_raw}
+\title{Sample of stats19 data (2017 vehicles)}
+\format{A data frame}
+\description{
+Sample of stats19 data (2017 vehicles)
+}
+\examples{
+\dontrun{
+# Obtained with:
+dl_stats19(year = 2017, type = "veh")
+vehicles_2017_raw = read_vehicles(year = 2017)
+set.seed(350)
+sel = sample(nrow(vehicles_2017_raw), 3)
+vehicles_sample_raw = vehicles_2017_raw[sel, ]
+vehicles_sample = format_vehicles(vehicles_sample_raw)
+}
+}
+\keyword{datasets}
