+++ man/casualties_sample.Rd
@@ -1,0 +1,23 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{casualties_sample}
+\alias{casualties_sample}
+\alias{casualties_sample_raw}
+\title{Sample of stats19 data (2017 casualties)}
+\format{A data frame}
+\description{
+Sample of stats19 data (2017 casualties)
+}
+\examples{
+\dontrun{
+# Obtained with:
+dl_stats19(year = 2017, type = "cas")
+casualties_2017_raw = read_casualties(year = 2017)
+set.seed(350)
+sel = sample(nrow(casualties_2017_raw), 3)
+casualties_sample_raw = casualties_2017_raw[sel, ]
+casualties_sample = format_casualties(casualties_sample_raw)
+}
+}
+\keyword{datasets}
