### 2018-12-12 31ed208
Begin refactor of format code
