### 2018-12-12 5707d26
Make schema a tibble, close #40
