### 2018-12-12 ef19881
Update schema - add formatted col names
