### 2018-12-12 4b56d44
Refactor factor code
fix format_vehicles()
