### 2018-12-12 a650773
Refactor factor code
fix format_vehicles()
