### 2018-12-12 6680933
Complete refactoring mission
