+++ R/format.R
@@ -7,40 +7,30 @@
 #' @export
 #' @examples
 #' \dontrun{
-#' dl_stats19(years = 2017, type = "accident")
-#' x = read_accidents()
+#' dl_stats19(year = 2017, type = "accident")
+#' x = read_accidents(year = 2017)
 #' crashes = format_accidents(x)
 #' }
 #' @export
 format_accidents = function(x) {
+
+  # rename colums
   old_names = names(x)
   new_names = format_column_names(old_names)
   names(x) = new_names
 
   # create lookup table
-  lkp = stats19_variables[stats19_variables$table == "accidents", ]
-  lkp = lkp[lkp$type == "character", ]
-  lkp$schema_variable = stats19_vname_switch(lkp$variable)
-  lkp$new_name = gsub(pattern = " ", replacement = "_", lkp$schema_variable)
-  lkp$new_name = stats19_vname_raw(lkp$new_name)
+  lkp = stats19_variables[stats19_variables$table == "accidents",]
 
-  vars_to_change = which(old_names %in% lkp$new_name)
-  # old_names[vars_to_change]
+  vkeep = new_names %in% stats19_schema$variable_formatted # new way with summary(vkeep) 12 true
+  vars_to_change = which(vkeep)
 
-  # # testing:
-  # perfect_matches = lkp$new_name %in% old_names
-  # summary(perfect_matches)
-  # lkp$new_name[perfect_matches]
-  # lkp$new_name[!perfect_matches]
+  # # testing: - todo: remove these when function works for all vars
+  message("Keeping these variables: ", paste(old_names[!vkeep], collapse = ", "))
 
-  # doing it as a for loop for now as easier to debug - could convert to lapply
   for(i in vars_to_change) {
-    # format 1 column for testing
-    lookup_col_name = lkp$schema_variable[lkp$new_name == old_names[i]]
-    lookup = stats19_schema[stats19_schema$variable == lookup_col_name, 1:2]
-    if(length(lookup_col_name) != 1) {
-      message("No single match for ", lookup_col_name)
-    }
+    lookup_col_name = lkp$column_name[lkp$column_name == new_names[i]]
+    lookup = stats19_schema[stats19_schema$variable_formatted == lookup_col_name, 1:2]
     x[[i]] = lookup$label[match(x[[i]], lookup$code)]
   }
   x
@@ -66,22 +56,15 @@ format_casualties = function(x) {
 
   # create lookup table
   lkp = stats19_variables[stats19_variables$table == "casualties", ]
-  lkp = lkp[lkp$type == "character", ] # old solution
-
-  lkp$schema_variable = stats19_vname_switch(lkp$variable)
-  lkp$new_name = gsub(pattern = " ", replacement = "_", lkp$schema_variable)
-  lkp$new_name = stats19_vname_raw(lkp$new_name)
 
   # vkeep = old_names %in% lkp$new_name # old way with summary(vkeep) 5 true
-  vkeep = new_names %in% lkp$column_name # new way with summary(vkeep) 12 true
+  vkeep = new_names %in% stats19_schema$variable_formatted # new way with summary(vkeep) 12 true
   vars_to_change = which(vkeep)
 
   # # testing: - todo: remove these when function works for all vars
-  # message("Changing these variables: ", paste(old_names[vkeep], collapse = ", "))
-  message("Changing these variables: ", paste(old_names[!vkeep], collapse = ", "))
+  message("Keeping these variables: ", paste(old_names[!vkeep], collapse = ", "))
 
   for(i in vars_to_change) {
-    # format 1 column for testing
     lookup_col_name = lkp$column_name[lkp$column_name == new_names[i]]
[…]
