+++ man/format_vehicles.Rd
@@ -19,8 +19,8 @@ This function formats raw stats19 data
 
 \examples{
 \dontrun{
-dl_stats19(years = 2017, type = "vehicles")
-vehicles_raw = read_vehicles()
+dl_stats19(year = 2017, type = "vehicles")
+x = read_vehicles(year = 2017)
 vehicles = format_vehicles(x)
 }
 }
