### 2018-12-12 8f15a83
Complete refactoring mission
