+++ man/format_column_names.Rd
@@ -19,7 +19,7 @@ such as spaces and `-` are returned.
 }
 \examples{
 \dontrun{
-crashes_raw = read_accidents()
+crashes_raw = read_accidents(year = 2017)
 column_names = names(crashes_raw)
 column_names
 format_column_names(column_names = column_names)
