### 2018-12-12 748881a
Tidy-up, rebuild README
