### 2018-12-14 2891de2
Refactor formatting code - format_stats19()
