### 2018-12-17 49d2b28
Remove extra checks and add necessary ones, part of #44
