+++ R/read.R
@@ -19,9 +19,6 @@
 read_accidents = function(filename = "",
                           data_dir = tempdir(),
                           year = NULL) {
-  # check inputs
-  if (!is.null (year))
-    year = check_year (year)
   path = check_input_file(
     filename = filename,
     type = "accidents",
@@ -68,7 +65,6 @@ read_vehicles = function(filename = NULL,
                          data_dir = tempdir(),
                          year = NULL) {
   # check inputs
-  year = check_year (year)
   path = check_input_file(
     filename = filename,
     type = "vehicles",
@@ -102,7 +98,6 @@ read_casualties = function(filename = NULL,
                            year = NULL) {
 
   # check inputs
-  year = check_year (year)
   path = check_input_file(
     filename = filename,
     type = "casualties",
@@ -129,7 +124,8 @@ check_input_file = function(filename = NULL,
                             type = NULL,
                             data_dir = NULL,
                             year = NULL) {
-  year = check_year (year)
+  if (!is.null (year))
+    year = check_year (year)
   path = locate_one_file(
     type = type,
     filename = filename,
