### 2018-12-17 a4a1907
Adding test cases to close #44
