### 2018-12-17 6546738
Replace JS code :( with R and close #23
