### 2018-12-17 #266 @layik
Just going through the review @daranzolin, RE vignette, I think what you needed to do was
```r
devtools::install(build_vignettes = TRUE)
```
then you would be able to load the vignette.
