### 2018-12-18 4bc5637
PR to manage locate_one_file in #44, not a fix but clarity and consistency
