+++ R/read.R
@@ -133,10 +133,12 @@ check_input_file = function(filename = NULL,
     year = year
   )
   # have we NOT found a csv to read?
-  if (!endsWith(path, ".csv") | !file.exists(path)) {
-    # locate_files malfunctioned or just foo/bar path returned with no filename
+  if (length(path) != 1 | !endsWith(path, ".csv") | !file.exists(path)) {
+    # locate_files malfunctioned or just path returned with no filename
     message(path)
     stop("Change data_dir, filename, year or run dl_stats19() first.")
   }
+  if(identical(path, "More than one csv file found."))
+    stop("Your parameters return identical filenames under different directories.")
   return(path)
 }
