### 2018-12-18 68a88f4
PR to manage locate_one_file in #44, not a fix but clarity and consistency
