### 2018-12-18 df7205a
Minor change as per review by @robinlovelace
