### 2018-12-18 cd5709f
Make locate_files internal closes #47
