### 2018-12-18 437e9a0
Make locate_files internal closes #47
