+++ NAMESPACE
@@ -8,7 +8,6 @@ export(format_casualties)
 export(format_column_names)
 export(format_sf)
 export(format_vehicles)
-export(locate_files)
 export(locate_one_file)
 export(read_accidents)
 export(read_casualties)
