+++ man/read_accidents.Rd
@@ -4,15 +4,15 @@
 \alias{read_accidents}
 \title{Import and Stats19 data on road traffic casualties}
 \usage{
-read_accidents(filename = "", data_dir = tempdir(), year = NULL)
+read_accidents(year = NULL, filename = "", data_dir = tempdir())
 }
 \arguments{
+\item{year}{Single year for which data are to be read}
+
 \item{filename}{Character string of the filename of the .csv to read, if this is given, type and
 years determine whether there is a target to read, otherwise disk scan would be needed.}
 
 \item{data_dir}{Where sets of downloaded data would be found.}
-
-\item{year}{Single year for which data are to be read}
 }
 \description{
 Import and Stats19 data on road traffic casualties
