### 2018-12-19 e6eb7a4
Close #48
