### 2018-12-19 ITSLeeds/stats19#49 @layik
Review responses checklist
- [x] add a map and a time series plot showing the scale of the data. Also show a sample of the additional casualty and vehicle tables.
- [x] paper.md
- [x] package name
- [x] review roxygen2 docs
- [x] reply to the remaining comments on responses1.Rmd
- [x] update the vignette in light of changes
Note: items could be added/removed any time.
