### 2018-12-19 3a100ab
Basic paper.md, link #49
