+++ paper.md
@@ -1,0 +1,34 @@
+---
+title: 'stats19: An R package to consume DfT road safety data.'
+tags:
+  - stats19
+  - crashes
+  - dft
+  - road safety
+authors:
+ - name: R Lovelace
+   orcid: 0000-0001-5679-6536
+   affiliation: 1
+ - name: M Morgan
+   affiliation: 1
+ - name: L hama
+   affiliations: 1
+ - name: M Padgam
+   affiliations: 2
+affiliations:
+ - name: ITS/LIDA, University of Leeds
+   index: 1
+ - name: ATFutures GmbH.
+   index: 2
+date: 19 December 2018
+---
+
+# Summary
+
+This package simplifies downloading, reading and formatting DfT (Department for Transport), UK gov datasets on road safety from 1979 until present. 
+
+# Statement
+
+The aim of the package is to save R users to download the `zip` files released by the DfT and convert them into clean `df` R objects ready for processing. We have used the package internally to look at the state of the road safety and generate various maps using inhouse and R mapping packages. Therefore, although it is developed with road safety relevant people in mind, it can also be used by academics and educators as the basis for teaching.
+
+
