### 2018-12-19 29fd94e
Basic paper.md, link #49
