### 2018-12-19 c7e52dc
Minor change to authors paper.md
