+++ paper.md
@@ -11,7 +11,8 @@ authors:
    affiliation: 1
  - name: M Morgan
    affiliation: 1
- - name: L hama
+ - name: L Hama
+   orcid: 0000-0003-1912-4890
    affiliations: 1
  - name: M Padgam
    affiliations: 2
