### 2018-12-19 e9d81bb
Update readme ready for paper updates
