### 2018-12-19 5a1f62e
Update readme ready for paper updates
