### 2018-12-19 904488b
Update paper.md
