+++ paper.bib
@@ -1,0 +1,38 @@
+@Article{lovelace_stplanr_Inpress,
+  title = {Stplanr: {{A Package}} for {{Transport Planning}}},
+  abstract = {stplanr - R package providing functions and data access for transport research},
+  journal = {The R Journal},
+  author = {Robin Lovelace and Richard Ellison},
+  year = {In press},
+  file = {/home/robin/Zotero/storage/AEGWQPFR/stplanr-paper.html},
+}
+
+@Article{lovelace_who_2016,
+  series = {Bicycling and bicycle safety},
+  title = {Who, Where, When: The Demographic and Geographic Distribution of Bicycle Crashes in {{West Yorkshire}}},
+  volume = {41, Part B},
+  issn = {13698478},
+  doi = {10.1016/j.trf.2015.02.010},
+  journal = {Transportation Research Part F: Traffic Psychology and Behaviour},
+  author = {Robin Lovelace and Hannah Roberts and Ian Kellar},
+  year = {2016},
+  keywords = {Risk,geographical factors,corresponding author,risk,s institution,first author,university of sheffield,Cycling,safety,msc,robin lovelace,mr,exposure,Safety,Exposure,Geographical factors},
+  file = {/home/robin/Zotero/storage/GDJH6AUM/S136984781500039X.html;/home/robin/Zotero/storage/XC7A9MA2/S136984781500039X.html},
+}
+
+@Article{sarkar_street_2018,
+  title = {Street Morphology and Severity of Road Casualties: {{A}} 5-Year Study of {{Greater London}}},
+  volume = {12},
+  issn = {1556-8318},
+  shorttitle = {Street Morphology and Severity of Road Casualties},
+  doi = {10.1080/15568318.2017.1402972},
+  abstract = {Built environment factors, especially street-scale design and traffic casualties, are intrinsically interlinked. Starting from Alker Tripp's seminal ideas about city design, street morphology, and accident risk, this article summarizes results from an increasingly sophisticated line of enquiry at the boundaries between transport geography, network modeling, urban geography, and planning. It goes on to specify what we believe to be the most comprehensive study yet, based on five years' worth of road casualty data from London; GIS data on street morphology and physical features captured at a street-link unit of analysis; socio-economics and other determinants of accidents; and individual data about casualty victims. We test hypotheses about links between urban morphology and casualty severity using multi-level models with individual victim attributes at level-one, street-link morphology attributes (various measure of network connectivity) at level-two, and neighborhood descriptors at level-three.Results indicate that street-level morphology and design (expressed in terms of betweenness, divergence ratio, and hull radius), together with traffic volume and physical features of streets are all significantly associated with odds of ``Killed and Seriously Injured'' (KSI) causality incidents. We find the strongest evidence yet recorded that London's 20-MPH speed-restricted residential zones reduce the incidence of KSI; while neighborhood-level factors such as population density, deprivation of living environment, and access to services are also significant predictors of KSI indicating that selective urban territorial enclosure can save lives.},
+  number = {7},
+  journal = {International Journal of Sustainable Transportation},
+  author = {Chinmoy Sarkar and Chris Webster and Sarika Kumari},
+  month = {aug},
+  year = {2018},
+  keywords = {urban design,KSI,morphometrics,multilevel,STATS19,Traffic casualty severity},
+  pages = {510-525},
+  file = {/home/robin/Zotero/storage/CKF3JGXX/Sarkar et al. - 2018 - Street morphology and severity of road casualties.pdf;/home/robin/Zotero/storage/FNTF9I3H/15568318.2017.html},
+}
