### 2018-12-19 fe658af
Update paper.md
