### 2018-12-19 e9a40e0
Update readme with placeholders for cas and veh tables
