### 2018-12-19 1a9692a
Fucntion docs reviewed, link #47
