### 2018-12-19 5a75c8b
Fucntion docs reviewed, link #47
