+++ man/locate_one_file.Rd
@@ -16,7 +16,7 @@ otherwise disk scan would be needed.}
 
 \item{year}{Single year for which file is to be found.}
 
-\item{type}{One of: 'Accidents', 'Casualties', 'Vehicles'; defaults to 'Accidents', ignores case.}
+\item{type}{One of: 'Accidents', 'Casualties', 'Vehicles'; ignores case.}
 }
 \value{
 One of: path for one file, a message `More than one file found` or error if none found.
