+++ man/check_year.Rd
@@ -7,7 +7,7 @@
 check_year(year)
 }
 \arguments{
-\item{year}{Single year for which file is to be downloaded}
+\item{year}{Single year for which file is to be downloaded.}
 }
 \description{
 check and convert year argument
