### 2018-12-19 efd2e00
Major update to readme, address reviewer comments
