+++ README.md
@@ -98,7 +98,7 @@ dl_stats19(year = 2017, type = "Accidents")
 #> Files identified: dftRoadSafetyData_Accidents_2017.zip
 #> Attempt downloading from:
 #>    http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/dftRoadSafetyData_Accidents_2017.zip
-#> Data saved at /tmp/RtmphiBH2G/dftRoadSafetyData_Accidents_2017/Acc.csv
+#> Data saved at /tmp/RtmpIzUU39/dftRoadSafetyData_Accidents_2017/Acc.csv
 ```
 
 Currently, these files are downloaded to a default location of “tempdir”
@@ -146,40 +146,100 @@ section), they can then be read-in as follows:
 ``` r
 crashes_2017_raw = read_accidents(year = 2017)
 #> Reading in:
-#> /tmp/RtmphiBH2G/dftRoadSafetyData_Accidents_2017/Acc.csv
+#> /tmp/RtmpIzUU39/dftRoadSafetyData_Accidents_2017/Acc.csv
+nrow(crashes_2017_raw)
+#> [1] 129982
+ncol(crashes_2017_raw)
+#> [1] 32
 crashes_2017 = format_accidents(crashes_2017_raw)
 ```
 
 What just happened? We read-in data on all road crashes recorded by the
-police in 2017 across Great Britain. `read_accidents()` imports the
-‘raw’ Stats19 data without cleaning messy column names or
-re-categorising the outputs. `format_accidents()` does this work,
-automating the process of matching column names with variable names and
-labels in a [`.xls`
+police in 2017 across Great Britain. The dataset contains 32 columns
+(variables) for 129982 crashes.
+
+This work was done by `read_accidents()`, which imported the ‘raw’
+Stats19 data without cleaning messy column names or re-categorising the
+outputs. `format_accidents()` does the formatting work, automating the
+process of matching column names with variable names and labels in a
+[`.xls`
 file](http://data.dft.gov.uk/road-accidents-safety-data/Road-Accident-Safety-Data-Guide.xls)
 provided by the DfT. This means `crashes_2017` is much more usable than
-`crashes_2017_raw`, as shown below, which shows three records and some
-key variables in the messy and clean datasets:
+`crashes_2017_raw`, as shown below, which shows some key variables in
+the messy and clean datasets:
 
 ``` r
-key_patt = "severity|speed|light|human"
-key_vars = grep(key_patt, x = names(crashes_2017_raw), ignore.case = TRUE)
-random_n = sample(x = nrow(crashes_2017_raw), size = 3)
-crashes_2017_raw[random_n, key_vars]
-#> # A tibble: 3 x 4
-#>   Accident_Severity Speed_limit `Pedestrian_Crossing-Huma… Light_Conditions
-#>               <int>       <int>                      <int>            <int>
-#> 1                 3          30                          0                1
-#> 2                 2          30                          0                1
-#> 3                 3          60                          0                1
-crashes_2017[random_n, key_vars]
-#> # A tibble: 3 x 4
-#>   accident_severity speed_limit pedestrian_crossing_human… light_conditions
-#>   <chr>                   <int> <chr>                      <chr>           
-#> 1 Slight                     30 None within 50 metres      Daylight        
-#> 2 Serious                    30 None within 50 metres      Daylight        
-#> 3 Slight                     60 None within 50 metres      Daylight
+crashes_2017_raw[c(7, 18, 23, 25)]
+#> # A tibble: 129,982 x 4
+#>    Accident_Severity Speed_limit `Pedestrian_Crossing-Hum… Light_Conditions
+#>                <int>       <int>                     <int>            <int>
+#>  1                 1          30                         0                4
+#>  2                 3          30                         0                4
[…]
