### 2018-12-19 2c0bf6a
Don't evaluate empty strings, double ||
