### 2018-12-19 7cef5fd
Consistent sanity checks with coverage
