### 2018-12-19 95a9b9e
read.R: minor, cover and more logical flow
