### 2018-12-19 fc5b09b
Add maps and analyis to vignette
