+++ README.md
@@ -98,7 +98,7 @@ dl_stats19(year = 2017, type = "Accidents")
 #> Files identified: dftRoadSafetyData_Accidents_2017.zip
 #> Attempt downloading from:
 #>    http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/dftRoadSafetyData_Accidents_2017.zip
-#> Data saved at /tmp/RtmpIzUU39/dftRoadSafetyData_Accidents_2017/Acc.csv
+#> Data saved at /tmp/Rtmp1K5WYn/dftRoadSafetyData_Accidents_2017/Acc.csv
 ```
 
 Currently, these files are downloaded to a default location of “tempdir”
@@ -146,7 +146,7 @@ section), they can then be read-in as follows:
 ``` r
 crashes_2017_raw = read_accidents(year = 2017)
 #> Reading in:
-#> /tmp/RtmpIzUU39/dftRoadSafetyData_Accidents_2017/Acc.csv
+#> /tmp/Rtmp1K5WYn/dftRoadSafetyData_Accidents_2017/Acc.csv
 nrow(crashes_2017_raw)
 #> [1] 129982
 ncol(crashes_2017_raw)
@@ -258,7 +258,7 @@ dl_stats19(year = 2017, type = "casualties")
 #> Files identified: dftRoadSafetyData_Casualties_2017.zip
 #> Attempt downloading from:
 #>    http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/dftRoadSafetyData_Casualties_2017.zip
-#> Data saved at /tmp/RtmpIzUU39/dftRoadSafetyData_Casualties_2017/Cas.csv
+#> Data saved at /tmp/Rtmp1K5WYn/dftRoadSafetyData_Casualties_2017/Cas.csv
 casualties_2017_raw = read_casualties(year = 2017)
 nrow(casualties_2017_raw)
 #> [1] 170993
@@ -322,7 +322,7 @@ dl_stats19(year = 2017, type = "vehicles")
 #> Files identified: dftRoadSafetyData_Vehicles_2017.zip
 #> Attempt downloading from:
 #>    http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/dftRoadSafetyData_Vehicles_2017.zip
-#> Data saved at /tmp/RtmpIzUU39/dftRoadSafetyData_Vehicles_2017/Veh.csv
+#> Data saved at /tmp/Rtmp1K5WYn/dftRoadSafetyData_Vehicles_2017/Veh.csv
 vehicles_2017_raw = read_vehicles(year = 2017)
 nrow(vehicles_2017_raw)
 #> [1] 238926
@@ -373,6 +373,118 @@ names(vehicles_2017)
 ```
 
 <!-- More data can be read-in as follows: -->
+
+## Creating geographic crash data
+
+An important feature of STATS19 data is that the ‘accidents’ table
+contains geographic coordinates. These are provided at ~10m resolution
+in the UK’s official coordinate reference system (the Ordnance Survey
+National Grid, EPSG code 27700). **stats19** converts the non-geographic
+tables created by `format_accidents()` into a geographic (`sf`) data
+form with the function `format_sf()` as follows:
+
+``` r
+crashes_sf = format_sf(crashes_2017)
+#> 19 rows removed with no coordinates
+```
+
+Note: rows without coordinates are removed, NAs are not permitted in
+`sf` coordinates. Once the data is in this form you can do geographic
+operations on it. The following code chunk, for example, returns all
+crashes in the Leeds local authority (requires the **ukboundaries**
+GitHub package):
+
+``` r
+library(sf)
[…]
