### 2018-12-19 7b3e846
Add time-series analysis plots
