+++ README.md
@@ -98,7 +98,7 @@ dl_stats19(year = 2017, type = "Accidents")
 #> Files identified: dftRoadSafetyData_Accidents_2017.zip
 #> Attempt downloading from:
 #>    http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/dftRoadSafetyData_Accidents_2017.zip
-#> Data saved at /tmp/Rtmp1K5WYn/dftRoadSafetyData_Accidents_2017/Acc.csv
+#> Data saved at /tmp/RtmpsAf5xr/dftRoadSafetyData_Accidents_2017/Acc.csv
 ```
 
 Currently, these files are downloaded to a default location of “tempdir”
@@ -146,7 +146,7 @@ section), they can then be read-in as follows:
 ``` r
 crashes_2017_raw = read_accidents(year = 2017)
 #> Reading in:
-#> /tmp/Rtmp1K5WYn/dftRoadSafetyData_Accidents_2017/Acc.csv
+#> /tmp/RtmpsAf5xr/dftRoadSafetyData_Accidents_2017/Acc.csv
 nrow(crashes_2017_raw)
 #> [1] 129982
 ncol(crashes_2017_raw)
@@ -258,7 +258,7 @@ dl_stats19(year = 2017, type = "casualties")
 #> Files identified: dftRoadSafetyData_Casualties_2017.zip
 #> Attempt downloading from:
 #>    http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/dftRoadSafetyData_Casualties_2017.zip
-#> Data saved at /tmp/Rtmp1K5WYn/dftRoadSafetyData_Casualties_2017/Cas.csv
+#> Data saved at /tmp/RtmpsAf5xr/dftRoadSafetyData_Casualties_2017/Cas.csv
 casualties_2017_raw = read_casualties(year = 2017)
 nrow(casualties_2017_raw)
 #> [1] 170993
@@ -322,7 +322,7 @@ dl_stats19(year = 2017, type = "vehicles")
 #> Files identified: dftRoadSafetyData_Vehicles_2017.zip
 #> Attempt downloading from:
 #>    http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/dftRoadSafetyData_Vehicles_2017.zip
-#> Data saved at /tmp/Rtmp1K5WYn/dftRoadSafetyData_Vehicles_2017/Veh.csv
+#> Data saved at /tmp/RtmpsAf5xr/dftRoadSafetyData_Vehicles_2017/Veh.csv
 vehicles_2017_raw = read_vehicles(year = 2017)
 nrow(vehicles_2017_raw)
 #> [1] 238926
@@ -429,7 +429,7 @@ cas_types = casualties_leeds %>%
   summarise(n = sum(n)) %>% 
   spread(casualty_type, n, fill = 0) 
 cas_types$Total = rowSums(cas_types[-1])
-crashes_joined = left_join(crashes_leeds[1], cas_types)
+crashes_joined = left_join(crashes_leeds, cas_types)
 #> Joining, by = "accident_index"
 ```
 
@@ -458,19 +458,60 @@ speed limits tend to be higher and where there are comparatively high
 volumes of motor traffic, compared with the city centre.
 
 Another way of visualising the data, in this case to show the spatial
-distribution of crashes causing pedestrian casualties by numbers of
-pedestrians hurt, is with **ggplot2**:
+distribution of crashes causing pedestrian casualties by number of
+pedestrians hurt and crash severity, is with **ggplot2**:
 
 ``` r
 ggplot(crashes_joined, aes(colour = Total)) +
   geom_sf() +
-  facet_wrap(vars(Pedestrian))
[…]
