### 2018-12-19 084f597
Call format function within read_*
