### 2018-12-19 93ce5c3
Call format function within read_*
