### 2018-12-19 bec18c8
Update readme for format = TRUE
