### 2018-12-19 2a1a807
Fix travis - cc @Layik
