### 2018-12-19 36ce52b
Fix travis - cc @Layik
