### 2018-12-19 0b4427f
Update readme - fix typo
