### 2018-12-20 2c017cf
Reply to some of the comments and TODO to others
