+++ responses1.Rmd
@@ -46,12 +46,12 @@ TODO: add a map and a time series plot showing the scale of the data. Also show 
 
 > I couldn't load the vignette from the console:
 
-We also could not see the vignette even when installing with `build_vignettes = TRUE`. Is this linked to https://github.com/STAT545-UBC/Discussion/issues/68#issuecomment-67113218 ?
+We also could not see the vignette when installing using `devtools::install_github(build_vignettes = TRUE`. But we can see the vignette if we intall locally.
 
 This was the code we ran:
 
 ```{r}
-devtools::install_github("ITSLeeds/stats19", build_vignettes = TRUE)
+devtools::install(build_vignettes = TRUE)
 vignette(package = "stats19")
 ```
 
@@ -61,11 +61,13 @@ These have now been fixed - thanks for testing and reporting.
 
 > I couldn't find any explicit contributing guidelines in the README, and there is no CONTRIBUTING document.
 
+A CONTRIBUTING will be added at some point.
+
 > The package has an obvious research application according to JOSS's definition
 
 > There is no paper.md.
 
-We will add one with:
+One is added with:
 
 - A short summary describing the high-level functionality of the software
 - Authors: A list of authors with their affiliations
@@ -76,15 +78,28 @@ Review Comments
 
 > A superb and essential package--we need this data and we need it in these formats. The download-format-read-explore workflow is intuitive and relatively frictionless. I have only some brief comments:
 
+Thank you.
+
 > I wonder you could possibly merge the formatting and reading step with a raw = TRUE or format = TRUE argument in the read_* functions. But perhaps that's my tendency towards abstraction. Something like ac = read_accidents(year = 2017, format = TRUE)
+
+Done, appreciate your input.
 
 > My personal preference would be to have the schema from dl_schema lazily loaded with the package.
 
+TODO
+
 > According to the vignette, the dl_* functions are interactive, although the interactivity is commented out in the code. Will the interactivity be returning? Or does the vignette need to be updated?
+
+Back in, as stated above.
 
 > Out of curiosity, what's happening with https://github.com/cyipt/stats19? It was updated recently.
 
+As it used for CyIPT project, it might still stay where it is, but clearly this package is taking over and it is just a matter of adapting the use by the CyIPT and other users.
+
 > I confess I wish the package name was more expressive--stats19 sounds like an introductory statistics class.
+
+Internal discussions are ongoing. 
 
 > This data will be used to make many maps. I personally would love a nudge in that direction in either the README or the vignette.
 
+Definitely. Thank you very much for your input.
