### 2018-12-20 2827243
More respones, contributing doc added
