### 2018-12-20 29f9921
More respones, contributing doc added
