+++ CONTRIBUTING.md
@@ -1,0 +1,86 @@
+This document was adapted from @peterdesmet's [template](https://gist.github.com/peterdesmet/e90a1b0dc17af6c12daf6e8b2f044e7c)
+
+# Contributing to stats19
+
+First of all, thanks for considering contributing to stats19! 👍 It's people like you that make it rewarding for us - the project maintainers - to work on stats19. 😊
+
+stats19 is an open source project, maintained by people who care. We are not directly funded to do so.
+
+[repo]: https://github.com/ITSLeeds/stats19
+[issues]: https://github.com/ITSLeeds/stats19/issues
+[new_issue]: https://github.com/ITSLeeds/stats19/issues/new
+[website]: https://ITSLeeds.github.io/stats19
+[citation]: https://ITSLeeds.github.io/stats19/authors.html
+[email]: maintainer_email
+
+## Code of conduct
+
+Please note that this project is released with a [Contributor Code of Conduct](CODE_OF_CONDUCT.md). By participating in this project you agree to abide by its terms.
+
+## How you can contribute
+
+There are several ways you can contribute to this project. If you want to know more about why and how to contribute to open source projects like this one, see this [Open Source Guide](https://opensource.guide/how-to-contribute/).
+
+### Share the love ❤️
+
+Think stats19 is useful? Let others discover it, by telling them in person, via Twitter or a blog post.
+
+Using stats19 for a paper you are writing? Consider [citing it][citation].
+
+### Ask a question ⁉️
+
+Using stats19 and got stuck? Browse the [documentation][website] to see if you can find a solution. Still stuck? Post your question as an [issue on GitHub][new_issue]. While we cannot offer user support, we'll try to do our best to address it, as questions often lead to better documentation or the discovery of bugs.
+
+Want to ask a question in private? Contact the package maintainer by [email][email].
+
+### Propose an idea 💡
+
+Have an idea for a new stats19 feature? Take a look at the [documentation][website] and [issue list][issues] to see if it isn't included or suggested yet. If not, suggest your idea as an [issue on GitHub][new_issue]. While we can't promise to implement your idea, it helps to:
+
+* Explain in detail how it would work.
+* Keep the scope as narrow as possible.
+
+See below if you want to contribute code for your idea as well.
+
+### Report a bug 🐛
+
+Using stats19 and discovered a bug? That's annoying! Don't let others have the same experience and report it as an [issue on GitHub][new_issue] so we can fix it. A good bug report makes it easier for us to do so, so please include:
+
+* Your operating system name and version (e.g. Mac OS 10.13.6).
+* Any details about your local setup that might be helpful in troubleshooting.
+* Detailed steps to reproduce the bug.
+
+### Improve the documentation 📖
+
+Noticed a typo on the website? Think a function could use a better example? Good documentation makes all the difference, so your help to improve it is very welcome!
+
+#### The website
+
+[This website][website] is generated with [`pkgdown`](http://pkgdown.r-lib.org/). That means we don't have to write any html: content is pulled together from documentation in the code, vignettes, [Markdown](https://guides.github.com/features/mastering-markdown/) files, the package `DESCRIPTION` and `_pkgdown.yml` settings. If you know your way around `pkgdown`, you can [propose a file change](https://help.github.com/articles/editing-files-in-another-user-s-repository/) to improve documentation. If not, [report an issue][new_issue] and we can point you in the right direction.
+
+#### Function documentation
+
+Functions are described as comments near their code and translated to documentation using [`roxygen2`](https://klutometis.github.io/roxygen/). If you want to improve a function description:
+
+1. Go to `R/` directory in the [code repository][repo].
+2. Look for the file with the name of the function.
[…]
