### 2018-12-20 e865fd7
Small changes to responses
