### 2018-12-20 ee17fb7
Update vignette and more reviews, #49
