### 2018-12-20 74823e2
Update vignette and more reviews
