+++ vignettes/stats19.Rmd
@@ -35,11 +35,39 @@ devtools::install_github("ITSLeeds/stats19")
 library(stats19)
 ```
 
-The package has 3 main types of functions, corresponding with the main tasks undertaken by the package:
+**stats19** has three main types of functions, corresponding with the three stages of data processing:
 
-- Download functions beginning with `dl_`
-- Read functions beginning with `read_`
-- Format functions beginning with `format_`
+- **Download**: This stage is taken care of by the `dl_stats19()`
+    function. `year`, `type` and `filename` arguments make it easy to
+    find the right file for your research question.
+  - **Read**: STATS19 data is provided in a particular format that
+    benefits from being read-in with pre-specified column types. This is
+    taken care of with `read_*()` functions corresponding to the 3 main
+    tables in STATS19 data:
+    
+      - `read_accidents()` reads-in the crash data (which has one row
+        per incident)
+      - `read_casualties()` reads-in the casualty data (which has one
+        row per person injured or killed)
+      - `read_vehicles()` reads-in the vehicles table, which contains
+        information on the vehicles involved in the crashes (one row per
+        vehicle)
+
+  - **Format**: in this stage labels are added to the tables. The raw
+    data provided by the DfT contains only integers. Running the
+    following commands converts these integer values to the
+    corresponding character variables, for each of the three tables:
+    
+      - `read_accidents()` adds labels to the accidents table. The
+        values in the `accident_severity` column, for example, are
+        converted from `1`, `2` and `3` into `Slight`, `Serious` or
+        `Fatal`.
+      - `read_casualties()` formats the casualty data
+      - `read_vehicles()`formats the vehicle data
+
+By default, stages 2 and 3 are done automatically in the `read_*()`
+functions. To read-in the raw data, without formatting, set `format =
+FALSE`.
 
 Multiple functions are needed for each step because of the structure of STATS19 data, which are divided into 3 tables, as described in an article that analyses 2005-2015 STATS19 data with reproducible R code by Christoph Freier ([2018](https://kawameicha.gitlab.io/inCodeWeTrust/2018/01/10-years-and-3.3-million-accidented-vehicles-in-the-uk/)):
 
@@ -55,10 +83,9 @@ goes back to 1979, providing a long period for comparison.
 
 ## Download STATS19 data
 
-The main function to download individual files `dl_stats19`.
-The function takes three parameters.
-If `file_name` has been provided, this means the user is aware of what to download and the other two parameters will be ignored.
-You can also use `years` and `type` to "search" through the file names, which are stored in a datset called `stats19::file_names`.
+The function to download individual files `dl_stats19`. The function takes three parameters. If `file_name` has been provided, this means the user is aware of what to download and the other two parameters will be ignored.
+You can also use `years` and `type` to "search" through the file names, which are stored in a datset called `stats19::file_names`, which are lazy loaded.
+
 You can find out the names of files that can be downloaded with `names(stats19::file_names)`, an example of which is shown below:
 
 ```{r}
