### 2018-12-20 3e2bc45
Update vignette and more reviews
