### 2018-12-20 fe8147e
Update vignette and more reviews, #49
