### 2018-12-20 6077d6d
unexport read_schema()
