### 2018-12-20 81a16c0
PR to move read_schema to data.R
