### 2018-12-20 60f430e
PR to move read_schema to data.R
