### 2018-12-20 cc629f0
Move chunk to examples in data.R
