### 2018-12-20 1ea58a5
Comment out example comments
