+++ README.md
@@ -1,4 +1,3 @@
-
 [![The API of a maturing package has been roughed out, but finer details
 likely to
 change.](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)
@@ -18,58 +17,59 @@ downloads](http://cranlogs.r-pkg.org/badges/stats19)](http://www.r-pkg.org/pkg/s
 **stats19** provides functions for downloading and formatting road crash
 data. Specifically, it enables access to the UK’s official road traffic
 casualty database,
-[STATS19](https://data.gov.uk/dataset/cb7ae6f0-4be6-4935-9277-47e5ce24a11f/road-safety-data)
-(the name comes from the form used by the police to record car crashes
+[STATS19](https://data.gov.uk/dataset/cb7ae6f0-4be6-4935-9277-47e5ce24a11f/road-safety-data).
+(The name comes from the form used by the police to record car crashes
 and other incidents resulting in casualties on the roads).
 
-The data, in its original form, is provided as a series of .csv files
-that are stored in dozens of `.zip` files. Finding, reading-in and
-formatting the data for research can be a time consuming process subject
-to human error. **stats19** speeds-up these vital data access and
-cleaning stages of the research process, enabling reproducibility, by
-streamlining the work into 3 stages:
+**Note**: The Department for Transport refers to “accidents”, but
+“crashes” is a more appropriate term, as emphasised in the “crash not
+accident” arguments of road safety advocacy groups such as
+[RoadPeace](http://www.roadpeace.org/take-action/crash-not-accident/).
+We use the term “accidents” only in reference to nomenclature within the
+data as provided.
 
-  - **Download**: This stage is taken care of by the `dl_stats19()`
-    function. `year`, `type` and `filename` arguments make it easy to
-    find the right file for your research question.
+The data is provided by the UK’s Department for Transport (DfT) as a
+series of `.csv` files that are stored in dozens of `.zip` files.
+Finding, reading-in and formatting the data for research can be a time
+consuming process subject to human error. **stats19** speeds up these
+vital data access and cleaning stages of the research process, enabling
+reproducibility, by streamlining the work into 3 stages:
+
+  - **Download**: This single `dl_stats19()` function accepts `year`,
+    `type` and `filename` arguments to make it easy to find the right
+    file for your research question.
 
   - **Read**: STATS19 data is provided in a particular format that
     benefits from being read-in with pre-specified column types. This is
-    taken care of with `read_*()` functions corresponding to the 3 main
-    tables in STATS19 data:
+    taken care of with `read_*()` functions providing access to the 3
+    main tables in STATS19 data:
     
       - `read_accidents()` reads-in the crash data (which has one row
         per incident)
       - `read_casualties()` reads-in the casualty data (which has one
         row per person injured or killed)
       - `read_vehicles()` reads-in the vehicles table, which contains
-        information on the vehicles involved in the crashes (one row per
-        vehicle)
+        information on the vehicles involved in the crashes (and has one
+        row per vehicle)
 
-  - **Format**: in this stage labels are added to the tables. The raw
-    data provided by the DfT contains only integers. Running the
-    following commands converts these integer values to the
[…]
