### 2018-12-21 #266 @Robinlovelace
Thanks @daranzolin, which years do you get parsing failures for? Cannot reproduce for 2017 data, as illustrated in the reprex below:

``` r
devtools::install_github("ITSLeeds/stats19")
#> Skipping install of 'stats19' from a github remote, the SHA1 (225bce03) has not changed since last install.
#>   Use `force = TRUE` to force installation

# casualties
dl_stats19(year = 2017, type = "casualties")
#> Error in dl_stats19(year = 2017, type = "casualties"): could not find function "dl_stats19"
ca = read_casualties(year = 2017)
#> Error in read_casualties(year = 2017): could not find function "read_casualties"
head(ca$accident_index)
#> Error in head(ca$accident_index): object 'ca' not found

# vehicles
dl_stats19(year = 2017, type = "vehicles")
#> Error in dl_stats19(year = 2017, type = "vehicles"): could not find function "dl_stats19"
ve = read_vehicles(year = 2017)
#> Error in read_vehicles(year = 2017): could not find function "read_vehicles"
head(ve$accident_index)
#> Error in head(ve$accident_index): object 've' not found
```
