### 2018-12-21 ITSLeeds/stats19#49 @layik
https://github.com/ropensci/software-review/issues/266
