### 2018-12-21 0750002
Move dl_schema to data.R inline with single use read_schema change
