+++ NAMESPACE
@@ -1,6 +1,5 @@
 # Generated by roxygen2: do not edit by hand
 
-export(dl_schema)
 export(dl_stats19)
 export(find_file_name)
 export(format_accidents)
