### 2018-12-21 e6d5176
Move dl_schema to data.R inline with single use read_schema change
