+++ vignettes/stats19.Rmd
@@ -14,7 +14,8 @@ bibliography: references.bib
 ```{r setup, include = FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
-  comment = "#>"
+  comment = "#>",
+  out.width = "100%"
 )
 ```
 
@@ -113,7 +114,7 @@ dl_stats19(file_name = paste0(d17, ".zip"))
 Note that this is the same file (containing the 'accidents' table from 2017) downloaded previously with the command `dl_stats19(year = 2017, type = "Accidents")`.
 If the file is already present, `dl_stats19()` will skip the download process.
 
-## Read STATS19 data
+## Read and format STATS19 data
 
 In a similar approach to the download section before, we can read files downloaded using a `data_dir` location of the file and the `filename` to read.
 The code below will download the `dftRoadSafetyData_Accidents_2017.zip` file from the DfT servers and read its content.
@@ -134,8 +135,243 @@ class(crashes_2017_raw)
 dim(crashes_2017_raw)
 ```
 
+There are three `read_*()` functions, corresponding to the three different classes of data provided by the DfT:
+1. `read_accidents()`
+2. `read_casualties()`
+3. `read_vehicles()`
+
+In all cases, a default parameter `read_*(format = TRUE)` returns the data in
+formatted form, as described above. Data can also be imported in the
+form directly provided by the DfT by passing `format = FALSE`, and then
+subsequently formatted with additional `format_*()` functions, as described in a
+final section of this vignette. Each of these `read*()` functions is now
+described in more detail.
+
+
+### Crash data
+
+After raw data files have been downloaded as described in the previous section, they can then be read-in as follows:
+
+```{r read2017-raw-format}
+crashes_2017_raw = read_accidents(year = 2017, format = FALSE)
+crashes_2017 = format_accidents(crashes_2017_raw)
+nrow(crashes_2017_raw)
+ncol(crashes_2017_raw)
+nrow(crashes_2017)
+ncol(crashes_2017)
+```
+
+What just happened?
+We read-in data on all road crashes recorded by the police in 2017 across Great Britain.
+The dataset contains `r ncol(crashes_2017_raw)` columns (variables) for `r format(nrow(crashes_2017_raw), big.mark = ",")` crashes.
+
+This work was done by `read_accidents(format = FALSE)`, which imported the "raw" Stats19 data without cleaning messy column names or re-categorising the outputs.
+`format_accidents()` automates the process of matching column names with variable names and labels in a [`.xls` file](http://data.dft.gov.uk/road-accidents-safety-data/Road-Accident-Safety-Data-Guide.xls) provided by the DfT.
+This means `crashes_2017` is much more usable than `crashes_2017_raw`, as shown below, which shows some key variables in the messy and clean datasets:
+
+```{r crashes2017-columns}
+crashes_2017_raw[c(7, 18, 23, 25)]
+crashes_2017[c(7, 18, 23, 25)]
+```
+
+By default, `format = TRUE`, meaning that the two stages of `read_accidents(format = FALSE)` and `format_accidents()` yield the same result as `read_accidents(format = TRUE)`.
+For the full list of columns, run `names(crashes_2017)` or see the vignette.
+
+<!-- This means `crashes_2017` is much more usable than `crashes_2017_raw`, as shown below, which shows three records and some key variables in the messy and clean datasets: -->
+
+```{r, echo=FALSE, eval=FALSE}
+# commented out as confusing...
+key_patt = "severity|speed|light|human"
+key_vars = grep(key_patt, x = names(crashes_2017_raw), ignore.case = TRUE)
[…]
