### 2018-12-21 b184dda
rm ukboundaries dep from vignette; add local data
