### 2018-12-21 3443ec3
Add Azure Pipeline status badge
