### 2018-12-21 c585e9f
Add Azure Pipeline status badge
