### 2018-12-21 5aa64de
Test travis install from CRAN than github
