### 2018-12-21 c05051a
Travis install tidyverse not package dependency to build vignette
