### 2018-12-21 091538f
Build vignette, add suggested packages
