### 2018-12-21 58a9996
Build vignette, add suggested packages
