### 2018-12-21 53e31ce
Build vignette, needs lubridate in suggets
