+++ DESCRIPTION
@@ -23,12 +23,13 @@ Imports:
     curl (>= 3.2),
     tibble
 Suggests: 
-    knitr,
-    rmarkdown,
-    testthat,
-    tidyr,
     dplyr,
     ggplot2,
-    stringr
+    knitr,
+    lubridate,
+    rmarkdown,
+    stringr,
+    testthat,
+    tidyr
 VignetteBuilder: knitr
 RoxygenNote: 6.1.1
