### 2018-12-23 #266 @layik
This shows, as @daranzolin says, both "cas" and "veh" suffer from the same issue between 2009:2013 inclusive but no issues after 2014. Created a [gist](https://gist.github.com/layik/0f25386b6e4de26c5725f369d711c19d) rather polluting this thread with lengthy reprex.

Will check read/format in due course.
