### 2018-12-23 c041d74
Check available column name before asking readr to read it
