### 2018-12-23 74c070b
Test coverage for last commit an dremove unecessary check in read_accidentes
