+++ R/read.R
@@ -29,42 +29,22 @@ read_accidents = function(year = NULL,
   )
   message("Reading in: ")
   message(path)
-  h = utils::read.csv(path, nrows = 1)
-  if(identical(names(h)[1], "Accident_Index")) {
-    # read the data in
-    suppressWarnings({
-      ac = readr::read_csv(
-        path,
-        col_types = readr::cols(
-          .default = readr::col_integer(),
-          Accident_Index = readr::col_character(),
-          Longitude = readr::col_double(),
-          Latitude = readr::col_double(),
-          Date = readr::col_character(),
-          Time = readr::col_character(),
-          `Local_Authority_(Highway)` = readr::col_character(),
-          LSOA_of_Accident_Location = readr::col_character()
-        )
+  # read the data in
+  suppressWarnings({
+    ac = readr::read_csv(
+      path,
+      col_types = readr::cols(
+        .default = readr::col_integer(),
+        Accident_Index = readr::col_character(),
+        Longitude = readr::col_double(),
+        Latitude = readr::col_double(),
+        Date = readr::col_character(),
+        Time = readr::col_character(),
+        `Local_Authority_(Highway)` = readr::col_character(),
+        LSOA_of_Accident_Location = readr::col_character()
       )
-    })
-  } else {
-    # read the data in
-    suppressWarnings({
-      ac = readr::read_csv(
-        path,
-        col_types = readr::cols(
-          .default = readr::col_integer(),
-          Acc_Index = readr::col_character(),
-          Longitude = readr::col_double(),
-          Latitude = readr::col_double(),
-          Date = readr::col_character(),
-          Time = readr::col_character(),
-          `Local_Authority_(Highway)` = readr::col_character(),
-          LSOA_of_Accident_Location = readr::col_character()
-        )
-      )
-    })
-  }
+    )
+  })
 
   if(format)
     return(format_accidents(ac))
