### 2018-12-23 9f0410d
Split-out read_ve_ca into function
