+++ man/read_casualties.Rd
@@ -28,7 +28,7 @@ in the stats19 dataset.
 
 \examples{
 \dontrun{
-dl_stats19(years = 2017, type = "casualties")
-casualties = read_casualties()
+dl_stats19(year = 2017, type = "casualties")
+casualties = read_casualties(year = 2017)
 }
 }
