+++ R/read.R
@@ -16,6 +16,8 @@
 #' \dontrun{
 #' dl_stats19(year = 2017, type = "Accidents")
 #' ac = read_accidents(year = 2017)
+#' dl_stats19(year = 2009, type = "Accidents")
+#' ac_2009 = read_accidents(year = 2009)
 #' }
 read_accidents = function(year = NULL,
                           filename = "",
@@ -49,7 +51,6 @@ read_accidents = function(year = NULL,
   if(format)
     return(format_accidents(ac))
   ac
-
 }
 
 #' Read in stats19 road safety data from .csv files downloaded.
