### 2018-12-23 #266 @Robinlovelace
Update: `index` branch has now been merged.
