+++ responses2.Rmd
@@ -1,0 +1,145 @@
+---
+output: github_document
+---
+
+## Responses to review 2 of stats19
+
+Many thanks for the detailed review.
+We think all the suggestions make sense, and link to previous discussions about combining the 3 stage (`dl`, `read`, `format`) process into a single function call, which we were thinking of calling `get_stats19()`.
+There reason for splitting the process up is to ensure maximum transparency and to give the user control over what the package is doing.
+However, as long as it is properly documents, we think the benefits of a `get_stats19()` function will outweigh any possible negatives we can think of so we plan to go ahead and create this function.
+
+We would very much welcome a pull request that address some of the other issues you mention.
+
+Responses to other issues/questions/comments are provided below.
+
+
+> Clearly much effort has gone into this package. I greatly support the sentiment behind making these data available in R having done the same for other data sets, myself. This package should be a great benefit to researchers in this area. I really appreciate the slim dependencies. It will make this package much easier to maintain into the future.
+
+Thanks for the comments, we have indeed tried to keep dependencies to a minimum but consider `readr` and `tibble` worthwhile.
+
+> I found the package to be well documented, the vignette is helpful in illustrating some use cases for the data along with how to access it and the code is clear.
+
+> Some of the functionality I find to be mildly confusing like downloading and then importing the files into the R session and then formatting. As a user I'd prefer it all in one step, but there are likely use cases I'm not aware of that mean that this is useful so some examples of this would be useful I think.
+
+> My general comments on the code follow and sections for the DESCRTIPTION and Vignette as well. I've commented quite a bit on grammar and spelling as I think that the polish of a package is important as it lends to the perception of the quality.
+
+
+--------------------------------------------------------------------------------
+
+- Per rOpenSci policy, avoid start-up messages, rely on the documentation for
+citation information:
+https://ropensci.github.io/dev_guide/building.html#recommended-scaffolding.
+
+- Avoid long lines >80 chars
+
+- Inconsistent use of white spaces in code, see `find_file_name()` `if` statements for examples.
+
+- The package does not pass `R CMD check`. *curl*, *readxl* and *tibble* are all listed as an Imports in DESCRIPTION but not imported from. With *curl* being used in tests, this means it should be in Suggests, I think. The others should be removed.
+
+- I don't think it's good form to have an example that won't work on Windows in the help file for `stats19_schema`, from data.R - line 17? Most of what I see there would be better served in a `data_raw` folder showing how the data were created with the documentation actually documenting what the variables are not how they were created, see <http://r-pkgs.had.co.nz/data.html> and for an example, <https://github.com/ropensci/GSODR/tree/master/data-raw>.
+
+- I would suggest to use proper formatting in help files, when naming packages, e.g. \pkg{stats19} and when referring to documented functions or data, e.g. \code{\link{stats19_schema}}, or with single quotes around abbreviations, e.g. 'DfT'. @ColinFay has an excellent page that outlines the formatting options and when/how to use them, <https://colinfay.me/writing-r-extensions/writing-r-documentation-files.html>. This will greatly enhance the users' experience when using the help files by making them more readable.
+
+- I also would suggest making use of `@seealso` in documentation. For example, the `dl_stats19()` example works great in the help files, but from there I have the data but it's not in R. Using the `@seealso` you can let the user know about the `read_*()` functions.
+
+- I downloaded files using `dl_stats19()`, selecting "Casualties", and then ran `read_accidents()` and got 
+```r
+Error in check_input_file(filename = filename, type = "accidents", data_dir = data_dir,  : 
[…]
