### 2018-12-29 024f9be
Demote readxl and curl to suggests
