### 2018-12-29 e402706
Demote readxl and curl to suggests
