### 2018-12-29 de34cf2
Shorten startup msg
