+++ responses2.Rmd
@@ -4,7 +4,7 @@ output: github_document
 
 ## Responses to review 2 of stats19
 
-Many thanks for the detailed review.
+Thanks for the detailed review.
 We think all the suggestions make sense, and link to previous discussions about combining the 3 stage (`dl`, `read`, `format`) process into a single function call, which we were thinking of calling `get_stats19()`.
 There reason for splitting the process up is to ensure maximum transparency and to give the user control over what the package is doing.
 However, as long as it is properly documents, we think the benefits of a `get_stats19()` function will outweigh any possible negatives we can think of so we plan to go ahead and create this function.
@@ -35,34 +35,62 @@ Agreed.
 
 --------------------------------------------------------------------------------
 
-- Per rOpenSci policy, avoid start-up messages, rely on the documentation for
+> Per rOpenSci policy, avoid start-up messages, rely on the documentation for
 citation information:
 https://ropensci.github.io/dev_guide/building.html#recommended-scaffolding.
 
-- Avoid long lines >80 chars
+The guidance is to 'Only use package startup messages when necessary'.
+A case can be made that this is necessary.
+As with `osmdata`, the package provides access to data that has a license that requires it to be cited.
+The `osmdata` load message is as follows:
 
-- Inconsistent use of white spaces in code, see `find_file_name()` `if` statements for examples.
+```{r}
+library(osmdata)
+```
 
-- The package does not pass `R CMD check`. *curl*, *readxl* and *tibble* are all listed as an Imports in DESCRIPTION but not imported from. With *curl* being used in tests, this means it should be in Suggests, I think. The others should be removed.
+We fully agree with the reasoning behind remove package startup messages however.
+As a compromise, we've shortened the startup from 4 lines to 2:
 
-- I don't think it's good form to have an example that won't work on Windows in the help file for `stats19_schema`, from data.R - line 17? Most of what I see there would be better served in a `data_raw` folder showing how the data were created with the documentation actually documenting what the variables are not how they were created, see <http://r-pkgs.had.co.nz/data.html> and for an example, <https://github.com/ropensci/GSODR/tree/master/data-raw>.
+```{r}
+# before:
+# Data provided under the conditions of the Open Government License.
+# If you use data from this package, mention the source
+# (UK Department for Transport), cite the package, and link to:
+# www.nationalarchives.gov.uk/doc/open-government-licence/version/3/.
+```
 
-- I would suggest to use proper formatting in help files, when naming packages, e.g. \pkg{stats19} and when referring to documented functions or data, e.g. \code{\link{stats19_schema}}, or with single quotes around abbreviations, e.g. 'DfT'. @ColinFay has an excellent page that outlines the formatting options and when/how to use them, <https://colinfay.me/writing-r-extensions/writing-r-documentation-files.html>. This will greatly enhance the users' experience when using the help files by making them more readable.
+```{r}
+# after:
+library(stats19)
+```
 
-- I also would suggest making use of `@seealso` in documentation. For example, the `dl_stats19()` example works great in the help files, but from there I have the data but it's not in R. Using the `@seealso` you can let the user know about the `read_*()` functions.
 
-- I downloaded files using `dl_stats19()`, selecting "Casualties", and then ran `read_accidents()` and got 
+
+
+> Avoid long lines >80 chars
+
+> Inconsistent use of white spaces in code, see `find_file_name()` `if` statements for examples.
+
+> The package does not pass `R CMD check`. *curl*, *readxl* and *tibble* are all listed as an Imports in DESCRIPTION but not imported from. With *curl* being used in tests, this means it should be in Suggests, I think. The others should be removed.
+
[…]
