### 2018-12-29 b1ed3a6
Shorten startup msg
