+++ R/stats19-package.R
@@ -1,9 +1,7 @@
 .onAttach = function(libname, pkgname) {
-  msg = paste0 (
-      "Data provided under the conditions of the Open Government License.\n",
-      "If you use data from this package, ",
-      "mention the source\n(UK Department for Transport), cite the package, and link to:\n",
-      "www.nationalarchives.gov.uk/doc/open-government-licence/version/3/."
+  msg = paste0(
+      "Data provided under OGL v3.0. Cite the source and link to:\n",
+      "www.nationalarchives.gov.uk/doc/open-government-licence/version/3/"
     )
-  packageStartupMessage (msg)
+  packageStartupMessage(msg)
 }
