+++ R/data.R
@@ -276,3 +276,24 @@ NULL
 #' @aliases vehicles_sample_raw
 #' @format A data frame
 NULL
+
+#' Police force boundaries in England (2016)
+#'
+#' See http://geoportal.statistics.gov.uk/
+#'
+#' @examples
+#' \dontrun{
+#' # Obtained with:
+#' library(sf)
+#' u = paste0("https://opendata.arcgis.com/",
+#'   "datasets/3e5a096a8c7c456fb6d3164a3f44b005_3.kml"
+#'   )
+#' police_boundaries_wgs = read_sf(u)
+#' police_boundaries = st_transform(police_boundaries_wgs, 27700)
+#' police_boundaries = police_boundaries[c("pfa16cd", "pfa16nm")]
+#' }
+#' @docType data
+#' @keywords datasets
+#' @name police_boundaries
+#' @format An sf data frame
+NULL
