### 2018-12-29 4e0cf5e
Shorten code, lines < 80 chars
