### 2018-12-29 2b82ae5
Shorten code, lines < 80 chars
