### 2018-12-29 c5f8028
Update vignette to work with new police_boundaries dataset
