+++ vignettes/stats19.Rmd
@@ -114,7 +114,7 @@ dl_stats19(file_name = paste0(d17, ".zip"))
 Note that this is the same file (containing the 'accidents' table from 2017) downloaded previously with the command `dl_stats19(year = 2017, type = "Accidents")`.
 If the file is already present, `dl_stats19()` will skip the download process.
 
-## Read and format STATS19 data
+## Read STATS19 data
 
 In a similar approach to the download section before, we can read files downloaded using a `data_dir` location of the file and the `filename` to read.
 The code below will download the `dftRoadSafetyData_Accidents_2017.zip` file from the DfT servers and read its content.
@@ -189,6 +189,90 @@ crashes_2017[random_n, key_vars]
 ```
 
 **Note**: As indicated above, the term "accidents" is only used as directly provided by the DfT; "crashes" is a more appropriate term, hence we call our resultant datasets `crashes_*`.
+
+## Format STATS19 data
+
+It is also possible to import the "raw" data as provided by the DfT.
+A [`.xls` file](http://data.dft.gov.uk/road-accidents-safety-data/Road-Accident-Safety-Data-Guide.xls) provided by the DfT defines the column names for the datasets provided.
+The packaged datasets `stats19_variables` and `stats19_schema` provide summary information about the contents of this data guide.
+These contain the full variable names in the guide (`stats19_variables`) and a complete lookup table relating integer values to the `.csv` files provided by the DfT and their labels (`stats19_schema`).
+The first rows of each dataset are shown below:
+
+```{r variables-and-schema}
+stats19_variables
+stats19_schema
+```
+
+The code that generated these small datasets can be found in their help pages (accessed with `?stats19_variables` and `?stats19_schema` respectively).
+`stats19_schema` is used internally to automate the process of formatting the downloaded `.csv` files.
+Column names are formatted by the function `format_column_names()`, as illustrated below:
+
+```{r format-col-names}
+format_column_names(stats19_variables$variable[1:3])
+```
+
+Previous approaches to data formatting `STATS19` data involved hard-coding results.
+This more automated approach to data cleaning is more consistent and fail-safe.
+Three functions, `format_accidents()`, `format_vehicles()` and
+`format_casualties()` do the data formatting on the respective data frames, as
+illustrated below:
+
+```{r format-main}
+crashes_2017 = format_accidents(crashes_2017_raw)
+
+# vehicle data for 2017
+dl_stats19(year = 2017, type = "vehicles")
+vehicles_2017_raw = read_vehicles(year = 2017)
+vehicles_2017 = format_vehicles(vehicles_2017_raw)
+
+# casualties data for 2017
+dl_stats19(year = 2017, type = "casualties")
+casualties_2017 = read_casualties(year = 2017)
+```
+
+The package automates this two-step `read_*` and `format_*`
+process by defaulting in all cases to `data_year = read_*(year, format = TRUE)`.
+`read_*` functions return, by default, formatted data.
+The two-step process may nevertheless be important for reference to the official
+nomenclature and values as provided by the DfT.
+
+A summary of the outputs for each of the three tables is shown below.
+
+```{r summarise-stats19}
+summarise_stats19 = function(x) {
+  data.frame(row.names = 1:length(x),
+    name = substr(names(x), 1, 19),
+    class = sapply(x, class),
+    n_unique = sapply(x, function(v) length(unique(v))),
+    first_label = sapply(x, function(v) substr(unique(v)[1], 1, 9)),
[…]
