### 2018-12-29 be27594
Update vignette to work with new police_boundaries dataset
