### 2018-12-30 6800196
Update vignette and readme
