+++ responses2.Rmd
@@ -75,11 +75,14 @@ running `goodpractice::gp()` found the following lines with > 80 lines:
     R/utils.R:167:1
 ```
 
-All these have been fixed - see 
+All these have been fixed.
 
 > Inconsistent use of white spaces in code, see `find_file_name()` `if` statements for examples.
 
 > The package does not pass `R CMD check`. *curl*, *readxl* and *tibble* are all listed as an Imports in DESCRIPTION but not imported from. With *curl* being used in tests, this means it should be in Suggests, I think. The others should be removed.
+
+`curl` is used in the tests and `readxl` is used in the examples.
+
 
 > I don't think it's good form to have an example that won't work on Windows in the help file for `stats19_schema`, from data.R - line 17? Most of what I see there would be better served in a `data_raw` folder showing how the data were created with the documentation actually documenting what the variables are not how they were created, see <http://r-pkgs.had.co.nz/data.html> and for an example, <https://github.com/ropensci/GSODR/tree/master/data-raw>.
 
