+++ README.md
@@ -1,9 +1,11 @@
+
 [![The API of a maturing package has been roughed out, but finer details
 likely to
 change.](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)
 [![Travis build
 status](https://travis-ci.org/ITSLeeds/stats19.svg?branch=master)](https://travis-ci.org/ITSLeeds/stats19)
-[![Azure Status](https://dev.azure.com/tralh/stats19/_apis/build/status/ITSLeeds.stats19?branchName=master)](https://dev.azure.com/tralh/stats19/_build/latest?definitionId=1?branchName=master)
+[![Azure
+Status](https://dev.azure.com/tralh/stats19/_apis/build/status/ITSLeeds.stats19?branchName=master)](https://dev.azure.com/tralh/stats19/_build/latest?definitionId=1?branchName=master)
 [![codecov](https://codecov.io/gh/ITSLeeds/stats19/branch/master/graph/badge.svg)](https://codecov.io/gh/ITSLeeds/stats19)
 [![Gitter
 chat](https://badges.gitter.im/ITSLeeds/stats19.png)](https://gitter.im/stats19/Lobby?source=orgpage)
@@ -78,10 +80,8 @@ devtools::install_github("ITSLeeds/stats19")
 
 ``` r
 library(stats19)
-#> Data provided under the conditions of the Open Government License.
-#> If you use data from this package, mention the source
-#> (Department for Transport), cite the package and link to:
-#> www.nationalarchives.gov.uk/doc/open-government-licence/version/3/.
+#> Data provided under OGL v3.0. Cite the source and link to:
+#> www.nationalarchives.gov.uk/doc/open-government-licence/version/3/
 ```
 
 <!-- You can install the released version of stats19 from [CRAN](https://CRAN.R-project.org) with: -->
@@ -103,7 +103,7 @@ dl_stats19(year = 2017, type = "Accidents")
 #> Files identified: dftRoadSafetyData_Accidents_2017.zip
 #> Attempt downloading from:
 #>    http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/dftRoadSafetyData_Accidents_2017.zip
-#> Data saved at /tmp/RtmpjsJsN9/dftRoadSafetyData_Accidents_2017/Acc.csv
+#> Data saved at /tmp/RtmpprL8mF/dftRoadSafetyData_Accidents_2017/Acc.csv
 ```
 
 Currently, these files are downloaded to a default location of `tempdir`
@@ -148,7 +148,7 @@ section, they can then be read-in as follows:
 ``` r
 crashes_2017_raw = read_accidents(year = 2017, format = FALSE)
 #> Reading in:
-#> /tmp/RtmpjsJsN9/dftRoadSafetyData_Accidents_2017/Acc.csv
+#> /tmp/RtmpprL8mF/dftRoadSafetyData_Accidents_2017/Acc.csv
 crashes_2017 = format_accidents(crashes_2017_raw)
 nrow(crashes_2017_raw)
 #> [1] 129982
@@ -228,7 +228,7 @@ dl_stats19(year = 2017, type = "casualties")
 #> Files identified: dftRoadSafetyData_Casualties_2017.zip
 #> Attempt downloading from:
 #>    http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/dftRoadSafetyData_Casualties_2017.zip
-#> Data saved at /tmp/RtmpjsJsN9/dftRoadSafetyData_Casualties_2017/Cas.csv
[…]
