### 2018-12-30 fa2580b
Update vignette and readme
