+++ README.Rmd
@@ -132,7 +132,7 @@ crashes_2017[c(7, 18, 23, 25)]
 ```
 
 By default, `format = TRUE`, meaning that the two stages of `read_accidents(format = FALSE)` and `format_accidents()` yield the same result as `read_accidents(format = TRUE)`.
-For the full list of columns, run `names(crashes_2017)` or see the vignette.
+For the full list of columns, run `names(crashes_2017)` or see the [vignette](https://github.com/ITSLeeds/stats19/blob/master/vignettes/stats19.Rmd).
 
 <!-- This means `crashes_2017` is much more usable than `crashes_2017_raw`, as shown below, which shows three records and some key variables in the messy and clean datasets: -->
 
@@ -195,23 +195,6 @@ The full list of column names in the `vehicles` dataset is:
 names(vehicles_2017)
 ```
 
-<!-- More data can be read-in as follows: -->
-
-```{r, eval=FALSE, echo=FALSE}
-# old code to be up-dated
-d14 = "Stats19_Data_2005-2014"
-crashes_2005_2014 = read_accidents(data_dir = d14)
-crashes_2005_2014_f = format_stats19_2005_2014_ac(crashes_2005_2014)
-d15 = "RoadSafetyData_2015"
-crashes_2015 = read_accidents(data_dir = d15, filename = "Accidents_2015.csv")
-crashes_2015_f = format_stats19_2015_ac(crashes_2015)
-d16 = "dftRoadSafety_Accidents_2016"
-crashes_2016 = read_accidents(data_dir = d16, filename = "dftRoadSafety_Accidents_2016.csv")
-crashes_2016_f = format_stats19_2016_ac(crashes_2016)
-all_crashes = rbind(crashes_2015_f, crashes_2016_f, crashes_2017_f)
-table(ac$Accident_Severity)
-```
-
 ## Creating geographic crash data
 
 An important feature of STATS19 data is that the "accidents" table contains geographic coordinates.
@@ -226,19 +209,18 @@ The note arises because `NA` values are not permitted in `sf` coordinates, and s
 Having the data in a standard geographic form allows various geographic operations to be performed on it.
 The following code chunk, for example, returns all crashes within the boundary of the Leeds local authority. 
 This requires the **ukboundaries** GitHub package:
-```{r ukboundaries, eval = FALSE}
-devtools::install_github("Robinlovelace/ukboundaries")
-```
 
 ```{r crashes-leeds}
 library(sf)
-leeds_osgb = st_transform(ukboundaries::leeds, crs = 27700)
-crashes_leeds = crashes_sf[leeds_osgb, ]
+west_yorkshire =
+  police_boundaries[police_boundaries$pfa16nm == "West Yorkshire", ]
+crashes_wy = crashes_sf[west_yorkshire, ]
 nrow(crashes_sf)
-nrow(crashes_leeds)
+nrow(crashes_wy)
 ```
-This subsetting has selected the `r format(nrow(crashes_leeds), big.mark = ",")`
-crashes which occurred within the Leeds boundary.
+
+This subsetting has selected the `r format(nrow(crashes_wy), big.mark = ",")`
+crashes which occurred within West Yorksire in 2017.
 
 
 ## Joining tables
@@ -247,31 +229,34 @@ The three main tables we have just read-in can be joined by shared key variables
 This is demonstrated in the code chunk below, which subsets all casualties that took place in Leeds, and counts the number of casualties by severity for each crash:
 
 ```{r table-join, message = FALSE}
-library(tidyverse)
-sel = casualties_2017$accident_index %in% crashes_leeds$accident_index
-casualties_leeds = casualties_2017[sel, ]
[…]
