+++ vignettes/references.bib
@@ -61,3 +61,26 @@
   pages = {510-525},
   file = {/home/robin/Zotero/storage/CKF3JGXX/Sarkar et al. - 2018 - Street morphology and severity of road casualties.pdf;/home/robin/Zotero/storage/FNTF9I3H/15568318.2017.html},
 }
+
+@article{feleke_comparative_2018,
+  title = {Comparative Fatality Risk for Different Travel Modes by Age, Sex, and Deprivation},
+  volume = {8},
+  issn = {22141405},
+  doi = {10.1016/j.jth.2017.08.007},
+  abstract = {Background: Cycling is perceived as an unsafe travel mode in many countries. However, road deaths in England have fallen sharply since 2007. We explored whether differences in fatality rates by age, gender and mode persist, and the associations of deprivation with these. Methods: Using ONS (cycling, pedestrian) and Stats19 (driving) 2007\textendash{}2012 data for travel-related deaths, including pedestrian falls, and National Travel Surveys 2007\textendash{}2012 travel data, we calculated fatality rates for England by distance (f/bnkm) and time travelled (million hours' use, f/mhu) by age, travel mode, and gender or residential Index of Multiple Deprivation. Results: Fatality rates fell significantly 2007\textendash{}2009 to 2010\textendash{}2012: male f/bnkm from 2.8 (95\%CI 2.7\textendash{}2.9) to 2.0 (1.9\textendash{}2.1) for driving; 32.1 (28.5\textendash{}36.0) to 20.8 (18.1\textendash{}23.9) for cycling; and 51.4 (48.5\textendash{}54.4) to 36.7 (34.3\textendash{}39.3) for walking. Fatality rates varied by age, gender, and mode. Driving and walking fatality rate ratios were generally higher for males than females. For males 17\textendash{}20y, fatality rates were 0.76 (0.69\textendash{}0.83)/mhu for driving and 0.28 (0.18\textendash{}0.42)/mhu for cycling but were similar by distance. Age-specific rates were J-shape for cycling, U-shape for driving, and increased exponentially with age for walking. Fatality rates aged 80+ were an order of magnitude higher in each mode than the all-age mean. Compared with those aged 17\textendash{}20, rate ratios were significantly lower for male drivers 21+ and female drivers 21\textendash{}74, but were higher for male cyclists aged 55+ and pedestrians 45+ (male) and 65+ (female). People living in the most deprived quintile generally had higher fatality rates than those in the least deprived quintile overall (three modes combined) and for walking but not for cycling; Rate ratios were highest for pedestrians 35\textendash{}64 and drivers 35\textendash{}54. Conclusions: Fatality rates for walking, cycling and driving are higher for males than females at almost every age and vary more by age than by travel mode. Deprivation exacerbates walking and driving fatality rates.},
+  language = {English},
+  journal = {Journal of Transport \& Health},
+  author = {Feleke, Robel and Scholes, Shaun and Wardlaw, Malcolm and Mindell, Jennifer S.},
+  month = mar,
+  year = {2018},
+  pages = {307-320},
+  file = {/home/robin/Zotero/storage/5VD6FC5Z/S2214140517301457.html}
+}
+
+@book{elvik_handbook_2009,
+  title = {The Handbook of Road Safety Measures},
+  publisher = {{Emerald Group Publishing}},
+  author = {Elvik, Rune and Vaa, Truls and Erke, Alena and Sorensen, Michael},
+  year = {2009},
+  file = {/home/robin/Zotero/storage/4F7KDLLY/Elvik et al. - 2009 - The handbook of road safety measures.pdf;/home/robin/Zotero/storage/ABXBEI7Y/books.html}
+}
