### 2018-12-30 dc05436
Update responses
