### 2018-12-30 b45c561
Tidy code - remove whitespace
