+++ man/file_names.Rd
@@ -16,7 +16,7 @@ The file names were generated as follows:
 library(rvest)
 #> Loading required package: xml2
 library(stringr)
-page <- read_html("https://data.gov.uk/dataset/cb7ae6f0-4be6-4935-9277-47e5ce24a11f/road-safety-data")
+page = read_html("https://data.gov.uk/dataset/cb7ae6f0-4be6-4935-9277-47e5ce24a11f/road-safety-data")
 
 r = page %>%
   html_nodes("a") %>%       # find all links
