### 2018-12-30 326d69b
Tidy code - remove whitespace
