### 2018-12-30 #266 @adamhsparks
Sorry, I missed the paper.md in the root dir. I've updated my review to reflect this.
