### 2018-12-30 1ee2751
Update DESCRIPTION
Add Mark's ORCID and flesh out Description field, including link to the data as per rOpenSci guidelines.
