### 2018-12-30 485e108
Spellcheck paper.md file
