+++ paper.md
@@ -18,7 +18,7 @@ authors:
    orcid: 0000-0003-2172-5265
    affiliations: 2
 affiliations:
- - name: Insititute for Transport Studies (ITS) and Leeds Institute for Data Analytics (LIDA), University of Leeds
+ - name: Institute for Transport Studies (ITS) and Leeds Institute for Data Analytics (LIDA), University of Leeds
    index: 1
  - name: ATFutures GmbH.
    index: 2
@@ -40,7 +40,7 @@ Finding, reading-in and formatting the data for research can be a time consuming
 Functions for each stage are named `dl_stats19()`, `read_*()` and `format_*()`, with `*` representing the type of data to be read-in: STATS19 data consists of `accidents`, `casualties` and `vehicles` tables, which correspond to incident records, people injured or killed, and vehicles involved, respectively.
 
 The package is needed because currently downloading and formatting STATS19 data is a time-consuming and error-prone process.
-By abstracting the process to its fundamental steps (donwload, read, format), the package makes it easy to get the data into appropriate formats (of classes `tbl`, `data.frame` and `sf`), ready for for further processing and analysis steps.
+By abstracting the process to its fundamental steps (download, read, format), the package makes it easy to get the data into appropriate formats (of classes `tbl`, `data.frame` and `sf`), ready for for further processing and analysis steps.
 We developed the package for road safety research, building on a clear need for reproducibility in the field [@lovelace_who_2016] and the importance of the geo-location in STATS19 data for assessing the effectiveness of interventions aimed to make roads safer and save lives [@sarkar_street_2018].
 A useful feature of the package is that it enables creation of geographic representations of the data, geo-referenced to the correct coordinate reference system, in a single function call (`format_sf()`).
 The package will be of use and interest to road safety data analysts working at local authority and national levels in the UK.
