### 2018-12-30 #266 @Robinlovelace
One question, probably more for @sckott: does JOSS accept figures and code chunks? Thinking that could liven it up a bit...
