### 2018-12-30 #266 @Robinlovelace
OK just merged those 2 mini PRs. That raises another question: are both reviewers happy to be in the author list with the new `rev` tag? More info here: https://ropensci.org/blog/2018/03/16/thanking-reviewers-in-metadata/
