### 2018-12-31 #266 @daranzolin
@Robinlovelace `rev` is great for me! Thanks for asking.
