### 2019-01-01 9405d0b
First attempt at get_stats19
