+++ NAMESPACE
@@ -7,6 +7,7 @@ export(format_casualties)
 export(format_column_names)
 export(format_sf)
 export(format_vehicles)
+export(get_stats19)
 export(locate_one_file)
 export(read_accidents)
 export(read_casualties)
