### 2019-01-01 517847f
First attempt at get_stats19
