### 2019-01-02 #266 @Robinlovelace
Great, I've add both reviewers in https://github.com/ITSLeeds/stats19/pull/67/files
