### 2019-01-02 cf2dad0
A default 'type' might be necessary for get_stats19, RFC commit
