### 2019-01-02 9aeac89
Making type required solves possible wrong type choice by user, test cov and consistent skip_download
