### 2019-01-02 67a3fa9
Making type required solves possible wrong type choice by user, test cov and consistent skip_download
