### 2019-01-02 #266 @sckott
@Robinlovelace will ask, not sure about the JOSS question
