### 2019-01-02 ITSLeeds/stats19#63 @sckott
regarding the `ukboundaries` not on CRAN thing, some have had luck with putting pkgs in a drat repo then adding the drat repository url to the DESCRIPTION file
