### 2019-01-02 ITSLeeds/stats19#63 @Robinlovelace
Thanks for the supportive comment @sckott. Good news: the issue was fixed by adding a new dataset `police_boundaries` to the package (not in `ukboundaries`, yet...) that is actually much more relevant to the package. Will update the response accordingly!
