### 2019-01-03 4822e6d
Correct test flow for get_stats19
