### 2019-01-04 aae6769
Readme, dl_stats19: update readme with get_stats19 and fix 'yes|y' interactivity, #63, .md fails.
