+++ man/get_stats19.Rd
@@ -5,7 +5,7 @@
 \title{Download, read and format STATS19 data in one function.}
 \usage{
 get_stats19(year = NULL, type = "accidents", data_dir = tempdir(),
-  file_name = NULL)
+  file_name = NULL, format = TRUE)
 }
 \arguments{
 \item{year}{Single year for which file is to be downloaded.}
@@ -16,6 +16,8 @@ Or any variation of to search the file names with such as "acc" or "accid".}
 \item{data_dir}{Parent directory for all downloaded files. Defaults to `tempdir()`.}
 
 \item{file_name}{The file name (DfT named) to download.}
+
+\item{format}{Switch to return raw read from file, default is `TRUE`.}
 }
 \description{
 Download, read and format STATS19 data in one function.
