+++ data-raw/schema.Rmd
@@ -1,0 +1,188 @@
+
+WARNING: this does not work on Windows.
+
+This function generates the data object `stats19_schema` in a reproducible way
+using DfT's schema definition (see function [dl_schema()]).
+
+The function also generates `stats19_variables` (see the function's source code
+for details).
+
+# Load stats19 schema
+
+```{r}
+read_schema = function(
+  data_dir = tempdir(),
+  filename = "Road-Accident-Safety-Data-Guide.xls",
+  sheet = NULL
+) {
+
+  file_path = file.path(data_dir, filename)
+  if(!file.exists(file_path)) {
+    dl_schema()
+  }
+  if(is.null(sheet)) {
+    export_variables = readxl::read_xls(path = file_path,
+                                        sheet = 2,
+                                        skip = 2)
+    export_variables_accidents = tibble::tibble(
+      table = "accidents",
+      variable = export_variables$`Accident Circumstances`
+    )
+    export_variables_vehicles =  tibble::tibble(
+      table = "vehicles",
+      variable = export_variables$Vehicle
+    )
+    export_variables_casualties =  tibble::tibble(
+      table = "casualties",
+      variable = export_variables$Casualty
+    )
+    export_variables_long = rbind(
+      export_variables_accidents,
+      export_variables_casualties,
+      export_variables_vehicles
+    )
+    stats19_variables = stats::na.omit(export_variables_long)
+    stats19_variables$type = stats19_vtype(stats19_variables$variable)
+
+    # add variable linking to names in formatted data
+    names_acc = names(accidents_sample)
+    names_veh = names(vehicles_sample)
+    names_cas = names(casualties_sample)
+    names_all = c(names_acc, names_veh, names_cas)
+
+    variables_lower = schema_to_variable(stats19_variables$variable)
+    # test result
+    # variables_lower[!variables_lower %in% names_all]
+    # names_all[!names_all %in% variables_lower]
+    stats19_variables$column_name = variables_lower
+    # head(stats19_variables)
+
+    #' export result: usethis::use_data(stats19_variables, overwrite = TRUE)
+
+    sel_character = stats19_variables$type == "character"
+    character_vars = stats19_variables$variable[sel_character]
+    character_vars = stats19_vname_switch(character_vars)
+
+    schema_list = lapply(
+      X = seq_along(character_vars),
+      FUN = function(i) {
+        x = readxl::read_xls(path = file_path, sheet = character_vars[i])
+        names(x) = c("code", "label")
+        x
+      }
+    )
+
+    stats19_schema = do.call(what = rbind, args = schema_list)
+    n_categories = vapply(schema_list, nrow, FUN.VALUE = integer(1))
+    stats19_schema$variable = rep(character_vars, n_categories)
+
+    character_cols = stats19_variables$column_name[sel_character]
+    stats19_schema$variable_formatted = rep(character_cols, n_categories)
+
+  } else {
+    stats19_schema = readxl::read_xls(path = file_path, sheet = sheet)
+  }
+  stats19_schema
+}
+```
+
+# Download schema
+
+```{r}
+dl_schema = function(data_dir = tempdir()) {
+u = paste0(
+  "http://data.dft.gov.uk/road-accidents-safety-data/",
+  "Road-Accident-Safety-Data-Guide.xls"
+)
+destfile = file.path(data_dir, "Road-Accident-Safety-Data-Guide.xls")
[…]
