### 2019-01-04 bd42138
Format help files
