+++ man/stats19_schema.Rd
@@ -7,13 +7,13 @@
 \title{Stats19 schema and variables}
 \description{
 `stats19_schema` and `stats19_variables` contain
-metadata on stats19 data.
+metadata on \pkg{stats19} data.
 `stats19_schema` is a look-up table matching
 codes provided in the raw stats19 dataset with
 character strings.
 }
 \note{
-The schema data can be (re-)generated using the script given in the
+The schema data can be (re-)generated using the script in the
 `data-raw` directory.
 }
 \keyword{datasets}
