+++ man/read_casualties.Rd
@@ -2,7 +2,7 @@
 % Please edit documentation in R/read.R
 \name{read_casualties}
 \alias{read_casualties}
-\title{Read in stats19 road safety data from .csv files downloaded.}
+\title{Read in STATS19 road safety data from .csv files downloaded.}
 \usage{
 read_casualties(year = NULL, filename = "", data_dir = tempdir(),
   format = TRUE)
@@ -18,12 +18,12 @@ years determine whether there is a target to read, otherwise disk scan would be 
 \item{format}{Switch to return raw read from file, default is `TRUE`.}
 }
 \description{
-Read in stats19 road safety data from .csv files downloaded.
+Read in STATS19 road safety data from .csv files downloaded.
 }
 \section{Details}{
 
 The function returns a data frame, in which each record is a reported casualty
-in the stats19 dataset.
+in the STATS19 dataset.
 }
 
 \examples{
