+++ man/file_names.Rd
@@ -12,16 +12,17 @@ how these were obtained and would be kept up to date.
 }
 \details{
 The file names were generated as follows:
-
+}
+\examples{
+\dontrun{
 library(rvest)
-#> Loading required package: xml2
 library(stringr)
 page = read_html("https://data.gov.uk/dataset/cb7ae6f0-4be6-4935-9277-47e5ce24a11f/road-safety-data")
 
-r = page %>%
-  html_nodes("a") %>%       # find all links
-  html_attr("href") %>%     # get the url
-  str_subset("\\.zip")
+r = page \%>\%
+  html_nodes("a") \%>\%       # find all links
+  html_attr("href") \%>\%     # get the url
+  str_subset("\\\\.zip")
 
 dr = c()
 for(i in 1:length(r)) {
@@ -32,5 +33,6 @@ for(i in 1:length(r)) {
 }
 file_names = setNames(as.list(file_names), file_names)
 usethis::use_data(file_names)
+}
 }
 \keyword{datasets}
