+++ man/format_sf.Rd
@@ -7,10 +7,10 @@
 format_sf(x, lonlat = FALSE)
 }
 \arguments{
-\item{x}{Data frame created with `read_accidents()`}
+\item{x}{Data frame created with \code{read_accidents()}}
 
 \item{lonlat}{Should the results be returned in longitude/latitude?
-By default `FALSE`, meaning the British National Grid (EPSG code: 27700)
+By default \code{FALSE}, meaning the British National Grid (EPSG code: 27700)
 is used.}
 }
 \description{
