+++ DESCRIPTION
@@ -44,3 +44,4 @@ Suggests:
     tidyr
 VignetteBuilder: knitr
 RoxygenNote: 6.1.1
+Roxygen: list(markdown = TRUE)
