+++ man/dl_stats19.Rd
@@ -26,8 +26,8 @@ This function downloads and unzips UK road crash data.
 It results in unzipped .csv files that are put
 in the temporary directory specified by \code{tempdir()} or provided \code{data_dir}.
 
-The file downloaded would be for a specific year (e.g 2017).
-It could also be a file containing data for a range of two (e.g 2005-2014).
+The file downloaded would be for a specific year (e.g. 2017).
+It could also be a file containing data for a range of two (e.g. 2005-2014).
 
 The \code{dl_*} functions can download many MB of data so ensure you
 have a sufficient internet access and hard disk space.
