+++ R/dl.R
@@ -21,7 +21,7 @@
 #'
 #' @export
 #' @examples
-#' \dontrun{
+#' \donttest{
 #' dl_stats19(year = 2017) # interactively select files...
 #' # now you can read-in the data
 #' dl_stats19(year = 2009)
