+++ R/read.R
@@ -13,7 +13,7 @@
 #'
 #' @export
 #' @examples
-#' \dontrun{
+#' \donttest{
 #' dl_stats19(year = 2017, type = "Accidents")
 #' ac = read_accidents(year = 2017)
 #' dl_stats19(year = 2009, type = "Accidents")
@@ -63,7 +63,7 @@ read_accidents = function(year = NULL,
 #'
 #' @export
 #' @examples
-#' \dontrun{
+#' \donttest{
 #' dl_stats19(year = 2009, type = "vehicles")
 #' ve = read_vehicles(year = 2009)
 #' }
@@ -96,7 +96,7 @@ read_vehicles = function(year = NULL,
 #'
 #' @export
 #' @examples
-#' \dontrun{
+#' \donttest{
 #' dl_stats19(year = 2017, type = "casualties")
 #' casualties = read_casualties(year = 2017)
 #' }
