+++ man/format_casualties.Rd
@@ -18,7 +18,7 @@ This function formats raw STATS19 data
 }
 
 \examples{
-\dontrun{
+\donttest{
 dl_stats19(year = 2017, type = "casualties")
 x = read_casualties(year = 2017)
 casualties = format_casualties(x)
