+++ man/format_accidents.Rd
@@ -18,7 +18,7 @@ This is a helper function to format raw STATS19 data
 }
 
 \examples{
-\dontrun{
+\donttest{
 dl_stats19(year = 2017, type = "accident")
 x = read_accidents(year = 2017)
 crashes = format_accidents(x)
