### 2019-01-04 f366bc3
Add call. arg to call funs
