+++ R/get.R
@@ -29,7 +29,7 @@ get_stats19 = function(year = NULL,
                       file_name = NULL,
                       format = TRUE) {
   if(!exists("type"))
-    stop("Type is required")
+    stop("Type is required", call. = FALSE)
   # download what the user wanted
   dl_stats19(year = year,
              type = type,
