+++ pkgdown.yml
@@ -1,0 +1,6 @@
+pandoc: 2.3.1
+pkgdown: 1.3.0
+pkgdown_sha: ~
+articles:
+  stats19: stats19.html
+
