+++ DESCRIPTION
@@ -13,12 +13,12 @@ Authors@R: c(
     person("Adam", "Sparks", role = c("rev", "ctb"),
       comment = c(ORCID = "0000-0002-0061-8359"))
   )
-Description: Tools to help process and analysie the UK road collision data collected using the
-  STATS19 form.  The data are provided as 'CSV' files with detailed road safety data about the
+Description: Tools to help process and analyse the UK road collision data collected using the
+  'STATS19' form. The data are provided as 'CSV' files with detailed road safety data about the
   circumstances of personal injury road accidents in Great Britain from 1979, the types
   (including make and model) of vehicles involved and the consequential casualties.  The
   statistics relate only to personal injury accidents on public roads that are reported
-  to the police, and subsequently recorded, using the STATS19 accident reporting form. See
+  to the police, and subsequently recorded, using the 'STATS19' accident reporting form. See
   the Department for Transportation website 
   <https://data.gov.uk/dataset/cb7ae6f0-4be6-4935-9277-47e5ce24a11f/road-safety-data> for more
   information on these data.
@@ -33,7 +33,6 @@ Imports:
     readr
 Suggests: 
     curl (>= 3.2),
-    readxl,
     dplyr,
     ggplot2,
     knitr,
@@ -45,3 +44,4 @@ Suggests:
 VignetteBuilder: knitr
 RoxygenNote: 6.1.1
 Roxygen: list(markdown = TRUE)
+Language: en-US
