### 2019-01-04 efb4709
Use remotes::install_github()
