### 2019-01-04 1458963
Use remotes::install_github()
