+++ README.Rmd
@@ -25,31 +25,16 @@ knitr::opts_chunk$set(
 # stats19
 
 **stats19** provides functions for downloading and formatting road crash data.
-Specifically, it enables access to the UK's official road traffic casualty database, [STATS19](https://data.gov.uk/dataset/cb7ae6f0-4be6-4935-9277-47e5ce24a11f/road-safety-data). (The name comes from the form used by the police to record car crashes and other incidents resulting in casualties on the roads).
+Specifically, it enables access to the UK's official road traffic casualty database, [STATS19](https://data.gov.uk/dataset/cb7ae6f0-4be6-4935-9277-47e5ce24a11f/road-safety-data). (The name comes from the form used by the police to record car crashes and other incidents resulting in casualties on the roads.)
 
-**Note**: The Department for Transport refers to "accidents", but "crashes" is a more appropriate term, as emphasised in the "crash not accident" arguments of road safety advocacy groups such as [RoadPeace](http://www.roadpeace.org/take-action/crash-not-accident/).
-We use the term "accidents" only in reference to nomenclature within the data as provided.
+A full overview of STATS19 variables be found in a [document](http://data.dft.gov.uk/road-accidents-safety-data/Brief-guide-to%20road-accidents-and-safety-data.doc) provided by the UK's Department for Transport (DfT).
 
-The data is provided by the UK's Department for Transport (DfT) as a series of `.csv` files that are stored in dozens of `.zip` files.
+The raw data is provided as a series of `.csv` files that are stored in dozens of `.zip` files.
 Finding, reading-in and formatting the data for research can be a time consuming process subject to human error.
-**stats19** speeds up these vital data access and cleaning stages of the research process, enabling reproducibility, by streamlining the work into 3 stages:
+**stats19** speeds up these vital but boring and error-prone stages of the research process with a single function: `get_stats19()`.
+By allowing public access to properly labelled road crash data, **stats19** aims to make road safety research more reproducible and accessible.
 
-- **Download**: This single `dl_stats19()` function accepts `year`, `type` and `filename` arguments to make it easy to find the right file for your research question.
-
-- **Read**: STATS19 data is provided in a particular format that benefits from being read-in with pre-specified column types. This is taken care of with `read_*()` functions providing access to the 3 main tables in STATS19 data:
-
-  - `read_accidents()` reads-in the crash data (which has one row per incident)
-  - `read_casualties()` reads-in the casualty data (which has one row per person injured or killed)
-  - `read_vehicles()` reads-in the vehicles table, which contains information on the vehicles involved in the crashes (and has one row per vehicle)
-  
-- **Format**: Each of the `read_*()` functions has a `format` parameter which, when `TRUE`, adds labels to the tables. The raw data provided by the DfT contains only integers. 
-Running `read_*(..., format = TRUE)` converts these integer values to the corresponding character variables for each of the three tables.
-For example, `read_accidents(format = TRUE)` converts values in the `accident_severity` column from `1`, `2` and `3` to `Slight`, `Serious` and `Fatal`.
-To read-in raw data without formatting, set `format = FALSE`.
-
-There is a wrapper function combining these three stages in one function called `get_stats19`.
-
-A full description of the stats19 data and variables they contain can be found in a [document](http://data.dft.gov.uk/road-accidents-safety-data/Brief-guide-to%20road-accidents-and-safety-data.doc) provided by the UK's Department for Transport (DfT).
[…]
