### 2019-01-04 fe6ca6c
Update, simplify README
