+++ README.Rmd
@@ -29,7 +29,7 @@ Specifically, it enables access to the UK's official road traffic casualty datab
 
 A full overview of STATS19 variables be found in a [document](http://data.dft.gov.uk/road-accidents-safety-data/Brief-guide-to%20road-accidents-and-safety-data.doc) provided by the UK's Department for Transport (DfT).
 
-The raw data is provided as a series of `.csv` files that are stored in dozens of `.zip` files.
+The raw data is provided as a series of `.csv` files that contain integers and which are stored in dozens of `.zip` files.
 Finding, reading-in and formatting the data for research can be a time consuming process subject to human error.
 **stats19** speeds up these vital but boring and error-prone stages of the research process with a single function: `get_stats19()`.
 By allowing public access to properly labelled road crash data, **stats19** aims to make road safety research more reproducible and accessible.
@@ -113,39 +113,23 @@ Code to read-in and format each of these tables is demonstrated below.
 ### Crash data
 
 Crash data was downloaded and read-in using the function `get_stats19()`, as described above.
-Crash data can also be read-in as follows:
+Crash data can also be read-in as follows (assuming the dataset has already been downloaded):
 
 ```{r read2017-raw-format}
-crashes_2017_raw = read_accidents(year = 2017, format = FALSE)
-crashes_2017 = format_accidents(crashes_2017_raw)
-nrow(crashes_2017_raw)
-ncol(crashes_2017_raw)
+crashes_2017 = read_accidents(year = 2017)
 nrow(crashes_2017)
 ncol(crashes_2017)
 ```
 
-The formatting work was done by `read_accidents(format = FALSE)`, which imported the "raw" Stats19 data without cleaning messy column names or re-categorising the outputs.
-`format_accidents()` automates the process of matching column names with variable names and labels in a [`.xls` file](http://data.dft.gov.uk/road-accidents-safety-data/Road-Accident-Safety-Data-Guide.xls) provided by the DfT.
-This means `crashes_2017` is much more usable than `crashes_2017_raw`, as shown below, which shows some key variables in the messy and clean datasets:
+Some of the key variables in this dataset include:
 
 ```{r crashes2017-columns}
-crashes_2017_raw[c(7, 18, 23, 25)]
 crashes_2017[c(7, 18, 23, 25)]
 ```
 
-By default, `format = TRUE`, meaning that the two stages of `read_accidents(format = FALSE)` and `format_accidents()` yield the same result as `read_accidents(format = TRUE)`.
 For the full list of columns, run `names(crashes_2017)` or see the [vignette](https://github.com/ITSLeeds/stats19/blob/master/vignettes/stats19.Rmd).
 
 <!-- This means `crashes_2017` is much more usable than `crashes_2017_raw`, as shown below, which shows three records and some key variables in the messy and clean datasets: -->
-
-```{r, echo=FALSE, eval=FALSE}
-# commented out as confusing...
-key_patt = "severity|speed|light|human"
-key_vars = grep(key_patt, x = names(crashes_2017_raw), ignore.case = TRUE)
-random_n = sample(x = nrow(crashes_2017_raw), size = 3)
-crashes_2017_raw[random_n, key_vars]
-crashes_2017[random_n, key_vars]
-```
 
 ## Casualties data
 
