+++ README.md
@@ -28,13 +28,14 @@ A full overview of STATS19 variables be found in a
 [document](http://data.dft.gov.uk/road-accidents-safety-data/Brief-guide-to%20road-accidents-and-safety-data.doc)
 provided by the UK’s Department for Transport (DfT).
 
-The raw data is provided as a series of `.csv` files that are stored in
-dozens of `.zip` files. Finding, reading-in and formatting the data for
-research can be a time consuming process subject to human error.
-**stats19** speeds up these vital but boring and error-prone stages of
-the research process with a single function: `get_stats19()`. By
-allowing public access to properly labelled road crash data, **stats19**
-aims to make road safety research more reproducible and accessible.
+The raw data is provided as a series of `.csv` files that contain
+integers and which are stored in dozens of `.zip` files. Finding,
+reading-in and formatting the data for research can be a time consuming
+process subject to human error. **stats19** speeds up these vital but
+boring and error-prone stages of the research process with a single
+function: `get_stats19()`. By allowing public access to properly
+labelled road crash data, **stats19** aims to make road safety research
+more reproducible and accessible.
 
 For transparency and modularity, each stage can be undertaken
 seperately, as documented in the [stats19
@@ -78,9 +79,9 @@ crashes = get_stats19(year = 2017, type = "accident")
 #> Files identified: dftRoadSafetyData_Accidents_2017.zip
 #> Attempt downloading from:
 #>    http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/dftRoadSafetyData_Accidents_2017.zip
-#> Data saved at /tmp/RtmpcjzbHt/dftRoadSafetyData_Accidents_2017/Acc.csv
+#> Data saved at /tmp/RtmpbUnCl0/dftRoadSafetyData_Accidents_2017/Acc.csv
 #> Reading in:
-#> /tmp/RtmpcjzbHt/dftRoadSafetyData_Accidents_2017/Acc.csv
+#> /tmp/RtmpbUnCl0/dftRoadSafetyData_Accidents_2017/Acc.csv
 ```
 
 What just happened? We read-in data on all road crashes recorded by the
@@ -98,12 +99,12 @@ casualties = get_stats19(year = 2017, type = "casualties")
 #> Files identified: dftRoadSafetyData_Casualties_2017.zip
 #> Attempt downloading from:
 #>    http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/dftRoadSafetyData_Casualties_2017.zip
-#> Data saved at /tmp/RtmpcjzbHt/dftRoadSafetyData_Casualties_2017/Cas.csv
+#> Data saved at /tmp/RtmpbUnCl0/dftRoadSafetyData_Casualties_2017/Cas.csv
 vehicles = get_stats19(year = 2017, type = "vehicles")
 #> Files identified: dftRoadSafetyData_Vehicles_2017.zip
 #> Attempt downloading from:
 #>    http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/dftRoadSafetyData_Vehicles_2017.zip
-#> Data saved at /tmp/RtmpcjzbHt/dftRoadSafetyData_Vehicles_2017/Veh.csv
+#> Data saved at /tmp/RtmpbUnCl0/dftRoadSafetyData_Vehicles_2017/Veh.csv
 ```
 
 The package also allows STATS19 files to be downloaded and read-in
@@ -147,48 +148,21 @@ Code to read-in and format each of these tables is demonstrated below.
 
 Crash data was downloaded and read-in using the function
 `get_stats19()`, as described above. Crash data can also be read-in as
-follows:
+follows (assuming the dataset has already been downloaded):
 
 ``` r
[…]
