### 2019-01-04 09e71ea
Remove pkgdown autodeploy
