### 2019-01-05 c5ade46
Updates for checks
