+++ R/data.R
@@ -35,7 +35,10 @@ NULL
 #' \dontrun{
 #' library(rvest)
 #' library(stringr)
-#' page = read_html("https://data.gov.uk/dataset/cb7ae6f0-4be6-4935-9277-47e5ce24a11f/road-safety-data")
+#' u = paste0(
+#'   "https://data.gov.uk/dataset/",
+#'   "cb7ae6f0-4be6-4935-9277-47e5ce24a11f/road-safety-data")
+#' page = read_html(u)
 #'
 #' r = page %>%
 #'   html_nodes("a") %>%       # find all links
