### 2019-01-05 b883f98
Revert "Remove pkgdown autodeploy"
This reverts commit 89ff5b133845f1aa0236eacdbe052c55a0d6e839.
