+++ .travis.yml
@@ -13,3 +13,9 @@ before_install:
 
 after_success:
   - Rscript -e 'covr::codecov()'
+
+before_deploy: Rscript -e 'remotes::install_cran("pkgdown")'
+deploy:
+  provider: script
+  script: Rscript -e 'pkgdown::deploy_site_github()'
+  skip_cleanup: true
