### 2019-01-05 2c9262a
Re-add pkgdown script to travis (test)
