+++ DESCRIPTION
@@ -40,7 +40,8 @@ Suggests:
     rmarkdown,
     stringr,
     testthat,
-    tidyr
+    tidyr,
+    pkgdown
 VignetteBuilder: knitr
 RoxygenNote: 6.1.1
 Roxygen: list(markdown = TRUE)
