### 2019-01-05 b16de04
Re-add pkgdown script to travis (test)
