+++ man/file_names.Rd
@@ -17,7 +17,10 @@ The file names were generated as follows:
 \dontrun{
 library(rvest)
 library(stringr)
-page = read_html("https://data.gov.uk/dataset/cb7ae6f0-4be6-4935-9277-47e5ce24a11f/road-safety-data")
+u = paste0(
+  "https://data.gov.uk/dataset/",
+  "cb7ae6f0-4be6-4935-9277-47e5ce24a11f/road-safety-data")
+page = read_html(u)
 
 r = page \%>\%
   html_nodes("a") \%>\%       # find all links
