+++ R/read.R
@@ -14,8 +14,9 @@
 #' @export
 #' @examples
 #' \donttest{
-#' dl_stats19(year = 2017, type = "Accidents")
-#' ac = read_accidents(year = 2017)
+#' dl_stats19(year = 2011, type = "Accidents")
+#' ac = read_accidents(year = 2011)
+#'
 #' dl_stats19(year = 2009, type = "Accidents")
 #' ac_2009 = read_accidents(year = 2009)
 #' }
