### 2019-01-05 c359a96
Fix example, fix travis (I hope)
