### 2019-01-05 dec373c
Fix example, fix travis (I hope)
