+++ vignettes/stats19.Rmd
@@ -24,7 +24,7 @@ knitr::opts_chunk$set(
 **stats19** enables access to and processing of the UK's official road traffic casualty database, [STATS19](https://data.gov.uk/dataset/cb7ae6f0-4be6-4935-9277-47e5ce24a11f/road-safety-data). 
 A description of variables in the database can be found in a [document](http://data.dft.gov.uk/road-accidents-safety-data/Brief-guide-to%20road-accidents-and-safety-data.doc) provided by the UK's Department for Transport (DfT).
 The datasets are collectively called STATS19 after the form used to report them, which can be found [here](http://docs.adrn.ac.uk/888043/mrdoc/pdf/888043_stats19-road-accident-injury-statistics-report-form.pdf). 
-This vignette focusses on how to use the **stats19** package to work with STATS19 data.
+This vignette focuses on how to use the **stats19** package to work with STATS19 data.
 
 **Note**: The Department for Transport refers to "accidents", but "crashes" is a more appropriate term, as emphasised in the "crash not accident" arguments of road safety advocacy groups such as [RoadPeace](http://www.roadpeace.org/take-action/crash-not-accident/).
 We use the term "accidents" only in reference to nomenclature within the data as provided.
@@ -56,9 +56,9 @@ To read-in raw data without formatting, set `format = FALSE`.
 
 Multiple functions are needed for each step because of the structure of STATS19 data, which are divided into 3 tables, as described in an article that analyses 2005-2015 STATS19 data with reproducible R code by Christoph Freier ([2018](https://kawameicha.gitlab.io/inCodeWeTrust/2018/01/10-years-and-3.3-million-accidented-vehicles-in-the-uk/)):
 
-1. "accident circumstances, with details about location, severity, weather, etc; 
+1. "accident circumstances, with details about location, severity, weather, etc.; 
 2. casualties, referencing knowledge about the victims; and
-3. vehicles, which contains more information about the vehicle type and maneuvers, as well the some information about the driver."
+3. vehicles, which contains more information about the vehicle type and manoeuvres, as well the some information about the driver."
 
 The data can be downloaded for many years.
 Datasets since 1979 are broadly consistent, meaning that STATS19 data represents a very rich historic geographic, as stated in the DfT's road casualties report in [2017](https://assets.publishing.service.gov.uk/government/uploads/system/uploads/attachment_data/file/744077/reported-road-casualties-annual-report-2017.pdf):
@@ -68,8 +68,8 @@ goes back to 1979, providing a long period for comparison.
 
 ## Download STATS19 data
 
-**stats19** enables download of raw stats19 data with `dl_*` functions.
-The following code chunk, for example, downloads and unzips a .zip file containing Stats19 data from 2017:
+**stats19** enables download of raw STATS19 data with `dl_*` functions.
+The following code chunk, for example, downloads and unzips a .zip file containing STATS19 data from 2017:
 
 ```{r dl2017-accidents}
 dl_stats19(year = 2017, type = "Accidents")
@@ -80,7 +80,7 @@ The `dl_stats19()` function prints out the location and final file name(s) of un
 
 `dl_stats19()` takes three parameters.
 Supplying a `file_name` is interpreted to mean that the user is aware of what to download and the other two parameters will be ignored.
-You can also use `years` and `type` to "search" through the file names, which are stored in a lazy-loaded datset called `stats19::file_names`.
+You can also use `years` and `type` to "search" through the file names, which are stored in a lazy-loaded dataset called `stats19::file_names`.
 
 You can find out the names of files that can be downloaded with `names(stats19::file_names)`, an example of which is shown below:
 
@@ -169,7 +169,7 @@ What just happened?
 We read-in data on all road crashes recorded by the police in 2017 across Great Britain.
[…]
