### 2019-01-06 8284d4f
Spellcheck on vignette, #63
