### 2019-01-06 70030a7
The rest of the checkpoints of vignette review responses in #63
