### 2019-01-06 a4c7df4
The rest of the checkpoints of vignette review responses in #63
