### 2019-01-06 159c872
The rest of the checkpoints of vignette review responses in #63
