### 2019-01-06 d71ce2c
Run format_sf() example
