### 2019-01-06 b3b3509
Update vignette and readme...
With better maps! Heads-up @Layik
