+++ vignettes/stats19.Rmd
@@ -21,7 +21,7 @@ knitr::opts_chunk$set(
 
 ## Introduction
 
-**stats19** enables access to and processing of the UK's official road traffic casualty database, [STATS19](https://data.gov.uk/dataset/cb7ae6f0-4be6-4935-9277-47e5ce24a11f/road-safety-data). 
+**stats19** enables access to and processing of Great Britain's official road traffic casualty database, [STATS19](https://data.gov.uk/dataset/cb7ae6f0-4be6-4935-9277-47e5ce24a11f/road-safety-data). 
 A description of variables in the database can be found in a [document](http://data.dft.gov.uk/road-accidents-safety-data/Brief-guide-to%20road-accidents-and-safety-data.doc) provided by the UK's Department for Transport (DfT).
 The datasets are collectively called STATS19 after the form used to report them, which can be found [here](http://docs.adrn.ac.uk/888043/mrdoc/pdf/888043_stats19-road-accident-injury-statistics-report-form.pdf). 
 This vignette focuses on how to use the **stats19** package to work with STATS19 data.
@@ -42,7 +42,13 @@ remotes::install_github("ITSLeeds/stats19")
 library(stats19)
 ```
 
-**stats19** has three main types of functions, corresponding with the three stages of data processing:
+## Functions
+
+The easiest way to get STATS19 data is with `get_stats19()`.
+This function takes 2 main arguments, `year` and `type`.
+The year can be any year between 1979 and 2017.
+The type can be one of `accidents`, `casualties` and `vehicles`, described below.
+`get_stats19()` performs 3 jobs, corresponding to three main types of functions:
 
 - **Download**: A `dl_stats19()` function accepts `year`, `type` and `filename` arguments to make it easy to find the right file to download only.
 
@@ -384,53 +390,86 @@ nrow(crashes_wy)
 ```
 
 This subsetting has selected the `r format(nrow(crashes_wy), big.mark = ",")`
-crashes which occurred within the Leeds boundary.
+crashes which occurred in West Yorkshire.
 
 
 ## Joining tables
 
 The three main tables we have just read-in can be joined by shared key variables.
-This is demonstrated in the code chunk below, which subsets all casualties that took place in Leeds, and counts the number of casualties by severity for each crash:
+This is demonstrated in the code chunk below, which subsets all casualties that took place in West Yorkshire, and counts the number of casualties by severity for each crash:
 
 ```{r table-join, message = FALSE}
 library(tidyr)
 library(dplyr)
 sel = casualties_2017$accident_index %in% crashes_wy$accident_index
-casualties_leeds = casualties_2017[sel, ]
-cas_types = casualties_leeds %>% 
+casualties_wy = casualties_2017[sel, ]
+cas_types = casualties_wy %>% 
   select(accident_index, casualty_type) %>% 
   mutate(n = 1) %>% 
   group_by(accident_index, casualty_type) %>% 
   summarise(n = sum(n)) %>% 
   spread(casualty_type, n, fill = 0) 
 cas_types$Total = rowSums(cas_types[-1])
-crashes_joined = left_join(crashes_wy, cas_types, by = "accident_index")
+crashes_joined = left_join(crashes_wy, cas_types)
 ```
 
-What just happened? We found the subset of casualties that took place in Leeds with reference to the `accident_index` variable.
+What just happened? We found the subset of casualties that took place in West Yorksire with reference to the `accident_index` variable.
 Then we used functions from the **tidyverse** package **dplyr** (and `spread()` from **tidyr**) to create a dataset with a column for each casualty type.
 We then joined the updated casualty data onto the `crashes_wy` dataset.
[…]
