### 2019-01-06 80e6f46
Update vignette and readme...
With better maps! Heads-up @Layik
