+++ README.md
@@ -79,9 +79,9 @@ crashes = get_stats19(year = 2017, type = "accident")
 #> Files identified: dftRoadSafetyData_Accidents_2017.zip
 #> Attempt downloading from:
 #>    http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/dftRoadSafetyData_Accidents_2017.zip
-#> Data saved at /tmp/RtmpbUnCl0/dftRoadSafetyData_Accidents_2017/Acc.csv
+#> Data saved at /tmp/RtmpV6HXXy/dftRoadSafetyData_Accidents_2017/Acc.csv
 #> Reading in:
-#> /tmp/RtmpbUnCl0/dftRoadSafetyData_Accidents_2017/Acc.csv
+#> /tmp/RtmpV6HXXy/dftRoadSafetyData_Accidents_2017/Acc.csv
 ```
 
 What just happened? We read-in data on all road crashes recorded by the
@@ -99,12 +99,12 @@ casualties = get_stats19(year = 2017, type = "casualties")
 #> Files identified: dftRoadSafetyData_Casualties_2017.zip
 #> Attempt downloading from:
 #>    http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/dftRoadSafetyData_Casualties_2017.zip
-#> Data saved at /tmp/RtmpbUnCl0/dftRoadSafetyData_Casualties_2017/Cas.csv
+#> Data saved at /tmp/RtmpV6HXXy/dftRoadSafetyData_Casualties_2017/Cas.csv
 vehicles = get_stats19(year = 2017, type = "vehicles")
 #> Files identified: dftRoadSafetyData_Vehicles_2017.zip
 #> Attempt downloading from:
 #>    http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/dftRoadSafetyData_Vehicles_2017.zip
-#> Data saved at /tmp/RtmpbUnCl0/dftRoadSafetyData_Vehicles_2017/Veh.csv
+#> Data saved at /tmp/RtmpV6HXXy/dftRoadSafetyData_Vehicles_2017/Veh.csv
 ```
 
 The package also allows STATS19 files to be downloaded and read-in
@@ -153,7 +153,7 @@ follows (assuming the dataset has already been downloaded):
 ``` r
 crashes_2017 = read_accidents(year = 2017)
 #> Reading in:
-#> /tmp/RtmpbUnCl0/dftRoadSafetyData_Accidents_2017/Acc.csv
+#> /tmp/RtmpV6HXXy/dftRoadSafetyData_Accidents_2017/Acc.csv
 nrow(crashes_2017)
 #> [1] 129982
 ncol(crashes_2017)
@@ -177,7 +177,7 @@ crashes_2017[c(7, 18, 23, 25)]
 #>  8 Slight                     30 Control by other autho… Darkness - lights…
 #>  9 Serious                    50 None within 50 metres   Darkness - lights…
 #> 10 Serious                    30 None within 50 metres   Darkness - lights…
-#> # ... with 129,972 more rows
+#> # … with 129,972 more rows
 ```
 
 For the full list of columns, run `names(crashes_2017)` or see the
@@ -196,7 +196,7 @@ dl_stats19(year = 2017, type = "casualties")
 #> Attempt downloading from:
 #>    http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/dftRoadSafetyData_Casualties_2017.zip
 #> Data already exists in data_dir, not downloading
-#> Data saved at /tmp/RtmpbUnCl0/dftRoadSafetyData_Casualties_2017/Cas.csv
[…]
