### 2019-01-06 3099c62
Fix typo in readme
