+++ README.Rmd
@@ -57,8 +57,7 @@ library(stats19)
 ## get_stats19()
 
 `get_stats19()` requires `year` and `type` parameters, mirroring the provision of STATS19 data files, which are categorised by year (from 1979 onwards) and type (with separate tables for crashes, casualties and vehicles, as outlined below).
-The following command, for example, gets crash data from 2017 (**note**: we use the name 'crashes', following the "crash not accident" campaign of [RoadPeace](http://www.roadpeace.org/take-action/crash-not-accident/).
-):
+The following command, for example, gets crash data from 2017 (**note**: we follow the "crash not accident" campaign of [RoadPeace](http://www.roadpeace.org/take-action/crash-not-accident/) in naming crashes, although the DfT refers to the relevant tables as 'accidents' data):
 
 ```{r}
 crashes = get_stats19(year = 2017, type = "accident")
