### 2019-01-06 #266 @layik
I learned a lot from this review @sckott, @daranzolin and @adamhsparks. I really appreciate all your contributions. Fantastic open source vibes.

Thanks @Robinlovelace and @mpadge awesome work.
