### 2019-01-06 ITSLeeds/stats19#63 @layik
https://github.com/ropensci/software-review/issues/266
