### 2019-01-06 1b74c35
Shorten README
