### 2019-01-06 fd928dd
Shorten README
