+++ README.md
@@ -80,9 +80,9 @@ crashes = get_stats19(year = 2017, type = "accident")
 #> Files identified: dftRoadSafetyData_Accidents_2017.zip
 #> Attempt downloading from:
 #>    http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/dftRoadSafetyData_Accidents_2017.zip
-#> Data saved at /tmp/RtmpTmFJL4/dftRoadSafetyData_Accidents_2017/Acc.csv
+#> Data saved at /tmp/Rtmpsc7HQK/dftRoadSafetyData_Accidents_2017/Acc.csv
 #> Reading in:
-#> /tmp/RtmpTmFJL4/dftRoadSafetyData_Accidents_2017/Acc.csv
+#> /tmp/Rtmpsc7HQK/dftRoadSafetyData_Accidents_2017/Acc.csv
 ```
 
 What just happened? We read-in data on all road crashes recorded by the
@@ -92,25 +92,12 @@ datasets provided by **stats19**, are outlined below and described in
 more detail in the [stats19
 vignette](https://itsleeds.github.io/stats19/articles/stats19.html).
 
-The following commands get the corresponding casualty and vehicle
-datasets for 2017:
-
-``` r
-casualties = get_stats19(year = 2017, type = "casualties")
-#> Files identified: dftRoadSafetyData_Casualties_2017.zip
-#> Attempt downloading from:
-#>    http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/dftRoadSafetyData_Casualties_2017.zip
-#> Data saved at /tmp/RtmpTmFJL4/dftRoadSafetyData_Casualties_2017/Cas.csv
-vehicles = get_stats19(year = 2017, type = "vehicles")
-#> Files identified: dftRoadSafetyData_Vehicles_2017.zip
-#> Attempt downloading from:
-#>    http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/dftRoadSafetyData_Vehicles_2017.zip
-#> Data saved at /tmp/RtmpTmFJL4/dftRoadSafetyData_Vehicles_2017/Veh.csv
-```
-
-The package also allows STATS19 files to be downloaded and read-in
-separately, allowing more control over what you download, and
-subsequently read-in, e.g. with `read_accidents()`.
+We will see below how the function also works to get the corresponding
+casualty and vehicle datasets for 2017. The package also allows STATS19
+files to be downloaded and read-in separately, allowing more control
+over what you download, and subsequently read-in, with
+`read_accidents()`, `read_casualties()` and `read_vehicles()`, as
+described in the vignette.
 
 ## Data download
 
@@ -125,16 +112,16 @@ dl_stats19(year = 2017)
 
     Multiple matches. Which do you want to download?
     
-    1: dftRoadSafetyData_Vehicles_2017.zip
-    2: dftRoadSafetyData_Casualties_2017.zip
+    1: dftRoadSafetyData_vehicles.zip
+    2: dftRoadSafetyData_casualties.zip
     3: dftRoadSafetyData_Accidents_2017.zip
     
     Selection: 
     Enter an item from the menu, or 0 to exit
 
-## Reading-in and formatting STATS19 data
+## Using the data
 
-As mentioned, STATS19 contains 3 main tables:
+STATS19 data consists of 3 main tables:
 
   - Accidents, the main table which contains information on the crash
     time, location and other variables (32 columns in total)
@@ -143,28 +130,24 @@ As mentioned, STATS19 contains 3 main tables:
   - Vehicles, containing data on vehicles involved in or causing each
     crash (23 columns in total)
[…]
