### 2019-01-06 #266 @adamhsparks
Thanks, @Robinlovelace, @layik and @mpadge. I'll have a look at the changes shortly and respond.
