### 2019-01-07 2524243
Update vignette plot
