+++ vignettes/stats19.Rmd
@@ -444,7 +444,7 @@ In other words, joins allow the casualties and vehicles tables to be geo-referen
 We can then explore the spatial distribution of different casualty types.
 The following figure, for example, shows the spatial distribution of pedestrians and car passengers hurt in car crashes across West Yorkshire in 2017:
 
-```{r sfplot, fig.show='hold', out.width="100%", fig.cap="Spatial distribution of serious and fatal crashes in West Yorkshire in which pedestrians and car passengers were hurt. Colour is related to the speed limit where the crash happened (red is faster) and size is proportional to the total number of people hurt in each crash (legend not shown)."}
+```{r sfplot, fig.show='hold', out.width="100%", fig.cap="Spatial distribution of serious and fatal crashes in West Yorkshire in which pedestrians and car passengers were hurt. Colour is related to the speed limit where the crash happened (red is faster) and size is proportional to the total number of people hurt in each crash (legend not shown).", fig.width=9, fig.height=7}
 library(ggplot2)
 crashes_types = crashes_joined %>% 
   filter(accident_severity != "Slight") %>% 
@@ -460,8 +460,9 @@ ggplot(crashes_types, aes(size = Total, colour = speed_limit)) +
     breaks = c(1:3, 12),
     labels = c(1:2, "3+", 12)
     ) +
-  scale_color_gradientn(colours = c("blue", "yellow", "red"), ) +
-  theme(legend.position = "none")
+  scale_color_gradientn(colours = c("blue", "yellow", "red"))
+ # +
+ #  theme(legend.position = "none")
 ```
 
 It is clear that pedestrians and car occupants tend to get hurt in different places.
