### 2019-01-07 764ae56
Updated readme figure
