+++ README.md
@@ -80,9 +80,9 @@ crashes = get_stats19(year = 2017, type = "accident")
 #> Files identified: dftRoadSafetyData_Accidents_2017.zip
 #> Attempt downloading from:
 #>    http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/dftRoadSafetyData_Accidents_2017.zip
-#> Data saved at /tmp/Rtmpz6teuo/dftRoadSafetyData_Accidents_2017/Acc.csv
+#> Data saved at /tmp/Rtmpbm2SeI/dftRoadSafetyData_Accidents_2017/Acc.csv
 #> Reading in:
-#> /tmp/Rtmpz6teuo/dftRoadSafetyData_Accidents_2017/Acc.csv
+#> /tmp/Rtmpbm2SeI/dftRoadSafetyData_Accidents_2017/Acc.csv
 ```
 
 What just happened? We read-in data on all road crashes recorded by the
@@ -179,7 +179,7 @@ casualties = get_stats19(year = 2017, type = "casualties")
 #> Files identified: dftRoadSafetyData_Casualties_2017.zip
 #> Attempt downloading from:
 #>    http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/dftRoadSafetyData_Casualties_2017.zip
-#> Data saved at /tmp/Rtmpz6teuo/dftRoadSafetyData_Casualties_2017/Cas.csv
+#> Data saved at /tmp/Rtmpbm2SeI/dftRoadSafetyData_Casualties_2017/Cas.csv
 nrow(casualties)
 #> [1] 170993
 ncol(casualties)
@@ -240,7 +240,7 @@ vehicles = get_stats19(year = 2017, type = "vehicles")
 #> Files identified: dftRoadSafetyData_Vehicles_2017.zip
 #> Attempt downloading from:
 #>    http://data.dft.gov.uk.s3.amazonaws.com/road-accidents-safety-data/dftRoadSafetyData_Vehicles_2017.zip
-#> Data saved at /tmp/Rtmpz6teuo/dftRoadSafetyData_Vehicles_2017/Veh.csv
+#> Data saved at /tmp/Rtmpbm2SeI/dftRoadSafetyData_Vehicles_2017/Veh.csv
 nrow(vehicles)
 #> [1] 238926
 ncol(vehicles)
@@ -393,16 +393,19 @@ enabling it to be mapped (for more advanced maps, see the
 [vignette](https://itsleeds.github.io/stats19/articles/stats19.html)):
 
 ``` r
-plot(crashes_joined["Total"], pch = 3, cex = 0.1)
+pal_alpha = function(n) sf.colors(n, alpha = 0.7)
+cex = crashes_joined$Total / 5
+plot(crashes_joined["speed_limit"], cex = cex, pal = pal_alpha)
 ```
 
 <img src="man/figures/README-unnamed-chunk-2-1.png" width="100%" />
 
 The spatial distribution of crashes in West Yorkshire clearly relates to
 the region’s geography. Crashes tend to happen on busy Motorway roads
-and city centres, of Leeds and Bradford in particular. This can be seen
-by comparing the previous map with an overview of the area, from
-(Lovelace, Roberts, and Kellar
+(with high speed limits, as shown in the map above) and city centres, of
+Leeds and Bradford in particular. This can be seen by comparing the
+previous map with an overview of the area, from (Lovelace, Roberts, and
+Kellar
 2016):
 
 <img src="https://raw.githubusercontent.com/ITSLeeds/stats19/master/vignettes/wy-overview.jpg" width="100%" />
