### 2019-01-07 a4af16a
Explain new map in readme
