+++ README.Rmd
@@ -230,13 +230,14 @@ crashes_joined[1:2, c(1, 5, 34)] %>% st_drop_geometry()
 The join operation added a geometry column to the casualty data, enabling it to be mapped (for more advanced maps, see the [vignette](https://itsleeds.github.io/stats19/articles/stats19.html)):
 
 ```{r}
-pal_alpha = function(n) sf.colors(n, alpha = 0.7)
-cex = crashes_joined$Total / 5
+pal_alpha = function(n) sf.colors(n, alpha = 0.5)
+cex = crashes_joined$Total / 3
 plot(crashes_joined["speed_limit"], cex = cex, pal = pal_alpha)
 ```
 
 The spatial distribution of crashes in West Yorkshire clearly relates to the region's geography.
 Crashes tend to happen on busy Motorway roads (with high speed limits, as shown in the map above) and city centres, of Leeds and Bradford in particular.
+The severity and number of people hurt (proportional to circle width in the map above) in crashes is related to the speed limit roads.
 This can be seen by comparing the previous map with an overview of the area, from [@lovelace_who_2016]:
 
 ```{r, echo=FALSE}
