### 2019-01-07 9929df6
Explain new map in readme
