### 2019-01-08 #266 @adamhsparks
Thank you @Robinlovelace, @layik and @mpadge for your consideration of my comments and the incorporation of many of them as you see fit. However, there are still a few issues (most of which you did attempt to previously address but seem to have slipped through) that I think should be fixed before I sign off on it.

1. Use `spelling::spell_check_package()` and correct the errors.

2. There's a line >80 chars. Not a hard-fast rule, but nice to observe and this one can be easily split over two lines. It is Line 149 of read.R.

3. In the data.R file there are several examples that say something like: "The file names were generated as follows:" IMO any example like this needs to be in the `data-raw` in Rmd files and only the data in the final package with proper documentation as to what the data are and a link to where they were obtained from. A link to the `data-raw` files would be appropriate in the documentation for each data set.

From Hadley's R packages - Data page: http://r-pkgs.had.co.nz/data.html

> Often, the data you include in data/ is a cleaned up version of raw data you’ve gathered from elsewhere. I highly recommend taking the time to include the code used to do this in the source version of your package. This will make it easy for you to update or reproduce your version of the data. I suggest that you put this code in data-raw/. **You don’t need it in the bundled version of your package**,...

Emphasis mine.

There is also an error in the examples for `police_boundaries`. I cannot reproduce the data set as you have in your example.

```r
  > library(sf)
  Linking to GEOS 3.7.1, GDAL 2.3.2, PROJ 5.2.0
  > u = paste0("https://opendata.arcgis.com/",
  +   "datasets/3e5a096a8c7c456fb6d3164a3f44b005_3.kml"
  +   )
  > police_boundaries_wgs = read_sf(u)
  Warning in CPL_read_ogr(dsn, layer, query, as.character(options), quiet,  :
    GDAL Error 1: JSON parsing error: unexpected character (at offset 0)
  > police_boundaries = st_transform(police_boundaries_wgs, 27700)
  > police_boundaries = police_boundaries[c("pfa16cd", "pfa16nm")]
  Error: Can't find columns `pfa16cd`, `pfa16nm` in `.data`.
  Execution halted
```

The warning with GDAL can be circumvented if I first download the file then import it? Not that it matters, the error still occurs since there are no columns named `pfa16cd` or `pfa16nm` in the data frame.
