### 2019-01-08 #266 @adamhsparks
One other suggestion for you is to add a CITATION file so that `cite("stats19")` can be used to generate a reference for citing it from R.
