### 2019-01-08 #266 @Robinlovelace
These all sound reasonable @adamhsparks. We'll get on it and hope to 'resubmit' by early next week. Note to self: add a link to show changes that have happened between now and then: https://help.github.com/articles/comparing-commits-across-time/
