### 2019-01-08 #266 @adamhsparks
If it helps, here's the CITATION file for `nasapower` that has a JOSS article and a manual (for the release version) both in it as well as info on citing the POWER data properly.

https://github.com/ropensci/nasapower/blob/master/inst/CITATION
