### 2019-01-08 b62e2b1
Second part of 2nd review, spelling::spell_check_package, #63
