### 2019-01-08 3f38d4c
Second part of 2nd review, spelling::spell_check_package, #63
