### 2019-01-08 4aed1d1
Second part of 2nd review, spelling::spell_check_package, #63
