+++ man/locate_one_file.Rd
@@ -9,7 +9,7 @@ locate_one_file(filename = NULL, data_dir = tempdir(), year = NULL,
 }
 \arguments{
 \item{filename}{Character string of the filename of the .csv to read, if this
-is given, type and years determin whether there is a target to read,
+is given, type and years determine whether there is a target to read,
 otherwise disk scan would be needed.}
 
 \item{data_dir}{Where sets of downloaded data would be found.}
