### 2019-01-08 ITSLeeds/stats19#63 @layik
I take pleasure of saying line 149 of read.R was 81 characters :)
