### 2019-01-08 4864b9a
Second part of 2nd review, tiny goodpractice 80 chrs, #63
