### 2019-01-08 9fc9736
Second part of 2nd review, tiny goodpractice 80 chrs, #63
