### 2019-01-08 1d83af9
Second part of 2nd review, tiny goodpractice 80 chrs, #63
