### 2019-01-08 403a6da
Second part of 2nd review, CITATION basics, #63
