### 2019-01-08 378266e
Second part of 2nd review, police_boundaries example fix, #63
