+++ DESCRIPTION
@@ -41,7 +41,8 @@ Suggests:
     stringr,
     testthat,
     tidyr,
-    pkgdown
+    pkgdown,
+    geojsonsf
 VignetteBuilder: knitr
 RoxygenNote: 6.1.1
 Roxygen: list(markdown = TRUE)
