### 2019-01-08 ITSLeeds/stats19#63 @layik
Now see if I can move the relevant bits to `raw` as per 3rd checkbox.
