### 2019-01-08 ITSLeeds/stats19#63 @layik
Sending a PR soon. `download.file()` for which sorry?
