### 2019-01-08 #266 @Robinlovelace
Great, thanks for showing us this example. I imagine the citation will change if/when it goes into JOSS.
