### 2019-01-08 b1688e2
Second part of 2nd review, move reproducibility from data.R to raw, #63
