### 2019-01-08 cc2f777
Second part of 2nd review, move reproducibility from data.R to raw, #63
