+++ man/vehicles_sample.Rd
@@ -11,13 +11,8 @@ Sample of stats19 data (2017 vehicles)
 }
 \examples{
 \donttest{
-# Obtained with:
-dl_stats19(year = 2017, type = "veh")
-vehicles_2017_raw = read_vehicles(year = 2017)
-set.seed(350)
-sel = sample(nrow(vehicles_2017_raw), 3)
-vehicles_sample_raw = vehicles_2017_raw[sel, ]
-vehicles_sample = format_vehicles(vehicles_sample_raw)
+nrow(vehicles_sample_raw)
+vehicles_sample_raw
 }
 }
 \keyword{datasets}
