### 2019-01-08 c4dfb16
Second part of 2nd review, move reproducibility from data.R to raw, #63
