### 2019-01-08 ITSLeeds/stats19#63 @layik
I see. Sure, can add commits to the branch.
@Robinlovelace, could you please check the "spellcheck" commit, too? I think it is done but you know what it is like.
