### 2019-01-08 fc8d5d5
Update spelling in readme
