+++ README.Rmd
@@ -34,7 +34,7 @@ Finding, reading-in and formatting the data for research can be a time consuming
 **stats19** speeds up these vital but boring and error-prone stages of the research process with a single function: `get_stats19()`.
 By allowing public access to properly labelled road crash data, **stats19** aims to make road safety research more reproducible and accessible.
 
-For transparency and modularity, each stage can be undertaken seperately, as documented in the [stats19 vignette](https://itsleeds.github.io/stats19/articles/stats19.html).
+For transparency and modularity, each stage can be undertaken separately, as documented in the [stats19 vignette](https://itsleeds.github.io/stats19/articles/stats19.html).
 
 ## Installation
 
@@ -56,7 +56,7 @@ library(stats19)
 
 ## get_stats19()
 
-`get_stats19()` requires `year` and `type` parameters, mirroring the provision of STATS19 data files, which are categorised by year (from 1979 onwards) and type (with separate tables for crashes, casualties and vehicles, as outlined below).
+`get_stats19()` requires `year` and `type` parameters, mirroring the provision of STATS19 data files, which are categorised by year (from 1979 onward) and type (with separate tables for crashes, casualties and vehicles, as outlined below).
 The following command, for example, gets crash data from 2017 (**note**: we follow the "crash not accident" campaign of [RoadPeace](http://www.roadpeace.org/take-action/crash-not-accident/) in naming crashes, although the DfT refers to the relevant tables as 'accidents' data):
 
 ```{r}
@@ -124,7 +124,7 @@ For the full list of columns, run `names(crashes)` or see the [vignette](https:/
 
 ### Casualties data
 
-As with `crashes`, casualty data for 2017 can be downloaded, read-in and formated as follows:
+As with `crashes`, casualty data for 2017 can be downloaded, read-in and formatted as follows:
 
 ```{r 2017-cas}
 casualties = get_stats19(year = 2017, type = "casualties")
@@ -147,7 +147,7 @@ names(casualties)
 
 ### Vehicles data
 
-Data for vehicles involved in crashes in 2017 can be downloaded, read-in and formated as follows:
+Data for vehicles involved in crashes in 2017 can be downloaded, read-in and formatted as follows:
 
 ```{r dl2017-vehicles}
 vehicles = get_stats19(year = 2017, type = "vehicles")
@@ -192,7 +192,7 @@ nrow(crashes_wy)
 ```
 
 This subsetting has selected the `r format(nrow(crashes_wy), big.mark = ",")`
-crashes which occurred within West Yorksire in 2017.
+crashes which occurred within West Yorkshire in 2017.
 
 
 ## Joining tables
