### 2019-01-08 ITSLeeds/stats19#74 @Robinlovelace
See https://github.com/ITSLeeds/stats19/commit/fc8d5d52a4529d9a6c888fba71fd41f3ddd81a6c
