### 2019-01-08 ITSLeeds/stats19#74 @Robinlovelace
Heads-up @adamhsparks I think the pkg is free from typos now. See the output below. If you spot any that are still wrong, please let use know. Ashamed that I misspelled Yorkshire, seeing it's where I live!

``` r
devtools::spell_check("/home/robin/ITSLeeds/stats19/")
#>   WORD             FOUND IN
#> acc              check_input_file.Rd:15
#>                  dl_stats19.Rd:14
#>                  get_stats19.Rd:14
#> accid            dl_stats19.Rd:14
#>                  get_stats19.Rd:14
#> accidents’     README.md:68
#> ae               description:8
#> amazonaws        get_url.Rd:26
#> analyse          description:1
#> Behaviour        README.md:492
#> bikeR            README.md:474
#>                  README.Rmd:294
#> bkm              stats19.Rmd:374
#> blahDIR          download_and_unzip.Rd:20,22
#> categorised      README.md:68
#>                  README.Rmd:59
#> categorising     stats19.Rmd:176
#> cb               description:8
#> CDRC             stats19.Rmd:531
#> ce               description:8
#> centre           stats19.Rmd:468,479
#> Centre           stats19.Rmd:531
#> centres          README.md:403
#>                  README.Rmd:238
#>                  stats19.Rmd:433
#> Chinmoy          README.md:498
#> Christoph        stats19.Rmd:66
#> codecov          README.md:9
#>                  README.Rmd:9
#> colour           stats19.Rmd:426
#> csv              check_input_file.Rd:11
#>                  dl_stats19.Rd:26
#>                  download_and_unzip.Rd:10
#>                  locate_one_file.Rd:11
#>                  read_accidents.Rd:5,13,21
#>                  read_casualties.Rd:5,13,21
#>                  read_vehicles.Rd:5,13,21
#> CyIPT            README.md:480
#>                  README.Rmd:296
#> df               get_stats19.Rd:28
#> dft              get_url.Rd:24,26
#> DfT              dl_stats19.Rd:18
#>                  file_names.Rd:10
#>                  get_stats19.Rd:18
#>                  README.md:27,68
#>                  README.Rmd:30,59
#>                  stats19.Rmd:24,61,87,132,151,156,176,199,203,239
#> DfT's            stats19.Rmd:72
#> dftRoadSafety    download_and_unzip.Rd:20,21,22
#>                  get_url.Rd:26
#> dir              read_vehicles.Rd:26
#> doi              README.md:492,501
#> dplyr            README.md:365
#>                  README.Rmd:220
#>                  stats19.Rmd:420
#> emphasised       stats19.Rmd:29
#> EPSG             format_sf.Rd:13
#>                  README.md:292
#>                  README.Rmd:173
#>                  stats19.Rmd:349
#> exdir            download_and_unzip.Rd:14
#> Freier           stats19.Rmd:66
#> geo              stats19.Rmd:441
#> geoportal        police_boundaries.Rd:9
#> ggplot           stats19.Rmd:471
#> github           stats19.Rmd:532
#> Gitter           README.md:11
#>                  README.Rmd:10
#> http             get_url.Rd:24,26
#>                  police_boundaries.Rd:9
#> https            description:8
#>                  README.md:492,501
#> ITSLeeds         stats19.Rmd:532
#> Kellar           README.md:403,452,488
#> Kumari           README.md:463,498
#> labelled         README.md:31
#>                  README.Rmd:32
#> manoeuvres       stats19.Rmd:70
#> mh               stats19.Rmd:374
#> modularity       README.md:40
#>                  README.Rmd:37
#> nd               get_url.Rd:28
#> normalised       stats19.Rmd:526
#> OpenStreetMap    stats19.Rmd:531
#> params           download_and_unzip.Rd:5
#>                  locate_files.Rd:25
#> pre              stats19.Rmd:55
#> repo             README.md:474
#>                  README.Rmd:294
#> retuns           get_stats19.Rd:28
#> RoadPeace        README.md:74
#>                  README.Rmd:60
#>                  stats19.Rmd:29
#> RoadSafetyData   get_url.Rd:24
#> RStudio          README.md:14
#>                  README.Rmd:12
#> Sarika           README.md:498
#> Sarkar           README.md:463,498
#> stplanr          README.md:477
#>                  README.Rmd:295
#> tibble           stats19.Rmd:144
#> tidyr            README.md:365
#>                  README.Rmd:220
#>                  stats19.Rmd:420
#> tidyverse        README.md:365
#>                  README.Rmd:220
#>                  stats19.Rmd:420
#> tmp              download_and_unzip.Rd:20,22
#> travelling       README.md:434
#>                  README.Rmd:265
#>                  stats19.Rmd:498
#> trf              README.md:492
#> uk               get_url.Rd:24,26
#>                  police_boundaries.Rd:9
#>                  description:8
#> UKDS             schema_original.Rd:6,9
#> visualising      stats19.Rmd:471
```
