### 2019-01-08 b04711f
Minor: remove geojsonsf suggest
