+++ man/vehicles_sample.Rd
@@ -9,6 +9,10 @@
 \description{
 Sample of stats19 data (2017 vehicles)
 }
+\note{
+These were generated using the script in the
+\code{data-raw} directory (\code{misc.Rmd} file).
+}
 \examples{
 \donttest{
 nrow(vehicles_sample_raw)
