+++ DESCRIPTION
@@ -41,8 +41,7 @@ Suggests:
     stringr,
     testthat,
     tidyr,
-    pkgdown,
-    geojsonsf
+    pkgdown
 VignetteBuilder: knitr
 RoxygenNote: 6.1.1
 Roxygen: list(markdown = TRUE)
