### 2019-01-08 c7b0f6e
Minor: remove geojsonsf suggest
