### 2019-01-08 078a38d
Minor: remove geojsonsf suggest
