+++ man/police_boundaries.Rd
@@ -8,6 +8,10 @@
 \description{
 See http://geoportal.statistics.gov.uk/
 }
+\note{
+These were generated using the script in the
+\code{data-raw} directory (\code{misc.Rmd} file).
+}
 \examples{
 \donttest{
 nrow(police_boundaries)
