### 2019-01-09 0b4862e
Improve documentation of polic_forces
