### 2019-01-09 811a520
Improve documentation of polic_forces
