### 2019-01-09 #266 @Robinlovelace
## Response to 2nd round of reviews

Many thanks for providing follow-up comments, they have been really useful and let to further improvements in the package. All the changes in response to them can be found here: https://github.com/ITSLeeds/stats19/compare/3f38d4cd9144497a0884c1c49c06f680b674acb7...master

Specifically:

1. Use spelling::spell_check_package() and correct the errors
  - See https://github.com/ITSLeeds/stats19/issues/74

2. In the data.R file there are several examples that say something like: "The file names were generated as follows:" IMO any example like this needs to be in the data-raw in Rmd files and only the data in the final package with proper documentation as to what the data are and a link to where they were obtained from. A link to the data-raw files would be appropriate in the documentation for each data set.
  - The examples have been changed to examples of using the data rather than getting the data.

The code to get `police_boundaries` (now in `data-raw`) has been updated to be more robust. It worked on my Ubuntu machine but suspect it failed on Mac. Maybe one for the `sf` issue tracker if so. In any case here's a reprex from the new code that should work on any OS:

``` r
library(sf)
#> Linking to GEOS 3.7.0, GDAL 2.3.2, PROJ 5.2.0
u = "https://opendata.arcgis.com/datasets/3e5a096a8c7c456fb6d3164a3f44b005_3.geojson"
police_boundaries_wgs = sf::st_read(u)
#> Reading layer `3e5a096a8c7c456fb6d3164a3f44b005_3' from data source `https://opendata.arcgis.com/datasets/3e5a096a8c7c456fb6d3164a3f44b005_3.geojson' using driver `GeoJSON'
#> Simple feature collection with 43 features and 9 fields
#> geometry type:  MULTIPOLYGON
#> dimension:      XY
#> bbox:           xmin: -6.402838 ymin: 49.86709 xmax: 1.761138 ymax: 55.81109
#> epsg (SRID):    4326
#> proj4string:    +proj=longlat +datum=WGS84 +no_defs
names(police_boundaries_wgs)
#>  [1] "objectid"       "pfa16cd"        "pfa16nm"        "bng_e"
#>  [5] "bng_n"          "long"           "lat"            "st_areashape"
#>  [9] "st_lengthshape" "geometry"
police_boundaries = st_transform(police_boundaries_wgs, 27700)
names(police_boundaries)
#>  [1] "objectid"       "pfa16cd"        "pfa16nm"        "bng_e"
#>  [5] "bng_n"          "long"           "lat"            "st_areashape"
#>  [9] "st_lengthshape" "geometry"
police_boundaries = police_boundaries[c("pfa16cd", "pfa16nm")]
```

I think that's everything. Any further feedback, very welcome.
