### 2019-01-09 #266 @adamhsparks
Looks great! I've amended my review and recommend accepting this package to rOpenSci now.
