### 2019-01-10 9ffbad4
Changes for ropensci/software-review/#266
