### 2019-01-10 af048b7
Changes for ropensci/software-review/#266
