### 2019-01-10 23af948
Changes for ropensci/software-review/#266
