+++ DESCRIPTION
@@ -46,3 +46,4 @@ VignetteBuilder: knitr
 RoxygenNote: 6.1.1
 Roxygen: list(markdown = TRUE)
 Language: en-US
+X-schema.org-keywords: stats19, road-safety, transport, car-crashes, ropensci, data
