### 2019-01-10 #266 @Robinlovelace
See here for my attempt to implement all those changes: https://github.com/ropensci/stats19/pull/80

Look good to you @sckott ?

One question: how does this process relate to the JOSS paper process? Wondering if we get a citation generated when that goes up, and when that will be.
