### 2019-01-10 #266 @layik
@Robinlovelace
```
From github.com:ropensci/stats19
 * [new branch]      ropensci   -> origin/ropensci
Already up-to-date.```

was it propagation issue?
