### 2019-01-10 #266 @sckott
permissions should be done now - sorry about my UTC - 8 time zone 😞

* For JOSS:
    * I just added a Zenodo web hook - check your webhooks to see it
    * If you haven't already, generate a new release associated with a git tag in your repo. This doesn't mean you have to submit to CRAN with that git tag/release, its just to get a Zenodo DOI - and you can put a zenodo badge in your readme, then after each git tag it will generate a new Zenodo release
    * Then submit your `paper.md` to JOSS at http://joss.theoj.org/papers/new

> can we keep hosting the website on our own org (ITSLeeds)?

Yes. I maintain a metadata csv file https://github.com/ropensci/biweekly/blob/gh-pages/data/pkg_metadata.csv with canonical docs links so can use whatevever link you want
