### 2019-01-10 #266 @stefaniebutland
Great to hear you'd like to contribute a post @Robinlovelace. We're always keen to help get more eyes on people's work.

This link will give you many examples of blog posts by authors of peer-reviewed packages so you can get an idea of the style and length you prefer: https://ropensci.org/tags/software-peer-review/. Technotes are here: https://ropensci.org/technotes/.

Here are some technical and editorial guidelines for contributing a post: https://github.com/ropensci/roweb2#contributing-a-blog-post. We ask that you submit your draft post via pull request a week before the planned publication date so we can give you some feedback. If you give me an idea of when you might submit a draft, we can choose a publication date.

Happy to answer any questions.
