### 2019-01-10 #266 @Robinlovelace
Thanks for all the info Stefanie, really useful background, sure we can do a good job.  How's early February for an ETA?
