### 2019-01-15 #266 @Robinlovelace
Heads-up @sckott I've submitted the package to CRAN and have added a Zenodo DOI badge as suggested. I've also submitted the package to JOSS for peer review. Next steps? Starting some work that may cite the package so will useful to have a 'cannonical' .bib file also.
