### 2019-01-15 #266 @sckott
sounds good @Robinlovelace  that's it. will close issue now. yeah, having a citation file is a good idea
