### 2019-01-15 #266 @Robinlovelace
Fantastic! What is best practice when creating a CITATION file? I can see how to create it from the official documentation here: https://cran.r-project.org/doc/manuals/R-exts.html#CITATION-files

But see that some packages, such as `dplyr` and [`drake`](https://github.com/ropensci/drake) lack one.
