### 2019-01-15 #266 @Robinlovelace
See here for an example of a CITATION file: https://github.com/cran/nlme/tree/master/inst
