### 2019-01-15 #266 @Robinlovelace
And also, what is the official citation info of the package? Will it change when it is published in JOSS?
