### 2019-01-15 #266 @sckott
official with respect to?

I think JOSS would be an additional citation you could be put in the citaiton file
