### 2019-01-15 #266 @Robinlovelace
Just 1 citation for the package is probably enough - that's the aim of the `CITATION` file, I believe, to allow authors to choose what the official citation info is, with respect to the output of `citation()`. This is a good example from the `sf` package:

``` r
citation("sf")
#>
#> To cite package sf in publications use:
#>
#>   Pebesma, E., 2018. Simple Features for R: Standardized Support
#>   for Spatial Vector Data. The R Journal,
#>   https://journal.r-project.org/archive/2018/RJ-2018-009/
#>
#> A BibTeX entry for LaTeX users is
#>
#>   @Article{,
#>     author = {Edzer Pebesma},
#>     title = {{Simple Features for R: Standardized Support for Spatial
#>           Vector Data}},
#>     year = {2018},
#>     journal = {{The R Journal}},
#>     url = {https://journal.r-project.org/archive/2018/RJ-2018-009/index.html},
#>   }
```

The associated `CITATION` file can be found here: https://github.com/r-spatial/sf/blob/master/inst/CITATION

I think it may be useful for authors to have guidance from rOpenSci on this, seeing as you provide really useful guidance on, and state preferences on (no problem with that!) many aspects of package design.
