### 2019-01-16 #266 @Robinlovelace
Many thanks @adamhsparks, very useful. I plan to create something similar to this: https://github.com/ropensci/nasapower/blob/master/inst/CITATION

Worth adding something on CITATION in the package guidance/process?
