### 2019-02-05 #266 @stefaniebutland
> How's early February for an ETA?

Hi @Robinlovelace. Can you suggest a date by which you'd like to submit a draft blog post?
