### 2019-02-06 #266 @Robinlovelace
Sure, how's a week tomorrow? In other words,

``` r
remotes::install_github("ATFutures/calendar")
#> Skipping install of 'calendar' from a github remote, the SHA1 (6c67c0fc) has not changed since last install.
#>   Use `force = TRUE` to force installation
library(calendar)
e = ic_event(start_time = "2019-02-14 10:00", end_time = "2019-02-14 10:00", summary = "Complete stats19 blog post")
e
#> # A tibble: 1 x 4
#>   UID                DTSTART             DTEND               SUMMARY
#>   <chr>              <dttm>              <dttm>              <chr>
#> 1 ical-f3df3945-5e5… 2019-02-14 10:00:00 2019-02-14 10:00:00 Complete sta…
class(e)
#> [1] "ical"       "tbl_df"     "tbl"        "data.frame"
ic_write(e, "e.ics")
```
