### 2019-02-06 #266 @Robinlovelace
Resulting Calendar invite (cc @layik ): https://github.com/ATFutures/calendar/releases/download/0.0.1/e.ics
