### 2019-02-15 #266 @Robinlovelace
Here's the blog post: https://github.com/ropensci/stats19/blob/master/vignettes/blog.Rmd

Any comments/suggestions welcome. Think this is good to go!
