### 2019-02-15 #266 @Robinlovelace
@stefaniebutland one thing: may be worth holding fire: I've just submitted stats19 0.2.0 to CRAN - will be good if the new features are available on the CRAN version when this drops.
