### 2019-02-15 #266 @Robinlovelace
Another update: after some tweaks the article can be found here: https://itsleeds.github.io/stats19/articles/blog.html
