### 2019-02-15 #266 @Robinlovelace
Update: just saw previos messages and am taking a look here https://github.com/ropensci/roweb2#contributing-a-blog-post
