### 2019-02-15 #266 @stefaniebutland
We can absolutely wait till it's on CRAN.

Thank you for getting the draft done. I'll do a proper review once it's submitted as a pull request (so I can check the technical issues at same time as the content)
