### 2019-02-18 #266 @layik
@Robinlovelace and @stefaniebutland

I have now gone through `hugo` instructions to get the post ready for a PR. Slight issue with parsing date or maybe that is even expected in dev mode.

I will send the PR and we can take it from there.
